Package: gbfs

The gbfs package allows users to query tidy datasets about bikeshare programs around the world by supplying a set of functions to interface with .json feeds following the General Bikeshare Feed Specification, a standard data release format developed by the North American Bikeshare Association.

Details

The main function exported by this package is get_gbfs(), which grabs every feed released by a city. Alternatively, the user can just grab information on specific feeds (or groups of feeds).

Each of the feeds described below can be queried with the get_suffix() function, where suffix is replaced with the name of the relevant feed.

Although all of the feeds are livestreamed, only a few of the datasets change often:

station_status: Supplies the number of available bikes and docks at each station as well as station availability

free_bike_status: Gives the coordinates and metadata on available bikes that are parked, but not at a station.

In this package, these two datasets are considered "dynamic", and can be specified as desired datasets by setting `feeds = "dynamic" in the main wrapper function in the package, get_gbfs.

Much of the data supplied in this specification can be considered static. If you want to grab all of these for a given city, set feeds = "static" when calling get_gbfs. Static feeds include:

system_information: Basic metadata about the bikeshare program

station_information: Information on the capacity and coordinates of stations

Several optional feeds: system_hours, system_calendar, system_regions, system_pricing_plans, and system_alerts

See also