Grab the free_bike_status feed.

Grab a dataframe giving the geographic location and other metadata of bikeshare bikes not parked at bikeshare stations. Metadata for this dataset can be found at: https://github.com/NABSA/gbfs/blob/master/gbfs.md

get_free_bike_status(
  city,
  directory = NULL,
  file = "free_bike_status.rds",
  output = NULL
)

Arguments

city
A character string that can be matched to a gbfs feed. The recommended argument is a system ID supplied in the output of [get_gbfs_cities()], but will also attempt to match to the URL of an active .json feed or city name.

directory
Optional. Path to a folder (or folder to be created) where the feed will be saved.

file
Optional. The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
Optional. The type of output method. If left as default, this argument is inferred from the directory argument. If output = "save", the object will be saved as an .rds object at # the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on the argument to output and directory. Either a saved .rds object generated from the current station_information feed, a dataframe object, or both. If a saved feed of the same type already exists at the filepath, the feed will be appended to rather than overwritten.

See also

Examples

# grab the free bike status feed for portland, oregon's bikeshare program 
get_free_bike_status(city = 
"https://gbfs.biketownpdx.com/gbfs/en/free_bike_status.json",
output = "return")

#>              type                          bike_id      lat is_disabled
#> 1   electric_bike 75415701bc2d7984e72283b4aeb11e00 45.51745           0
#> 2   electric_bike 45f0eb8dfc00d5fde99f1f86b89c7204 45.56453           0
#> 3   electric_bike 772a6ceb2fbc94f004ea0fbb8e545a01 45.52101           0
#> 4   electric_bike bb5b2f80fa76335082d9f40dd3a6f5fb 45.48932           0
#> 5   electric_bike 1b03487f83e7f8f4953035872b2f5876 45.52497           0
#> 6   electric_bike 58f3003058018a431ce57613e65bc003 45.56829           0
#> 7   electric_bike b2ecc442e76c28aab727430f996f4896 45.51642           0
#> 8   electric_bike d663e4406d05620aef1748a1117519ec 45.51727           0
#> 9   electric_bike 16646a9a3c9ce8a3997fd3b0835f1a20 45.54361           0
#> 10  electric_bike 07b121d3410aa00e91544571153e4df6 45.49467           0
#> 11  electric_bike e68821989f7d51e6f4dfb0e2d6953ef1 45.53348           0
#> 12  electric_bike b4e1d40ff6cffcbc6ce07a9ae23f1272 45.56018           0
#> 13  electric_bike 5b6d816fdb40746a3aee414dac4ac54b 45.54370           0
#> 14  electric_bike 6b2831c272a0810c0e59cb5fd2ab13e9 45.53347           0
#> 15  electric_bike 44f54a70164c316609c3a54ecaa7700c 45.52578           0
#> 16  electric_bike 262f888850773650f5f0e29b058d9a31 45.52298           0
#> 17  electric_bike c463454886a6df29b8aa89c72b250bb9 45.52217           0
#> 18  electric_bike 96582ddb48be8cddf316f7e298c9f4f5 45.54241           0
#> 19  electric_bike fb6768a715f0ffcc30adf4f184422cd9 45.53496           0
#> 20  electric_bike c9c226db45384d62ba712b2ac0bf6925 45.53341           0
#> 21  electric_bike e8d853d22241100f8676133ea6ab2362 45.55196           0
#> 22  electric_bike 9045da1e89926338fdc6b15187a701bc 45.52370           0
#> 23  electric_bike 13168d8d67e0890856992b9d9b611c81 45.48299           0
#> 24  electric_bike ada7ae800cd0c05c334f028d314465d1 45.49199           0
#> 25  electric_bike 35b6e9956abc9bd9da2c5eabd158d745 45.57720           0
#> 26  electric_bike f9cb002f01109fdb51bd0372601e90a9 45.52005           0
#> 27  electric_bike 65e2c40ddb136cd55b602e234e37fdf5 45.54755           0
#> 28  electric_bike 473a40f15a744b497d4db962c3d769d2 45.50426           0
#> 29  electric_bike 976363ae3f19bbaba4b3c91135100149 45.52905           0
#> 30  electric_bike f60d7bf118bd869ef3c2068e169baff8 45.53093           0
#> 31  electric_bike 4a2f2f7751f3f7cfe52ec7138ea65e9b 45.51842           0
#> 32  electric_bike ba5396f3ea79e192f77b851eb0f8d3da 45.53444           0
#> 33  electric_bike d79eb9b8f852f00b4a41e6cec2b12528 45.49038           0
#> 34  electric_bike 9b76db24dcde06be5344f1e0789db75e 45.52171           0
#> 35  electric_bike 83de5df5abf609821c38ccdf4c2771ce 45.52572           0
#> 36  electric_bike 5e6c6ec7844b61a1df813c4c881ab0f1 45.53340           0
#> 37  electric_bike aa0818867c191712799d9fdfef54ed63 45.51724           0
#> 38  electric_bike 712f66c8a31bc1d038111ea1a09aaac9 45.51967           0
#> 39  electric_bike 4babd16cb09a4578242cc8d96c0ed268 45.56036           0
#> 40  electric_bike 85e4cec1652aacd8acd46e95280d9be4 45.51513           0
#> 41  electric_bike 81a7ae499b54a433b819c06aa03a1f1f 45.53732           0
#> 42  electric_bike cb086b8652e63b738d3cffb353087d7b 45.56209           0
#> 43  electric_bike 552b65eb2d446fc226695e10b59fa9f8 45.51773           0
#> 44  electric_bike cf24185bf01c5a106521e10cb00cb89a 45.52732           0
#> 45  electric_bike 5c2e65222d9f82d5247f67fdb7d0d2b6 45.50232           0
#> 46  electric_bike 364bcbc9435478e26e709f433f7b76e7 45.54140           0
#> 47  electric_bike de3f8ff94682f5e88569d8b7656384b2 45.52320           0
#> 48  electric_bike b237d864c7fcc5a61bd0501d1dc79d9c 45.48927           0
#> 49  electric_bike 40b84a21736964b2f0fcc2db4367df5e 45.49718           0
#> 50  electric_bike e266ca7c841a9a4a1a897a62fae851b5 45.49399           0
#> 51  electric_bike 8ba377504dbbd2091369d57a69146648 45.56278           0
#> 52  electric_bike 027f92bd2734ab2ea5b1c26a500e09c4 45.50231           0
#> 53  electric_bike 4dc978392c47196e68f2174f7d00d5e9 45.49035           0
#> 54  electric_bike cdc2f36fad4b4347d9abfb2a74e5fefd 45.52768           0
#> 55  electric_bike 8785e76ec98e1d32d7c15e32a5efed16 45.52805           0
#> 56  electric_bike c0cf2935032877b30a4844baccd1f493 45.52141           0
#> 57  electric_bike 62f275d9cf9d56c599e2b9d901c24ffc 45.57063           0
#> 58  electric_bike 9506c51f5c7018b0d6f201421a5dd9ee 45.55554           0
#> 59  electric_bike f1df7de57205cd511c304fed69a87bfb 45.53431           0
#> 60  electric_bike 393ffbeab8745d03bcc7d2936f6f961e 45.52616           0
#> 61  electric_bike 3f2d79b2428c2a794ac8239686e2c25e 45.56449           0
#> 62  electric_bike 9436fa29550b069d316070d010045728 45.52289           0
#> 63  electric_bike 0aca2a1baf88d4eb8d240baf5b6c3408 45.48312           0
#> 64  electric_bike ed8c0169a4b6793f7c7b8b7518fd92ed 45.47972           0
#> 65  electric_bike a036b6d03c69505de69217b6e606dfe9 45.54205           0
#> 66  electric_bike d2992533c6458873c0330ec13282b861 45.49666           0
#> 67  electric_bike 602ba54234837abd3e89d7285565d709 45.54689           0
#> 68  electric_bike b2e9853ef9902625e66bf6c130ed2b0b 45.51190           0
#> 69  electric_bike e9b43f44617120753e333dd1fc78b66b 45.51205           0
#> 70  electric_bike 3b8e866af5b5c6831afe26c999ea6342 45.51159           0
#> 71  electric_bike 04b6ca25328661e47f44828c26e11c3d 45.51857           0
#> 72  electric_bike 4b9a7b8f44e0b2b30951bc238175fd64 45.56614           0
#> 73  electric_bike 5e2966ca772ed968961bec4b56fa5f65 45.57705           0
#> 74  electric_bike 92e431036b91a8f039aab23098deec14 45.53740           0
#> 75  electric_bike 9ad9d8ffa57e7ebc8b684afd7c69730b 45.51313           0
#> 76  electric_bike aeff9072c1baf5b027cd4bfb1f0a996e 45.56830           0
#> 77  electric_bike ba987c1125c40ddfe33491c41db3987f 45.57721           0
#> 78  electric_bike 24b15713c6b190a0bcaccaa6a2bbd6a1 45.51786           0
#> 79  electric_bike cffc9c11742d84ea539592823cdd1c0e 45.53458           0
#> 80  electric_bike 3cc801278af09d6edcc8695188f91a77 45.51921           0
#> 81  electric_bike 71970dedf769de8316e80f667138cf32 45.57285           0
#> 82  electric_bike 5da77805ca88d494d952abce4f22a8ae 45.50879           0
#> 83  electric_bike a97b0f5b7f62cd78b6ce80ddca64a63e 45.52251           0
#> 84  electric_bike faa4cd85527d09dcd6595371214c15c5 45.52174           0
#> 85  electric_bike ac299510b815f37f75ae82471c2f6d80 45.55175           0
#> 86  electric_bike 4306f3b294de79430229f1def324359d 45.53320           0
#> 87  electric_bike 2897abc9c7d4d3bd4fc9c062dc7d72a1 45.56688           0
#> 88  electric_bike ff32aaaf33b133b7637fdd8243c52e2b 45.52649           0
#> 89  electric_bike e73f660a3b39ebd68b6afef8929608be 45.51764           0
#> 90  electric_bike 6ab9725e401e97b7fa8baefb7f650a25 45.52109           0
#> 91  electric_bike 56b13062f7753783876906bc403f95bf 45.51701           0
#> 92  electric_bike 671ae9dc2e9f9760ff0ec6eb70566699 45.51932           0
#> 93  electric_bike 4e9b643d236fc91e61a3660731b9b11c 45.52367           0
#> 94  electric_bike e61716965d102041b62a9b5f9512af66 45.52399           0
#> 95  electric_bike 576519f2a55225b81d0a296c23c92165 45.57688           0
#> 96  electric_bike 05a090ab50722cb22db08e210cbfc2d0 45.53345           0
#> 97  electric_bike 8e2349bc310b520e0f8351da26a1b0a8 45.54830           0
#> 98  electric_bike 522874c9526a46e09aca4b616956a0a8 45.53531           0
#> 99  electric_bike 64f4382d4ecc3af03a5c75532b899613 45.52593           0
#> 100 electric_bike 7fa11f6fd945c9a4d254e54d8150480a 45.50544           0
#> 101 electric_bike e76a0b417ebb282c113dcba508d98f09 45.56517           0
#> 102 electric_bike 80c4de5f1cdcbcc45876005c27325391 45.55989           0
#> 103 electric_bike f212eef718aa8475f0d8a9fc833807b3 45.51209           0
#> 104 electric_bike c1ecfbd4e1abcf5b704cadce53f5e380 45.51312           0
#> 105 electric_bike 65c69590885cb1f2fb92d876dacb9d53 45.52280           0
#> 106 electric_bike 242ea3a9af50266aa7b8dceebe60bc5f 45.51912           0
#> 107 electric_bike ae633ad92779ea0d9543fd6769ae420c 45.53616           0
#> 108 electric_bike 1aad33ca94275a5763e8d416c94e7149 45.51733           0
#> 109 electric_bike 9c650480a2948cf071bd933aabe54d0a 45.51227           0
#> 110 electric_bike ebdd1199295e20f820c023b1b3ce8efb 45.48297           0
#> 111 electric_bike 140a5f385ad32ee3d2c7227ad209a6f2 45.53653           0
#> 112 electric_bike 29e7cd296d5f3b44ceac9a70519e4619 45.50791           0
#> 113 electric_bike 3a3668e782279e838416433ce323a972 45.52112           0
#> 114 electric_bike eba1ad8e98b092afcd8dbd6cef44ba7c 45.51762           0
#> 115 electric_bike 7303f53882f9c7f5a93116a27d3b7116 45.56401           0
#> 116 electric_bike 4903239d3e03890a219f056e2b01cf3e 45.53506           0
#> 117 electric_bike 43e9fbc2ace63426384c610b5ff43df5 45.52041           0
#> 118 electric_bike 306ce909ebde3bbfba86a126b8c60d00 45.51702           0
#> 119 electric_bike bfde8de14a2fdfda037c4b86975649c9 45.50583           0
#> 120 electric_bike 17592620d24e33fd56db8de61c7f4a33 45.51459           0
#> 121 electric_bike c1d7dc83210cba0443c65b91581d2ab0 45.50628           0
#> 122 electric_bike 416aaa6d21e998f536b344a0ca6f3d13 45.52843           0
#> 123 electric_bike c8cfb743a937009ba04fefef74f818b1 45.57724           0
#> 124 electric_bike ddf7433ff8cec4801c1090bc38d321a1 45.54367           0
#> 125 electric_bike ce576213f855358940c746f4fe6c0d0a 45.52744           0
#> 126 electric_bike d8b501e2898f96e071b7dcd7d8e3ff23 45.49099           0
#> 127 electric_bike da0ba12a3fab6dd226ba46063f9ff443 45.49415           0
#> 128 electric_bike ab3fba13ff02125d2ba865114120a2dd 45.52980           0
#> 129 electric_bike 1de2f8ac05ed752acbd0232186138167 45.49021           0
#> 130 electric_bike 22504ed1f10a6fda1cc75e32b20cfbc8 45.50089           0
#> 131 electric_bike d6ca2dd660c37a5112a125e0fc6800c4 45.51206           0
#> 132 electric_bike 13f5aecbec4b3052ded519a58db3814e 45.51732           0
#> 133 electric_bike c6bdc9d8037bd49feaaaf809607ba56d 45.54043           0
#> 134 electric_bike 35cb243aa9417a009075951942d6d933 45.51739           0
#> 135 electric_bike ccd8180de489255f93f1b76c093f6b56 45.56526           0
#> 136 electric_bike 34911291973204ecb243deba06d8aa7c 45.52663           0
#> 137 electric_bike b9e1e341ecc4fa5f8306bf35dd6e5748 45.53731           0
#> 138 electric_bike 0f76566dcf5059a40a5d8a6091db6fe6 45.51921           0
#> 139 electric_bike 56f9c8fc7bfb6fb0f1bf35cb052f442d 45.51884           0
#> 140 electric_bike b0de5bcf261c8ae2bb311b8d53d183ae 45.55929           0
#> 141 electric_bike e1ede6c129ca1a0c3ccb3317b45e05d8 45.51642           0
#> 142 electric_bike 7b886c67cd61aab029bed74a84de00d9 45.50164           0
#> 143 electric_bike 091fa64163acd4d3ac1b4ae04ee9d23c 45.52224           0
#> 144 electric_bike e9bbf8f3c1385400de3cac593cd612d1 45.52617           0
#> 145 electric_bike a48f57040d883d185a04f231f73fb253 45.51742           0
#> 146 electric_bike 4a12f1e794960c19302dd98d713d2a0d 45.52389           0
#> 147 electric_bike bfacd24a158d427915bda36cc5ab922d 45.49016           0
#> 148 electric_bike 5bf012738dfd7de78f93a24dec25dc1e 45.51009           0
#> 149 electric_bike bee9b8ec2db08eb2ec45c606038b264a 45.51193           0
#> 150 electric_bike 216df4979ac8ccebff9213b749b22915 45.52658           0
#>           lon                             name is_reserved
#> 1   -122.6801 75415701bc2d7984e72283b4aeb11e00           0
#> 2   -122.6952 45f0eb8dfc00d5fde99f1f86b89c7204           0
#> 3   -122.6751 772a6ceb2fbc94f004ea0fbb8e545a01           0
#> 4   -122.6510 bb5b2f80fa76335082d9f40dd3a6f5fb           0
#> 5   -122.6985 1b03487f83e7f8f4953035872b2f5876           0
#> 6   -122.6393 58f3003058018a431ce57613e65bc003           0
#> 7   -122.6338 b2ecc442e76c28aab727430f996f4896           0
#> 8   -122.6601 d663e4406d05620aef1748a1117519ec           0
#> 9   -122.6306 16646a9a3c9ce8a3997fd3b0835f1a20           0
#> 10  -122.6228 07b121d3410aa00e91544571153e4df6           0
#> 11  -122.6204 e68821989f7d51e6f4dfb0e2d6953ef1           0
#> 12  -122.6378 b4e1d40ff6cffcbc6ce07a9ae23f1272           0
#> 13  -122.6306 5b6d816fdb40746a3aee414dac4ac54b           0
#> 14  -122.5568 6b2831c272a0810c0e59cb5fd2ab13e9           0
#> 15  -122.6841 44f54a70164c316609c3a54ecaa7700c           0
#> 16  -122.6756 262f888850773650f5f0e29b058d9a31           0
#> 17  -122.6280 c463454886a6df29b8aa89c72b250bb9           0
#> 18  -122.6505 96582ddb48be8cddf316f7e298c9f4f5           0
#> 19  -122.6637 fb6768a715f0ffcc30adf4f184422cd9           0
#> 20  -122.7029 c9c226db45384d62ba712b2ac0bf6925           0
#> 21  -122.6474 e8d853d22241100f8676133ea6ab2362           0
#> 22  -122.6722 9045da1e89926338fdc6b15187a701bc           0
#> 23  -122.5777 13168d8d67e0890856992b9d9b611c81           0
#> 24  -122.6278 ada7ae800cd0c05c334f028d314465d1           0
#> 25  -122.6883 35b6e9956abc9bd9da2c5eabd158d745           0
#> 26  -122.6820 f9cb002f01109fdb51bd0372601e90a9           0
#> 27  -122.6640 65e2c40ddb136cd55b602e234e37fdf5           0
#> 28  -122.6611 473a40f15a744b497d4db962c3d769d2           0
#> 29  -122.5579 976363ae3f19bbaba4b3c91135100149           0
#> 30  -122.6987 f60d7bf118bd869ef3c2068e169baff8           0
#> 31  -122.6799 4a2f2f7751f3f7cfe52ec7138ea65e9b           0
#> 32  -122.6610 ba5396f3ea79e192f77b851eb0f8d3da           0
#> 33  -122.6278 d79eb9b8f852f00b4a41e6cec2b12528           0
#> 34  -122.6787 9b76db24dcde06be5344f1e0789db75e           0
#> 35  -122.6494 83de5df5abf609821c38ccdf4c2771ce           0
#> 36  -122.5445 5e6c6ec7844b61a1df813c4c881ab0f1           0
#> 37  -122.6645 aa0818867c191712799d9fdfef54ed63           0
#> 38  -122.6736 712f66c8a31bc1d038111ea1a09aaac9           0
#> 39  -122.6369 4babd16cb09a4578242cc8d96c0ed268           0
#> 40  -122.6495 85e4cec1652aacd8acd46e95280d9be4           0
#> 41  -122.6874 81a7ae499b54a433b819c06aa03a1f1f           0
#> 42  -122.6205 cb086b8652e63b738d3cffb353087d7b           0
#> 43  -122.6760 552b65eb2d446fc226695e10b59fa9f8           0
#> 44  -122.6415 cf24185bf01c5a106521e10cb00cb89a           0
#> 45  -122.6720 5c2e65222d9f82d5247f67fdb7d0d2b6           0
#> 46  -122.6587 364bcbc9435478e26e709f433f7b76e7           0
#> 47  -122.6902 de3f8ff94682f5e88569d8b7656384b2           0
#> 48  -122.5961 b237d864c7fcc5a61bd0501d1dc79d9c           0
#> 49  -122.6720 40b84a21736964b2f0fcc2db4367df5e           0
#> 50  -122.5946 e266ca7c841a9a4a1a897a62fae851b5           0
#> 51  -122.6668 8ba377504dbbd2091369d57a69146648           0
#> 52  -122.6720 027f92bd2734ab2ea5b1c26a500e09c4           0
#> 53  -122.6288 4dc978392c47196e68f2174f7d00d5e9           0
#> 54  -122.6767 cdc2f36fad4b4347d9abfb2a74e5fefd           0
#> 55  -122.6288 8785e76ec98e1d32d7c15e32a5efed16           0
#> 56  -122.6982 c0cf2935032877b30a4844baccd1f493           0
#> 57  -122.7068 62f275d9cf9d56c599e2b9d901c24ffc           0
#> 58  -122.6720 9506c51f5c7018b0d6f201421a5dd9ee           0
#> 59  -122.6332 f1df7de57205cd511c304fed69a87bfb           0
#> 60  -122.6786 393ffbeab8745d03bcc7d2936f6f961e           0
#> 61  -122.6530 3f2d79b2428c2a794ac8239686e2c25e           0
#> 62  -122.6757 9436fa29550b069d316070d010045728           0
#> 63  -122.5801 0aca2a1baf88d4eb8d240baf5b6c3408           0
#> 64  -122.5680 ed8c0169a4b6793f7c7b8b7518fd92ed           0
#> 65  -122.6239 a036b6d03c69505de69217b6e606dfe9           0
#> 66  -122.6686 d2992533c6458873c0330ec13282b861           0
#> 67  -122.6679 602ba54234837abd3e89d7285565d709           0
#> 68  -122.6277 b2e9853ef9902625e66bf6c130ed2b0b           0
#> 69  -122.6332 e9b43f44617120753e333dd1fc78b66b           0
#> 70  -122.6586 3b8e866af5b5c6831afe26c999ea6342           0
#> 71  -122.6791 04b6ca25328661e47f44828c26e11c3d           0
#> 72  -122.6824 4b9a7b8f44e0b2b30951bc238175fd64           0
#> 73  -122.6874 5e2966ca772ed968961bec4b56fa5f65           0
#> 74  -122.6536 92e431036b91a8f039aab23098deec14           0
#> 75  -122.5786 9ad9d8ffa57e7ebc8b684afd7c69730b           0
#> 76  -122.6382 aeff9072c1baf5b027cd4bfb1f0a996e           0
#> 77  -122.6975 ba987c1125c40ddfe33491c41db3987f           0
#> 78  -122.6542 24b15713c6b190a0bcaccaa6a2bbd6a1           0
#> 79  -122.6214 cffc9c11742d84ea539592823cdd1c0e           0
#> 80  -122.5823 3cc801278af09d6edcc8695188f91a77           0
#> 81  -122.6689 71970dedf769de8316e80f667138cf32           0
#> 82  -122.6599 5da77805ca88d494d952abce4f22a8ae           0
#> 83  -122.5527 a97b0f5b7f62cd78b6ce80ddca64a63e           0
#> 84  -122.6824 faa4cd85527d09dcd6595371214c15c5           0
#> 85  -122.6757 ac299510b815f37f75ae82471c2f6d80           0
#> 86  -122.6616 4306f3b294de79430229f1def324359d           0
#> 87  -122.6610 2897abc9c7d4d3bd4fc9c062dc7d72a1           0
#> 88  -122.6808 ff32aaaf33b133b7637fdd8243c52e2b           0
#> 89  -122.6515 e73f660a3b39ebd68b6afef8929608be           0
#> 90  -122.6615 6ab9725e401e97b7fa8baefb7f650a25           0
#> 91  -122.6850 56b13062f7753783876906bc403f95bf           0
#> 92  -122.6575 671ae9dc2e9f9760ff0ec6eb70566699           0
#> 93  -122.6634 4e9b643d236fc91e61a3660731b9b11c           0
#> 94  -122.6997 e61716965d102041b62a9b5f9512af66           0
#> 95  -122.6954 576519f2a55225b81d0a296c23c92165           0
#> 96  -122.6203 05a090ab50722cb22db08e210cbfc2d0           0
#> 97  -122.6043 8e2349bc310b520e0f8351da26a1b0a8           0
#> 98  -122.7086 522874c9526a46e09aca4b616956a0a8           0
#> 99  -122.6839 64f4382d4ecc3af03a5c75532b899613           0
#> 100 -122.5833 7fa11f6fd945c9a4d254e54d8150480a           0
#> 101 -122.6093 e76a0b417ebb282c113dcba508d98f09           0
#> 102 -122.6267 80c4de5f1cdcbcc45876005c27325391           0
#> 103 -122.6557 f212eef718aa8475f0d8a9fc833807b3           0
#> 104 -122.5786 c1ecfbd4e1abcf5b704cadce53f5e380           0
#> 105 -122.6597 65c69590885cb1f2fb92d876dacb9d53           0
#> 106 -122.6839 242ea3a9af50266aa7b8dceebe60bc5f           0
#> 107 -122.6668 ae633ad92779ea0d9543fd6769ae420c           0
#> 108 -122.6581 1aad33ca94275a5763e8d416c94e7149           0
#> 109 -122.6435 9c650480a2948cf071bd933aabe54d0a           0
#> 110 -122.5801 ebdd1199295e20f820c023b1b3ce8efb           0
#> 111 -122.6486 140a5f385ad32ee3d2c7227ad209a6f2           0
#> 112 -122.6548 29e7cd296d5f3b44ceac9a70519e4619           0
#> 113 -122.6807 3a3668e782279e838416433ce323a972           0
#> 114 -122.6761 eba1ad8e98b092afcd8dbd6cef44ba7c           0
#> 115 -122.6608 7303f53882f9c7f5a93116a27d3b7116           0
#> 116 -122.6528 4903239d3e03890a219f056e2b01cf3e           0
#> 117 -122.6743 43e9fbc2ace63426384c610b5ff43df5           0
#> 118 -122.5925 306ce909ebde3bbfba86a126b8c60d00           0
#> 119 -122.6644 bfde8de14a2fdfda037c4b86975649c9           0
#> 120 -122.6754 17592620d24e33fd56db8de61c7f4a33           0
#> 121 -122.6297 c1d7dc83210cba0443c65b91581d2ab0           0
#> 122 -122.6616 416aaa6d21e998f536b344a0ca6f3d13           0
#> 123 -122.6974 c8cfb743a937009ba04fefef74f818b1           0
#> 124 -122.6308 ddf7433ff8cec4801c1090bc38d321a1           0
#> 125 -122.6768 ce576213f855358940c746f4fe6c0d0a           0
#> 126 -122.5716 d8b501e2898f96e071b7dcd7d8e3ff23           0
#> 127 -122.5765 da0ba12a3fab6dd226ba46063f9ff443           0
#> 128 -122.6842 ab3fba13ff02125d2ba865114120a2dd           0
#> 129 -122.6195 1de2f8ac05ed752acbd0232186138167           0
#> 130 -122.6817 22504ed1f10a6fda1cc75e32b20cfbc8           0
#> 131 -122.6238 d6ca2dd660c37a5112a125e0fc6800c4           0
#> 132 -122.6595 13f5aecbec4b3052ded519a58db3814e           0
#> 133 -122.6094 c6bdc9d8037bd49feaaaf809607ba56d           0
#> 134 -122.6846 35cb243aa9417a009075951942d6d933           0
#> 135 -122.6820 ccd8180de489255f93f1b76c093f6b56           0
#> 136 -122.6779 34911291973204ecb243deba06d8aa7c           0
#> 137 -122.6874 b9e1e341ecc4fa5f8306bf35dd6e5748           0
#> 138 -122.5823 0f76566dcf5059a40a5d8a6091db6fe6           0
#> 139 -122.6640 56f9c8fc7bfb6fb0f1bf35cb052f442d           0
#> 140 -122.6205 b0de5bcf261c8ae2bb311b8d53d183ae           0
#> 141 -122.6478 e1ede6c129ca1a0c3ccb3317b45e05d8           0
#> 142 -122.6131 7b886c67cd61aab029bed74a84de00d9           0
#> 143 -122.6845 091fa64163acd4d3ac1b4ae04ee9d23c           0
#> 144 -122.5583 e9bbf8f3c1385400de3cac593cd612d1           0
#> 145 -122.6749 a48f57040d883d185a04f231f73fb253           0
#> 146 -122.6598 4a12f1e794960c19302dd98d713d2a0d           0
#> 147 -122.5978 bfacd24a158d427915bda36cc5ab922d           0
#> 148 -122.5695 5bf012738dfd7de78f93a24dec25dc1e           0
#> 149 -122.6278 bee9b8ec2db08eb2ec45c606038b264a           0
#> 150 -122.6863 216df4979ac8ccebff9213b749b22915           0
#>                     rental_uris.android                     rental_uris.ios
#> 1   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 2   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 3   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 4   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 5   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 6   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 7   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 8   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 9   https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 10  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 11  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 12  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 13  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 14  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 15  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 16  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 17  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 18  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 19  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 20  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 21  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 22  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 23  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 24  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 25  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 26  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 27  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 28  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 29  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 30  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 31  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 32  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 33  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 34  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 35  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 36  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 37  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 38  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 39  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 40  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 41  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 42  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 43  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 44  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 45  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 46  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 47  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 48  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 49  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 50  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 51  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 52  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 53  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 54  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 55  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 56  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 57  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 58  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 59  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 60  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 61  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 62  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 63  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 64  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 65  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 66  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 67  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 68  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 69  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 70  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 71  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 72  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 73  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 74  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 75  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 76  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 77  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 78  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 79  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 80  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 81  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 82  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 83  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 84  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 85  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 86  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 87  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 88  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 89  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 90  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 91  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 92  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 93  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 94  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 95  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 96  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 97  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 98  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 99  https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 100 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 101 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 102 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 103 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 104 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 105 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 106 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 107 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 108 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 109 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 110 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 111 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 112 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 113 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 114 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 115 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 116 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 117 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 118 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 119 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 120 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 121 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 122 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 123 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 124 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 125 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 126 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 127 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 128 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 129 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 130 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 131 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 132 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 133 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 134 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 135 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 136 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 137 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 138 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 139 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 140 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 141 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 142 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 143 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 144 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 145 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 146 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 147 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 148 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 149 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#> 150 https://pdx.lft.to/lastmile_qr_scan https://pdx.lft.to/lastmile_qr_scan
#>            last_updated year month day hour minute
#> 1   2021-01-05 17:40:05 2021     1   5   17     40
#> 2   2021-01-05 17:40:05 2021     1   5   17     40
#> 3   2021-01-05 17:40:05 2021     1   5   17     40
#> 4   2021-01-05 17:40:05 2021     1   5   17     40
#> 5   2021-01-05 17:40:05 2021     1   5   17     40
#> 6   2021-01-05 17:40:05 2021     1   5   17     40
#> 7   2021-01-05 17:40:05 2021     1   5   17     40
#> 8   2021-01-05 17:40:05 2021     1   5   17     40
#> 9   2021-01-05 17:40:05 2021     1   5   17     40
#> 10  2021-01-05 17:40:05 2021     1   5   17     40
#> 11  2021-01-05 17:40:05 2021     1   5   17     40
#> 12  2021-01-05 17:40:05 2021     1   5   17     40
#> 13  2021-01-05 17:40:05 2021     1   5   17     40
#> 14  2021-01-05 17:40:05 2021     1   5   17     40
#> 15  2021-01-05 17:40:05 2021     1   5   17     40
#> 16  2021-01-05 17:40:05 2021     1   5   17     40
#> 17  2021-01-05 17:40:05 2021     1   5   17     40
#> 18  2021-01-05 17:40:05 2021     1   5   17     40
#> 19  2021-01-05 17:40:05 2021     1   5   17     40
#> 20  2021-01-05 17:40:05 2021     1   5   17     40
#> 21  2021-01-05 17:40:05 2021     1   5   17     40
#> 22  2021-01-05 17:40:05 2021     1   5   17     40
#> 23  2021-01-05 17:40:05 2021     1   5   17     40
#> 24  2021-01-05 17:40:05 2021     1   5   17     40
#> 25  2021-01-05 17:40:05 2021     1   5   17     40
#> 26  2021-01-05 17:40:05 2021     1   5   17     40
#> 27  2021-01-05 17:40:05 2021     1   5   17     40
#> 28  2021-01-05 17:40:05 2021     1   5   17     40
#> 29  2021-01-05 17:40:05 2021     1   5   17     40
#> 30  2021-01-05 17:40:05 2021     1   5   17     40
#> 31  2021-01-05 17:40:05 2021     1   5   17     40
#> 32  2021-01-05 17:40:05 2021     1   5   17     40
#> 33  2021-01-05 17:40:05 2021     1   5   17     40
#> 34  2021-01-05 17:40:05 2021     1   5   17     40
#> 35  2021-01-05 17:40:05 2021     1   5   17     40
#> 36  2021-01-05 17:40:05 2021     1   5   17     40
#> 37  2021-01-05 17:40:05 2021     1   5   17     40
#> 38  2021-01-05 17:40:05 2021     1   5   17     40
#> 39  2021-01-05 17:40:05 2021     1   5   17     40
#> 40  2021-01-05 17:40:05 2021     1   5   17     40
#> 41  2021-01-05 17:40:05 2021     1   5   17     40
#> 42  2021-01-05 17:40:05 2021     1   5   17     40
#> 43  2021-01-05 17:40:05 2021     1   5   17     40
#> 44  2021-01-05 17:40:05 2021     1   5   17     40
#> 45  2021-01-05 17:40:05 2021     1   5   17     40
#> 46  2021-01-05 17:40:05 2021     1   5   17     40
#> 47  2021-01-05 17:40:05 2021     1   5   17     40
#> 48  2021-01-05 17:40:05 2021     1   5   17     40
#> 49  2021-01-05 17:40:05 2021     1   5   17     40
#> 50  2021-01-05 17:40:05 2021     1   5   17     40
#> 51  2021-01-05 17:40:05 2021     1   5   17     40
#> 52  2021-01-05 17:40:05 2021     1   5   17     40
#> 53  2021-01-05 17:40:05 2021     1   5   17     40
#> 54  2021-01-05 17:40:05 2021     1   5   17     40
#> 55  2021-01-05 17:40:05 2021     1   5   17     40
#> 56  2021-01-05 17:40:05 2021     1   5   17     40
#> 57  2021-01-05 17:40:05 2021     1   5   17     40
#> 58  2021-01-05 17:40:05 2021     1   5   17     40
#> 59  2021-01-05 17:40:05 2021     1   5   17     40
#> 60  2021-01-05 17:40:05 2021     1   5   17     40
#> 61  2021-01-05 17:40:05 2021     1   5   17     40
#> 62  2021-01-05 17:40:05 2021     1   5   17     40
#> 63  2021-01-05 17:40:05 2021     1   5   17     40
#> 64  2021-01-05 17:40:05 2021     1   5   17     40
#> 65  2021-01-05 17:40:05 2021     1   5   17     40
#> 66  2021-01-05 17:40:05 2021     1   5   17     40
#> 67  2021-01-05 17:40:05 2021     1   5   17     40
#> 68  2021-01-05 17:40:05 2021     1   5   17     40
#> 69  2021-01-05 17:40:05 2021     1   5   17     40
#> 70  2021-01-05 17:40:05 2021     1   5   17     40
#> 71  2021-01-05 17:40:05 2021     1   5   17     40
#> 72  2021-01-05 17:40:05 2021     1   5   17     40
#> 73  2021-01-05 17:40:05 2021     1   5   17     40
#> 74  2021-01-05 17:40:05 2021     1   5   17     40
#> 75  2021-01-05 17:40:05 2021     1   5   17     40
#> 76  2021-01-05 17:40:05 2021     1   5   17     40
#> 77  2021-01-05 17:40:05 2021     1   5   17     40
#> 78  2021-01-05 17:40:05 2021     1   5   17     40
#> 79  2021-01-05 17:40:05 2021     1   5   17     40
#> 80  2021-01-05 17:40:05 2021     1   5   17     40
#> 81  2021-01-05 17:40:05 2021     1   5   17     40
#> 82  2021-01-05 17:40:05 2021     1   5   17     40
#> 83  2021-01-05 17:40:05 2021     1   5   17     40
#> 84  2021-01-05 17:40:05 2021     1   5   17     40
#> 85  2021-01-05 17:40:05 2021     1   5   17     40
#> 86  2021-01-05 17:40:05 2021     1   5   17     40
#> 87  2021-01-05 17:40:05 2021     1   5   17     40
#> 88  2021-01-05 17:40:05 2021     1   5   17     40
#> 89  2021-01-05 17:40:05 2021     1   5   17     40
#> 90  2021-01-05 17:40:05 2021     1   5   17     40
#> 91  2021-01-05 17:40:05 2021     1   5   17     40
#> 92  2021-01-05 17:40:05 2021     1   5   17     40
#> 93  2021-01-05 17:40:05 2021     1   5   17     40
#> 94  2021-01-05 17:40:05 2021     1   5   17     40
#> 95  2021-01-05 17:40:05 2021     1   5   17     40
#> 96  2021-01-05 17:40:05 2021     1   5   17     40
#> 97  2021-01-05 17:40:05 2021     1   5   17     40
#> 98  2021-01-05 17:40:05 2021     1   5   17     40
#> 99  2021-01-05 17:40:05 2021     1   5   17     40
#> 100 2021-01-05 17:40:05 2021     1   5   17     40
#> 101 2021-01-05 17:40:05 2021     1   5   17     40
#> 102 2021-01-05 17:40:05 2021     1   5   17     40
#> 103 2021-01-05 17:40:05 2021     1   5   17     40
#> 104 2021-01-05 17:40:05 2021     1   5   17     40
#> 105 2021-01-05 17:40:05 2021     1   5   17     40
#> 106 2021-01-05 17:40:05 2021     1   5   17     40
#> 107 2021-01-05 17:40:05 2021     1   5   17     40
#> 108 2021-01-05 17:40:05 2021     1   5   17     40
#> 109 2021-01-05 17:40:05 2021     1   5   17     40
#> 110 2021-01-05 17:40:05 2021     1   5   17     40
#> 111 2021-01-05 17:40:05 2021     1   5   17     40
#> 112 2021-01-05 17:40:05 2021     1   5   17     40
#> 113 2021-01-05 17:40:05 2021     1   5   17     40
#> 114 2021-01-05 17:40:05 2021     1   5   17     40
#> 115 2021-01-05 17:40:05 2021     1   5   17     40
#> 116 2021-01-05 17:40:05 2021     1   5   17     40
#> 117 2021-01-05 17:40:05 2021     1   5   17     40
#> 118 2021-01-05 17:40:05 2021     1   5   17     40
#> 119 2021-01-05 17:40:05 2021     1   5   17     40
#> 120 2021-01-05 17:40:05 2021     1   5   17     40
#> 121 2021-01-05 17:40:05 2021     1   5   17     40
#> 122 2021-01-05 17:40:05 2021     1   5   17     40
#> 123 2021-01-05 17:40:05 2021     1   5   17     40
#> 124 2021-01-05 17:40:05 2021     1   5   17     40
#> 125 2021-01-05 17:40:05 2021     1   5   17     40
#> 126 2021-01-05 17:40:05 2021     1   5   17     40
#> 127 2021-01-05 17:40:05 2021     1   5   17     40
#> 128 2021-01-05 17:40:05 2021     1   5   17     40
#> 129 2021-01-05 17:40:05 2021     1   5   17     40
#> 130 2021-01-05 17:40:05 2021     1   5   17     40
#> 131 2021-01-05 17:40:05 2021     1   5   17     40
#> 132 2021-01-05 17:40:05 2021     1   5   17     40
#> 133 2021-01-05 17:40:05 2021     1   5   17     40
#> 134 2021-01-05 17:40:05 2021     1   5   17     40
#> 135 2021-01-05 17:40:05 2021     1   5   17     40
#> 136 2021-01-05 17:40:05 2021     1   5   17     40
#> 137 2021-01-05 17:40:05 2021     1   5   17     40
#> 138 2021-01-05 17:40:05 2021     1   5   17     40
#> 139 2021-01-05 17:40:05 2021     1   5   17     40
#> 140 2021-01-05 17:40:05 2021     1   5   17     40
#> 141 2021-01-05 17:40:05 2021     1   5   17     40
#> 142 2021-01-05 17:40:05 2021     1   5   17     40
#> 143 2021-01-05 17:40:05 2021     1   5   17     40
#> 144 2021-01-05 17:40:05 2021     1   5   17     40
#> 145 2021-01-05 17:40:05 2021     1   5   17     40
#> 146 2021-01-05 17:40:05 2021     1   5   17     40
#> 147 2021-01-05 17:40:05 2021     1   5   17     40
#> 148 2021-01-05 17:40:05 2021     1   5   17     40
#> 149 2021-01-05 17:40:05 2021     1   5   17     40
#> 150 2021-01-05 17:40:05 2021     1   5   17     40