Save gbfs feeds.

get_gbfs checks for the existence of gbfs feeds for a given city and saves the feeds as .rds objects in a directory that can be specified by the user. Go to `https://github.com/NABSA/gbfs/blob/master/gbfs.md` to see metadata for each dataset.

get_gbfs(city, feeds = "all", directory = "gbfs_data")

Arguments

city
A character string or a url to an active gbfs.json feed.

feeds
A character string specifying which feeds should be saved. Options are "all", "static", and "dynamic".

directory
The name of an existing folder or folder to be created, where the feeds will be saved.

Value

A folder containing the specified feeds saved as .rds objects.

Examples

get_gbfs(city = "boise")

get_gbfs(city = "http://biketownpdx.socialbicycles.com/opendata/gbfs.json", feeds = "dynamic")

get_gbfs("https://gbfs.bcycle.com/bcycle_greenbikeslc/gbfs.json", directory = "slcbikes")