Grab the station_information feed.

get_station_information grabs and tidies the station_information feed for a given city. This dataset contains locations, capacity, and other information about bikeshare stations. Metadata for this dataset can be found at: https://github.com/NABSA/gbfs/blob/master/gbfs.md

get_station_information(
  city,
  directory = NULL,
  file = "station_information.rds",
  output = NULL
)

Arguments

city
A character string that can be matched to a city or a url to an active gbfs .json feed. See [get_gbfs_cities()] for a current list of available cities.

directory
Optional. Path to a folder (or folder to be created) where the feed will be saved.

file
Optional. The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
Optional. The type of output method. If left as default, this argument is inferred from the directory argument. If output = "save", the object will be saved as an .rds object at # the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on argument to output and directory. Either a saved .rds object generated from the current feed, a dataframe object, or both.

See also

Examples

# we can grab the station information feed for memphis, 
# tennessee in one of several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "memphis", maybe,) for the `directory` argument 
# instead of `tempdir`.
get_station_information(city = 
"https://gbfs.bcycle.com/bcycle_memphis/station_information.json",  
                        directory = tempdir())
                    
# or, instead, just supply the name of 
# the city as a string. 
get_station_information(city = "memphis",
                        directory = tempdir())
                    
# instead of saving the output as a file, we can 
# just return the output as a dataframe
get_station_information(city = "memphis",  
                        output = "return")

#>          lon      lat                          address
#> 1  -90.04896 35.14742      Corner of B.B. King & Adams
#> 2  -90.04828 35.15682                   N. Main Street
#> 3  -90.05042 35.15087                   N. Main Street
#> 4  -90.06229 35.13294            West GE Patterson Ave
#> 5  -90.05385 35.14648                     Front Street
#> 6  -90.05231 35.14612                   N. Main Street
#> 7  -90.04841 35.14379                      Stadium Ct.
#> 8  -90.05369 35.14336                   S. Main Street
#> 9  -90.05092 35.14272      Autozone Park Main Entrance
#> 10 -90.05636 35.14063        Corner of Front and Beale
#> 11 -90.05411 35.14002 Between Main and Second on Beale
#> 12 -90.05179 35.14844                     Adams Avenue
#> 13 -90.05186 35.13914                     191 Beale St
#> 14 -90.05982 35.13963              251 Riverside Drive
#> 15 -90.05849 35.13775          Corner of Front & Vance
#> 16 -90.05748 35.13642                    S Main Street
#> 17 -90.06358 35.13568                  Riverside Drive
#> 18 -90.05813 35.13338                  450 Mulberry St
#> 19 -90.06195 35.12809                 7 W Carolina Ave
#> 20 -90.06915 35.12602   Parking Area on Virginia Ave W
#> 21 -90.03530 35.14080                      Madison Ave
#> 22 -90.04137 35.14203                      Madison Ave
#> 23 -90.03720 35.14087                  Monroe Ave Ext.
#> 24 -90.03011 35.14041                  930 Madison Ave
#> 25 -90.04035 35.14069                       Monroe Ave
#> 26 -90.03664 35.13408                        Vance Ave
#> 27 -90.03902 35.11969                 Mississippi Blvd
#> 28 -90.03262 35.11942     Corner of Neptune and Walker
#> 29 -90.04266 35.11860                  590 Jennette Pl
#> 30 -90.02147 35.11510                  S Bellevue Blvd
#> 31 -90.03391 35.10854            1400 Mississippi Blvd
#> 32 -90.03278 35.10682                   Genesis Circle
#> 33 -90.01420 35.15110                   N Cleveland St
#> 34 -89.99332 35.14634                Veterans Plaza Dr
#> 35 -90.01955 35.13906                 1245 Madison Ave
#> 36 -90.01632 35.13858                   S Cleveland St
#> 37 -90.00767 35.13829     Corner of Madison and Avalon
#> 38 -89.99209 35.13695                      Madison Ave
#> 39 -90.01690 35.13352         On S. Cleveland at Vance
#> 40 -90.00043 35.13229                    S McLean Blvd
#> 41 -89.99031 35.13177                        Cooper St
#> 42 -89.99065 35.12493                        Cooper St
#> 43 -89.96155 35.11421                      Waverly Ave
#> 44 -89.98725 35.10718                         Park Ave
#> 45 -89.97535 35.10862                        Dallas St
#> 46 -89.96831 35.10460                     Deadrick Ave
#> 47 -90.08542 35.13523   Entrance of Big River Crossing
#> 48 -89.99932 35.14101                    N McLean Blvd
#> 49 -90.13981 35.14608                East Broadway Ave
#> 50 -89.99483 35.14988                 2000 Prentiss Pl
#> 51 -89.99099 35.12097                              TBD
#> 52 -90.05623 35.16192                     Island Drive
#> 53 -90.05573 35.17546                     Island Drive
#> 54 -90.05622 35.18090                     Island Drive
#> 55 -89.98254 35.14655                              TBD
#> 56 -90.04573 35.14651                    296 Adams Ave
#> 57 -90.00735 35.14879            1567 Overton Park Ave
#> 58 -90.03888 35.14479                   652 Adams Ave.
#> 59 -90.03088 35.11589              926 E. McLemore Ave
#> 60 -90.05520 35.14675             51 N. Riverside Blvd
#> 61 -89.99104 35.11812               1000 Cooper Street
#> 62 -90.04922 35.15401                   Main & Jackson
#> 63 -90.05390 35.14557                   60 Madison Ave
#> 64 -90.04778 35.14275                 70 S. 4th Street
#> 65 -90.05540 35.14256                              TBD
#> 66 -90.05304 35.14422                   110 Monroe Ave
#> 67 -90.04430 35.15083                              TBD
#> 68 -90.05439 35.14095             Main @ Peabody Ave. 
#> 69 -90.03511 35.14353                 Manassas @ Adams
#> 70 -90.05048 35.14902               160 N. Main Street
#> 71 -90.05811 35.14937                              TBD
#> 72 -90.05870 35.14964                              TBD
#> 73 -90.05304 35.14918                              TBD
#> 74 -90.03675 35.13611     737 Union Ave - on MLK Drive
#> 75 -90.05542 35.14366                     47 Union Ave
#> 76 -90.04435 35.13353                     East Ratliff
#> 77 -90.04609 35.13375                     West Ratliff
#> 78 -90.04959 35.13698                              TBD
#>                             name          station_id
#> 1            B.B. King and Adams bcycle_memphis_3864
#> 2          Hudson Transit Center bcycle_memphis_3872
#> 3      Memphis Convention Center bcycle_memphis_3873
#> 4   Tennessee and G.E. Patterson bcycle_memphis_3874
#> 5                Front and Court bcycle_memphis_3875
#> 6                   Court Square bcycle_memphis_3924
#> 7                  Downtown YMCA bcycle_memphis_3925
#> 8                 Main and Union bcycle_memphis_3926
#> 9                  Autozone Park bcycle_memphis_3927
#> 10               Front and Beale bcycle_memphis_3928
#> 11                 2nd and Beale bcycle_memphis_3929
#> 12                Main and Adams bcycle_memphis_3930
#> 13             FedEx Forum Plaza bcycle_memphis_3931
#> 14          Beale Street Landing bcycle_memphis_3932
#> 15               Front and Vance bcycle_memphis_3933
#> 16               Main and Talbot bcycle_memphis_3934
#> 17                  Tom Lee Park bcycle_memphis_3935
#> 18  National Civil Rights Museum bcycle_memphis_3936
#> 19                   Loflin Yard bcycle_memphis_3937
#> 20       Big River Crossing-East bcycle_memphis_3938
#> 21          Madison and Manassas bcycle_memphis_3940
#> 22          Madison and Marshall bcycle_memphis_3941
#> 23                 Edge Triangle bcycle_memphis_3942
#> 24                   920 Madison bcycle_memphis_3943
#> 25           Marshall and Monroe bcycle_memphis_3944
#> 26              Vance and Walnut bcycle_memphis_3945
#> 27        Mississippi and Walker bcycle_memphis_3947
#> 28            Walker and Neptune bcycle_memphis_3948
#> 29               Knowledge Quest bcycle_memphis_3949
#> 30         Bellevue and McLemore bcycle_memphis_3951
#> 31  South Memphis Farmers Market bcycle_memphis_3952
#> 32                     The Works bcycle_memphis_3953
#> 33           Crosstown Concourse bcycle_memphis_3955
#> 34                  Overton Park bcycle_memphis_3957
#> 35 Southern College of Optometry bcycle_memphis_3959
#> 36         Madison and Cleveland bcycle_memphis_3960
#> 37            Madison and Avalon bcycle_memphis_3961
#> 38                Overton Square bcycle_memphis_3963
#> 39           Cleveland and Vance bcycle_memphis_3965
#> 40            Peabody and McLean bcycle_memphis_3966
#> 41            Cooper and Peabody bcycle_memphis_3967
#> 42          Memphis Made Brewery bcycle_memphis_3968
#> 43            Semmes and Waverly bcycle_memphis_3970
#> 44         Lamar Shopping Center bcycle_memphis_3971
#> 45 Orange Mound Community Center bcycle_memphis_3972
#> 46              Lewis Davis Park bcycle_memphis_3973
#> 47       Big River Crossing-West bcycle_memphis_3974
#> 48             McLean and Poplar bcycle_memphis_3976
#> 49               Big River Trail bcycle_memphis_3984
#> 50                   Memphis Zoo bcycle_memphis_3986
#> 51               Cooper & Oliver bcycle_memphis_4101
#> 52        Greenbelt Park - South bcycle_memphis_4196
#> 53      Greenbelt Park - Central bcycle_memphis_4197
#> 54        Greenbelt Park - North bcycle_memphis_4198
#> 55           Overton Park - East bcycle_memphis_4201
#> 56                 4th and Adams bcycle_memphis_4255
#> 57                     Evergreen bcycle_memphis_4264
#> 58             Victorian Village bcycle_memphis_4269
#> 59                   Stax Museum bcycle_memphis_4279
#> 60        Mississippi River Park bcycle_memphis_4281
#> 61            First Congo Church bcycle_memphis_4285
#> 62                Main & Jackson bcycle_memphis_4307
#> 63               Madison & Front bcycle_memphis_4308
#> 64               Fielders Square bcycle_memphis_4309
#> 65                Front & Gayoso bcycle_memphis_4313
#> 66       Main & Monroe Starbucks bcycle_memphis_4314
#> 67      Uptown Square Apartments bcycle_memphis_4328
#> 68                Main & Peabody bcycle_memphis_4330
#> 69              Manassas & Adams bcycle_memphis_4368
#> 70            Shelby Co Building bcycle_memphis_4369
#> 71               Mud island West bcycle_memphis_4433
#> 72              Mud Island Plaza bcycle_memphis_4434
#> 73               Mud Island East bcycle_memphis_4435
#> 74 SWT Community College - Union bcycle_memphis_4519
#> 75               Memphis Tourism bcycle_memphis_4520
#> 76                Ratliff - East bcycle_memphis_4535
#> 77                Ratliff - West bcycle_memphis_4536
#> 78              S. 4th St & MLK  bcycle_memphis_4574
#>                                                                                        rental_uris.ios
#> 1  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3864&platform=iOS
#> 2  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3872&platform=iOS
#> 3  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3873&platform=iOS
#> 4  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3874&platform=iOS
#> 5  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3875&platform=iOS
#> 6  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3924&platform=iOS
#> 7  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3925&platform=iOS
#> 8  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3926&platform=iOS
#> 9  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3927&platform=iOS
#> 10 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3928&platform=iOS
#> 11 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3929&platform=iOS
#> 12 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3930&platform=iOS
#> 13 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3931&platform=iOS
#> 14 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3932&platform=iOS
#> 15 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3933&platform=iOS
#> 16 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3934&platform=iOS
#> 17 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3935&platform=iOS
#> 18 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3936&platform=iOS
#> 19 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3937&platform=iOS
#> 20 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3938&platform=iOS
#> 21 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3940&platform=iOS
#> 22 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3941&platform=iOS
#> 23 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3942&platform=iOS
#> 24 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3943&platform=iOS
#> 25 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3944&platform=iOS
#> 26 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3945&platform=iOS
#> 27 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3947&platform=iOS
#> 28 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3948&platform=iOS
#> 29 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3949&platform=iOS
#> 30 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3951&platform=iOS
#> 31 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3952&platform=iOS
#> 32 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3953&platform=iOS
#> 33 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3955&platform=iOS
#> 34 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3957&platform=iOS
#> 35 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3959&platform=iOS
#> 36 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3960&platform=iOS
#> 37 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3961&platform=iOS
#> 38 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3963&platform=iOS
#> 39 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3965&platform=iOS
#> 40 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3966&platform=iOS
#> 41 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3967&platform=iOS
#> 42 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3968&platform=iOS
#> 43 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3970&platform=iOS
#> 44 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3971&platform=iOS
#> 45 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3972&platform=iOS
#> 46 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3973&platform=iOS
#> 47 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3974&platform=iOS
#> 48 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3976&platform=iOS
#> 49 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3984&platform=iOS
#> 50 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3986&platform=iOS
#> 51 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4101&platform=iOS
#> 52 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4196&platform=iOS
#> 53 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4197&platform=iOS
#> 54 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4198&platform=iOS
#> 55 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4201&platform=iOS
#> 56 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4255&platform=iOS
#> 57 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4264&platform=iOS
#> 58 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4269&platform=iOS
#> 59 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4279&platform=iOS
#> 60 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4281&platform=iOS
#> 61 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4285&platform=iOS
#> 62 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4307&platform=iOS
#> 63 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4308&platform=iOS
#> 64 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4309&platform=iOS
#> 65 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4313&platform=iOS
#> 66 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4314&platform=iOS
#> 67 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4328&platform=iOS
#> 68 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4330&platform=iOS
#> 69 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4368&platform=iOS
#> 70 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4369&platform=iOS
#> 71 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4433&platform=iOS
#> 72 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4434&platform=iOS
#> 73 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4435&platform=iOS
#> 74 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4519&platform=iOS
#> 75 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4520&platform=iOS
#> 76 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4535&platform=iOS
#> 77 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4536&platform=iOS
#> 78 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4574&platform=iOS
#>                                                                                        rental_uris.android
#> 1  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3864&platform=android
#> 2  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3872&platform=android
#> 3  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3873&platform=android
#> 4  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3874&platform=android
#> 5  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3875&platform=android
#> 6  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3924&platform=android
#> 7  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3925&platform=android
#> 8  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3926&platform=android
#> 9  https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3927&platform=android
#> 10 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3928&platform=android
#> 11 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3929&platform=android
#> 12 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3930&platform=android
#> 13 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3931&platform=android
#> 14 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3932&platform=android
#> 15 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3933&platform=android
#> 16 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3934&platform=android
#> 17 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3935&platform=android
#> 18 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3936&platform=android
#> 19 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3937&platform=android
#> 20 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3938&platform=android
#> 21 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3940&platform=android
#> 22 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3941&platform=android
#> 23 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3942&platform=android
#> 24 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3943&platform=android
#> 25 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3944&platform=android
#> 26 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3945&platform=android
#> 27 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3947&platform=android
#> 28 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3948&platform=android
#> 29 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3949&platform=android
#> 30 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3951&platform=android
#> 31 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3952&platform=android
#> 32 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3953&platform=android
#> 33 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3955&platform=android
#> 34 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3957&platform=android
#> 35 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3959&platform=android
#> 36 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3960&platform=android
#> 37 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3961&platform=android
#> 38 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3963&platform=android
#> 39 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3965&platform=android
#> 40 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3966&platform=android
#> 41 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3967&platform=android
#> 42 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3968&platform=android
#> 43 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3970&platform=android
#> 44 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3971&platform=android
#> 45 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3972&platform=android
#> 46 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3973&platform=android
#> 47 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3974&platform=android
#> 48 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3976&platform=android
#> 49 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3984&platform=android
#> 50 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_3986&platform=android
#> 51 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4101&platform=android
#> 52 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4196&platform=android
#> 53 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4197&platform=android
#> 54 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4198&platform=android
#> 55 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4201&platform=android
#> 56 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4255&platform=android
#> 57 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4264&platform=android
#> 58 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4269&platform=android
#> 59 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4279&platform=android
#> 60 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4281&platform=android
#> 61 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4285&platform=android
#> 62 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4307&platform=android
#> 63 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4308&platform=android
#> 64 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4309&platform=android
#> 65 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4313&platform=android
#> 66 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4314&platform=android
#> 67 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4328&platform=android
#> 68 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4330&platform=android
#> 69 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4368&platform=android
#> 70 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4369&platform=android
#> 71 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4433&platform=android
#> 72 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4434&platform=android
#> 73 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4435&platform=android
#> 74 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4519&platform=android
#> 75 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4520&platform=android
#> 76 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4535&platform=android
#> 77 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4536&platform=android
#> 78 https://www.bcycle.com/applink?system_id=bcycle_memphis&station_id=bcycle_memphis_4574&platform=android