Grab the station_status feed.

If the specified file does not exist, get_station_status saves the station_status feed for a given city as a .rds object. If the specified file does exist, get_station_status appends the new rows from the current station_status feed to the existing file. Go to `https://github.com/NABSA/gbfs/blob/master/gbfs.md` to see metadata for this dataset.

get_station_status(
  city,
  directory = NULL,
  file = "station_status.rds",
  output = "save"
)

Arguments

city
A character string or a url to an active gbfs.json feed. See get_gbfs_cities for a current list of available cities.

directory
The name of an existing folder or folder to be created, where the feed will will be saved. This argument is only required if `output = "save"` (the default option.) This argument will be ignored, if `output = "return"`

file
Name of an existing file or a new file to be saved. Must end in .rds. This argument will be ignored if `output = "return"`.

output
The type of output method. If `output = "save"`, the object will be saved as an .rds object at the given path. If `output = "return"`, the output will be returned as a dataframe object. Setting `output = "both"` will do both.

Value

The output of this function depends on argument to `output`: Either a saved .rds object generated from the current station_information feed, a dataframe object, or both.

Examples

# we can grab the station status feed for portland, 
# oregon in one of several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "pdx", maybe,) for the `directory` argument 
# instead of `tempdir`.
get_station_status(city = "http://biketownpdx.socialbicycles.com/opendata/station_status.json",
                   directory = tempdir())

# or, instead, just supply the name of 
# the city as a string. 
get_station_status(city = "portland",
                   directory = tempdir())

# instead of saving the output as a file, we can 
# just return the output as a dataframe
get_station_status(city = "portland",
                   output = "return")

#>     station_id num_bikes_available num_bikes_disabled num_docks_available
#> 1     hub_1512                   2                  0                  15
#> 2     hub_1513                   5                  0                   3
#> 3     hub_1514                   5                  0                   7
#> 4     hub_1515                  10                  0                   8
#> 5     hub_1516                   2                  0                  14
#> 6     hub_1517                   4                  0                  10
#> 7     hub_1518                   7                  0                   9
#> 8     hub_1519                   4                  0                  14
#> 9     hub_1521                   8                  0                  10
#> 10    hub_1522                   2                  0                  10
#> 11    hub_1523                   3                  0                  14
#> 12    hub_1524                   6                  0                   6
#> 13    hub_1527                   2                  0                   6
#> 14    hub_1528                   7                  0                  10
#> 15    hub_1529                   4                  0                   7
#> 16    hub_1535                   7                  0                  11
#> 17    hub_1536                   3                  0                   5
#> 18    hub_1537                   5                  0                  16
#> 19    hub_1538                   4                  0                  12
#> 20    hub_1539                   7                  0                   3
#> 21    hub_1540                   2                  0                  16
#> 22    hub_1541                  13                  0                   4
#> 23    hub_1542                   8                  0                   3
#> 24    hub_1543                   6                  0                   4
#> 25    hub_1544                   2                  0                  15
#> 26    hub_1545                   4                  0                  14
#> 27    hub_1546                   7                  0                  10
#> 28    hub_1547                   5                  0                   9
#> 29    hub_1548                   1                  0                  16
#> 30    hub_1549                   2                  0                  16
#> 31    hub_1550                   2                  0                  16
#> 32    hub_1551                   0                  0                  18
#> 33    hub_1552                   5                  0                   7
#> 34    hub_1554                   5                  0                   7
#> 35    hub_1555                   2                  0                  10
#> 36    hub_1556                   5                  0                   9
#> 37    hub_1558                   9                  0                   9
#> 38    hub_1559                   3                  0                  15
#> 39    hub_1560                   6                  0                   9
#> 40    hub_1561                   4                  0                  10
#> 41    hub_1562                   8                  0                   3
#> 42    hub_1563                   4                  0                   9
#> 43    hub_1564                   5                  0                  13
#> 44    hub_1566                   2                  0                  10
#> 45    hub_1567                   3                  0                   8
#> 46    hub_1568                   5                  0                  11
#> 47    hub_1569                   7                  0                  11
#> 48    hub_1570                   6                  0                  21
#> 49    hub_1571                   1                  0                  17
#> 50    hub_1572                   6                  0                  10
#> 51    hub_1573                   0                  0                   6
#> 52    hub_1574                   5                  0                  13
#> 53    hub_1575                  11                  0                   7
#> 54    hub_1577                   1                  0                  19
#> 55    hub_1578                   7                  0                  11
#> 56    hub_1579                   4                  0                   8
#> 57    hub_1580                   2                  0                  16
#> 58    hub_1581                   7                  0                  11
#> 59    hub_1582                   5                  0                  15
#> 60    hub_1583                   4                  0                  14
#> 61    hub_1584                   3                  0                  14
#> 62    hub_1585                   3                  0                   7
#> 63    hub_1586                   4                  0                   9
#> 64    hub_1588                   3                  0                  15
#> 65    hub_1590                   6                  0                   7
#> 66    hub_1591                   4                  0                  15
#> 67    hub_1592                   5                  0                  11
#> 68    hub_1593                   4                  0                  13
#> 69    hub_1594                   1                  0                  11
#> 70    hub_1595                   3                  0                  15
#> 71    hub_1597                   1                  0                  11
#> 72    hub_1598                   8                  0                  10
#> 73    hub_1599                   2                  0                  10
#> 74    hub_1600                  11                  0                   6
#> 75    hub_1601                   0                  0                  11
#> 76    hub_1602                   5                  0                  13
#> 77    hub_1603                   0                  0                   8
#> 78    hub_1604                   2                  0                   6
#> 79    hub_1605                   2                  0                  13
#> 80    hub_1606                   5                  0                  12
#> 81    hub_1607                   2                  0                  16
#> 82    hub_1608                   1                  0                  17
#> 83    hub_1609                   4                  0                  14
#> 84    hub_1611                   2                  0                  16
#> 85    hub_1612                   9                  0                   9
#> 86    hub_1613                   6                  0                   5
#> 87    hub_1614                   3                  0                  10
#> 88    hub_1615                   4                  0                  13
#> 89    hub_1616                   0                  0                   6
#> 90    hub_1617                   4                  0                   8
#> 91    hub_1713                   0                  0                  15
#> 92    hub_2275                   0                  0                  18
#> 93    hub_2276                   0                  0                  24
#> 94    hub_2598                   2                  0                   6
#> 95    hub_2677                   1                  0                  17
#> 96    hub_2678                   0                  0                  12
#> 97    hub_2679                   0                  0                  18
#> 98    hub_2680                   0                  0                  18
#> 99    hub_2681                   0                  0                  18
#> 100   hub_2682                   0                  0                  18
#> 101   hub_2683                   0                  0                  14
#> 102   hub_2684                   0                  0                  20
#> 103   hub_2685                   0                  0                  12
#> 104   hub_2686                   0                  0                  10
#> 105   hub_2687                   3                  0                   5
#> 106   hub_2688                   1                  0                  17
#> 107   hub_2689                   0                  0                  18
#> 108   hub_2695                   0                  0                  18
#> 109   hub_2696                   0                  0                  12
#> 110   hub_2698                   5                  0                  13
#> 111   hub_2707                   8                  0                   9
#> 112   hub_2708                   2                  0                   6
#> 113   hub_2789                   8                  0                   8
#> 114   hub_2798                   9                  0                   9
#> 115   hub_2875                   3                  0                  12
#> 116   hub_3240                   3                  0                   3
#> 117   hub_3929                   4                  0                   7
#> 118   hub_3957                   2                  0                  10
#> 119   hub_4034                   2                  0                   8
#> 120   hub_4050                   3                  0                  11
#> 121   hub_4312                   0                  0                   6
#> 122   hub_4313                   0                  0                   6
#> 123   hub_4315                   0                  0                   6
#> 124   hub_4316                   0                  0                   6
#> 125   hub_4317                   1                  0                   8
#> 126   hub_4318                   0                  0                   6
#> 127   hub_4319                   1                  0                   5
#> 128   hub_4320                   0                  0                  15
#> 129   hub_4321                   0                  0                   9
#> 130   hub_4322                   0                  0                   6
#> 131   hub_4323                   0                  0                  12
#> 132   hub_4324                   0                  0                   6
#> 133   hub_4325                   0                  0                  12
#> 134   hub_4326                   1                  0                   6
#> 135   hub_4327                   1                  0                  11
#> 136   hub_4328                   2                  0                  10
#> 137   hub_4329                   1                  0                  11
#> 138   hub_4368                   1                  0                   7
#> 139   hub_4381                   4                  0                   0
#> 140   hub_4482                   1                  0                  13
#> 141   hub_5277                   1                  0                   5
#> 142   hub_5311                   2                  0                   3
#> 143   hub_5316                   0                  0                   6
#> 144   hub_5357                   5                  0                   0
#> 145   hub_5414                   6                  0                  12
#> 146   hub_5470                   3                  0                   5
#> 147   hub_5473                   0                  0                   8
#> 148   hub_5474                   0                  0                   7
#> 149   hub_5476                   3                  0                   5
#> 150   hub_5491                   1                  0                   3
#> 151   hub_5493                   0                  0                   4
#> 152   hub_5498                   3                  0                   9
#> 153   hub_5567                   1                  0                   4
#> 154   hub_5597                   7                  0                  12
#>     is_installed is_renting is_returning        last_updated year month day
#> 1           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 2           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 3           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 4           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 5           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 6           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 7           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 8           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 9           TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 10          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 11          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 12          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 13          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 14          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 15          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 16          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 17          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 18          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 19          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 20          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 21          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 22          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 23          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 24          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 25          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 26          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 27          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 28          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 29          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 30          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 31          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 32          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 33          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 34          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 35          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 36          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 37          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 38          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 39          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 40          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 41          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 42          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 43          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 44          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 45          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 46          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 47          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 48          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 49          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 50          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 51          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 52          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 53          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 54          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 55          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 56          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 57          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 58          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 59          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 60          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 61          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 62          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 63          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 64          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 65          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 66          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 67          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 68          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 69          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 70          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 71          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 72          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 73          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 74          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 75          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 76          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 77          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 78          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 79          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 80          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 81          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 82          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 83          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 84          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 85          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 86          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 87          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 88          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 89          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 90          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 91          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 92          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 93          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 94          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 95          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 96          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 97          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 98          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 99          TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 100         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 101         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 102         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 103         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 104         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 105         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 106         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 107         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 108         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 109         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 110         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 111         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 112         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 113         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 114         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 115         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 116         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 117         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 118         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 119         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 120         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 121         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 122         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 123         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 124         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 125         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 126         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 127         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 128         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 129         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 130         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 131         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 132         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 133         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 134         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 135         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 136         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 137         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 138         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 139         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 140         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 141         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 142         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 143         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 144         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 145         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 146         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 147         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 148         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 149         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 150         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 151         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 152         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 153         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#> 154         TRUE       TRUE         TRUE 2019-12-11 17:45:39 2019    12  11
#>     hour minute
#> 1     17     45
#> 2     17     45
#> 3     17     45
#> 4     17     45
#> 5     17     45
#> 6     17     45
#> 7     17     45
#> 8     17     45
#> 9     17     45
#> 10    17     45
#> 11    17     45
#> 12    17     45
#> 13    17     45
#> 14    17     45
#> 15    17     45
#> 16    17     45
#> 17    17     45
#> 18    17     45
#> 19    17     45
#> 20    17     45
#> 21    17     45
#> 22    17     45
#> 23    17     45
#> 24    17     45
#> 25    17     45
#> 26    17     45
#> 27    17     45
#> 28    17     45
#> 29    17     45
#> 30    17     45
#> 31    17     45
#> 32    17     45
#> 33    17     45
#> 34    17     45
#> 35    17     45
#> 36    17     45
#> 37    17     45
#> 38    17     45
#> 39    17     45
#> 40    17     45
#> 41    17     45
#> 42    17     45
#> 43    17     45
#> 44    17     45
#> 45    17     45
#> 46    17     45
#> 47    17     45
#> 48    17     45
#> 49    17     45
#> 50    17     45
#> 51    17     45
#> 52    17     45
#> 53    17     45
#> 54    17     45
#> 55    17     45
#> 56    17     45
#> 57    17     45
#> 58    17     45
#> 59    17     45
#> 60    17     45
#> 61    17     45
#> 62    17     45
#> 63    17     45
#> 64    17     45
#> 65    17     45
#> 66    17     45
#> 67    17     45
#> 68    17     45
#> 69    17     45
#> 70    17     45
#> 71    17     45
#> 72    17     45
#> 73    17     45
#> 74    17     45
#> 75    17     45
#> 76    17     45
#> 77    17     45
#> 78    17     45
#> 79    17     45
#> 80    17     45
#> 81    17     45
#> 82    17     45
#> 83    17     45
#> 84    17     45
#> 85    17     45
#> 86    17     45
#> 87    17     45
#> 88    17     45
#> 89    17     45
#> 90    17     45
#> 91    17     45
#> 92    17     45
#> 93    17     45
#> 94    17     45
#> 95    17     45
#> 96    17     45
#> 97    17     45
#> 98    17     45
#> 99    17     45
#> 100   17     45
#> 101   17     45
#> 102   17     45
#> 103   17     45
#> 104   17     45
#> 105   17     45
#> 106   17     45
#> 107   17     45
#> 108   17     45
#> 109   17     45
#> 110   17     45
#> 111   17     45
#> 112   17     45
#> 113   17     45
#> 114   17     45
#> 115   17     45
#> 116   17     45
#> 117   17     45
#> 118   17     45
#> 119   17     45
#> 120   17     45
#> 121   17     45
#> 122   17     45
#> 123   17     45
#> 124   17     45
#> 125   17     45
#> 126   17     45
#> 127   17     45
#> 128   17     45
#> 129   17     45
#> 130   17     45
#> 131   17     45
#> 132   17     45
#> 133   17     45
#> 134   17     45
#> 135   17     45
#> 136   17     45
#> 137   17     45
#> 138   17     45
#> 139   17     45
#> 140   17     45
#> 141   17     45
#> 142   17     45
#> 143   17     45
#> 144   17     45
#> 145   17     45
#> 146   17     45
#> 147   17     45
#> 148   17     45
#> 149   17     45
#> 150   17     45
#> 151   17     45
#> 152   17     45
#> 153   17     45
#> 154   17     45