Grab the system_calendar feed.

get_system_calendar grabs and tidies the system_calendar feed for a given city. Metadata for this dataset can be found at: https://github.com/MobilityData/gbfs/blob/master/gbfs.md

get_system_calendar(
  city,
  directory = NULL,
  file = "system_calendar.rds",
  output = NULL
)

Arguments

city

A character string that can be matched to a gbfs feed. The recommended argument is a system ID supplied in the output of [get_gbfs_cities()], but will also attempt to match to the URL of an active .json feed or city name.

directory

Optional. Path to a folder (or folder to be created) where the feed will be saved.

file

Optional. The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output

Optional. The type of output method. If left as default, this argument is inferred from the directory argument. If output = "save", the object will be saved as an .rds object at # the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on argument to output

and directory. Either a saved .rds object generated from the current feed, a dataframe object, or both.

See also

Examples

# grab the system calendar feed for portland, oregon
get_system_calendar(city = 
"https://gbfs.lyft.com/gbfs/1.1/pdx/en/system_calendar.json",  
                    output = "return")
#>   start_month start_day end_day end_month
#> 1           1         1      31        12