Grab the system_information feed.

get_system_information grabs and tidies the system_information feed for a given city. Metadata for this dataset can be found at: https://github.com/NABSA/gbfs/blob/master/gbfs.md

get_system_information(
  city,
  directory = NULL,
  file = "system_information.rds",
  output = NULL
)

Arguments

city
A character string that can be matched to a gbfs feed. The recommended argument is a system ID supplied in the output of [get_gbfs_cities()], but will also attempt to match to the URL of an active .json feed or city name.

directory
Optional. Path to a folder (or folder to be created) where the feed will be saved.

file
Optional. The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
Optional. The type of output method. If left as default, this argument is inferred from the directory argument. If output = "save", the object will be saved as an .rds object at # the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on argument to output and directory. Either a saved .rds object generated from the current feed, a dataframe object, or both.

See also

Examples

# we can grab the free bike status feed for portland, 
# oregon's bikeshare program in  several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "pdx", maybe,) for the `directory` argument 
# instead of `tempdir`.
get_system_information(city = 
"https://gbfs.biketownpdx.com/gbfs/en/system_information.json",  
                       directory = tempdir())
                    
# or, instead, just supply the name of 
# the city as a string and return the output as a dataframe
get_system_information(city = "biketown_pdx",  
                       output = "return")

#>   name            timezone language system_id
#> 1 Lyft America/Los_Angeles       en  lyft_pdx