Grab the system_pricing_plans feed.

get_system_pricing_plans grabs and tidies the system_pricing_plans feed for a given city. Metadata for this dataset can be found at: https://github.com/NABSA/gbfs/blob/master/gbfs.md

get_system_pricing_plans(
  city,
  directory = NULL,
  file = "system_pricing_plans.rds",
  output = NULL
)

Arguments

city
A character string that can be matched to a city or a url to an active gbfs .json feed. See [get_gbfs_cities()] for a current list of available cities.

directory
Optional. Path to a folder (or folder to be created) where the feed will be saved.

file
Optional. The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
Optional. The type of output method. If left as default, this argument is inferred from the directory argument. If output = "save", the object will be saved as an .rds object at # the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on argument to output and directory. Either a saved .rds object generated from the current feed, a dataframe object, or both.

See also

Examples

# we can grab the system pricing plans feed for memphis, 
# tennessee in one of several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "memphis", maybe,) for the `directory` argument 
# instead of `tempdir`.
get_system_pricing_plans(city = 
"https://gbfs.bcycle.com/bcycle_memphis/system_pricing_plans.json",  
                         directory = tempdir())
                    
# or, instead, just supply the name of 
# the city as a string. 
get_system_pricing_plans(city = "memphis",
                         directory = tempdir())
                    
# instead of saving the output as a file, we can 
# just return the output as a dataframe
get_system_pricing_plans(city = "memphis",  
                         output = "return")

#>   price is_taxable                         url
#> 1    90          0 explorebikeshare.bcycle.com
#> 2    10          0 explorebikeshare.bcycle.com
#> 3    15          0 explorebikeshare.bcycle.com
#> 4     5          0 explorebikeshare.bcycle.com
#>                                                                                                                                                                                               description
#> 1   BEST VALUE - Get unlimited 60-minute rides for a full year! Rides longer than 60 min incur a fee of $1.25 per 15 minutes. Avoid added fees - lock at any bike station within 60 min & check back out.
#> 2 Enjoy unlimited rides, up to 60 minutes each, for SEVEN days. Rides longer than 60 min incur a fee of $1.25 per 15 minutes. Avoid added fees - lock at any bike station within 60 min & check back out.
#> 3   MOST POPULAR - Enjoy unlimited 60-minute rides for a month! Rides longer than 60 min incur a fee of $1.25 per 15 minutes. Avoid added fees - lock at any bike station within 60 min & check back out.
#> 4                                                 Enjoy your two-hour ride! Add additional time for $0.20 per minute.  Avoid parking fees by locking your bike securely at an Explore Bike Share station.
#>   currency        name            plan_id
#> 1      USD      Annual bcycle_memphis_423
#> 2      USD     Weekly  bcycle_memphis_424
#> 3      USD     Monthly bcycle_memphis_425
#> 4      USD 2 Hour Pass bcycle_memphis_648