Grab the system_regions feed.

get_system_regions grabs and tidies the system_regions feed for a given city. Go to `https://github.com/NABSA/gbfs/blob/master/gbfs.md` to see metadata for this dataset.

get_system_regions(
  city,
  directory = NULL,
  file = "system_regions.rds",
  output = "save"
)

Arguments

city
A character string that can be matched to a city or a url to an active gbfs .json feed. See get_gbfs_cities for a current list of available cities.

directory
The name of an existing folder or folder to be created, where the feed will will be saved. This argument is ignored if output = "return"

file
The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
The type of output method. If output = "save", the object will be saved as an .rds object at the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on the argument to output: Either a saved .rds object generated from the current system_regions feed, a dataframe object, or both.

Examples

# we can grab the system regions feed for portland, 
# oregon in one of several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "pdx", maybe,) for the `directory` argument 
# instead of `tempdir`.
get_system_regions(city = "http://biketownpdx.socialbicycles.com/opendata/system_regions.json",
                   directory = tempdir())

# or, instead, just supply the name of 
# the city as a string. 
get_system_regions(city = "portland",
                   directory = tempdir())

# instead of saving the output as a file, we can 
# just return the output as a dataframe
get_system_regions(city = "portland",
                   output = "return")

#>    region_id         name
#> 1 region_241 BIKETOWN PDX