Get dataframe of bikeshare feeds released by a city

Of the different types of feeds supplied by the gbfs, some are required, some are conditionally required, and some are optional. This function grabs a list of each of the feeds supplied by a given city, as well as the URLs to access them.

get_which_gbfs_feeds(city)

Source

North American Bikeshare Association, General Bikeshare Feed Specification https://github.com/MobilityData/gbfs/blob/master/gbfs.md

Arguments

city

A character string that can be matched to a gbfs feed. The recommended argument is a system ID supplied in the output of [get_gbfs_cities()], but will also attempt to match to the URL of an active .json feed or city name.

Value

A data.frame containing the feeds supplied by a city. . The `feed` column supplies the name of the relevant .json feeds, while the entries in the `URL` column supply the feeds themselves.

Examples

# grab all of the feeds released by portland
get_which_gbfs_feeds(city = "biketown_pdx")
#>                                                                url
#> 1                     https://gbfs.lyft.com/gbfs/2.3/pdx/gbfs.json
#> 2    https://gbfs.lyft.com/gbfs/2.3/pdx/en/system_information.json
#> 3   https://gbfs.lyft.com/gbfs/2.3/pdx/en/station_information.json
#> 4        https://gbfs.lyft.com/gbfs/2.3/pdx/en/station_status.json
#> 5      https://gbfs.lyft.com/gbfs/2.3/pdx/en/free_bike_status.json
#> 6          https://gbfs.lyft.com/gbfs/2.3/pdx/en/system_hours.json
#> 7       https://gbfs.lyft.com/gbfs/2.3/pdx/en/system_calendar.json
#> 8        https://gbfs.lyft.com/gbfs/2.3/pdx/en/system_regions.json
#> 9  https://gbfs.lyft.com/gbfs/2.3/pdx/en/system_pricing_plans.json
#> 10        https://gbfs.lyft.com/gbfs/2.3/pdx/en/system_alerts.json
#> 11        https://gbfs.lyft.com/gbfs/2.3/pdx/en/gbfs_versions.json
#> 12        https://gbfs.lyft.com/gbfs/2.3/pdx/en/vehicle_types.json
#>                    name
#> 1                  gbfs
#> 2    system_information
#> 3   station_information
#> 4        station_status
#> 5      free_bike_status
#> 6          system_hours
#> 7       system_calendar
#> 8        system_regions
#> 9  system_pricing_plans
#> 10        system_alerts
#> 11        gbfs_versions
#> 12        vehicle_types