Grab the free_bike_status feed.

If the specified file does not exist, get_free_bike_status saves the free_bike_status feed for a given city as a .rds object. If the specified file does exist, get_free_bike_status appends the current free_bike_status feed to the existing file. The resulting dataframe can alternatively be returned (rather than saved) using the `output` argument. Go to `https://github.com/NABSA/gbfs/blob/master/gbfs.md` to see metadata for this dataset.

get_free_bike_status(
  city,
  directory = NULL,
  file = "free_bike_status.rds",
  output = "save"
)

Arguments

city
A character string or a url to an active gbfs.json feed. See get_gbfs_cities for a current list of available cities.

directory
The name of an existing folder or folder to be created, where the feed will will be saved. This argument is only required if `output = "save"` (the default option.) If `output = "return"`, this argument will be ignored.

file
Name of an existing file or a new file to be saved. Must end in .rds. This argument will be ignored if `output = "return"`.

output
The type of output method. If `output = "save"`, the object will be saved as an .rds object at the given path. If `output = "return"`, the output will be returned as a dataframe object. Setting `output = "both"` will do both.

Value

The output of the function depends on argument to `output`: Either a saved .rds object generated from the current station_information feed, a dataframe object, or both.

Examples

# we can grab the free bike status feed for portland, 
# oregon in one of several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "pdx", maybe,) for the `directory` argument 
# instead of `tempdir`.
get_free_bike_status(city = "http://biketownpdx.socialbicycles.com/opendata/free_bike_status.json",
                     directory = tempdir())

# or, instead, just supply the name of 
# the city as a string. 
get_free_bike_status(city = "portland",
                     directory = tempdir())

# instead of saving the output as a file, we can 
# just return the output as a dataframe
get_free_bike_status(city = "portland",
                     output = "return")

#>        bike_id                                      name       lon      lat
#> 1    bike_6606                             0836 BIKETOWN -122.6616 45.56117
#> 2    bike_6806                           0761 BIKETOWN   -122.6549 45.50868
#> 3    bike_6661                             0768 BIKETOWN -122.6476 45.54044
#> 4    bike_6852                           0820 BIKETOWN   -122.6466 45.53578
#> 5    bike_6446                             0258 BIKETOWN -122.6050 45.54819
#> 6    bike_6094                             0531 BIKETOWN -122.6579 45.54135
#> 7   bike_19526                            0215 BIKETOWN  -122.6570 45.51716
#> 8    bike_6035                             0575 BIKETOWN -122.6800 45.51702
#> 9    bike_6266                             0162 BIKETOWN -122.6839 45.52802
#> 10   bike_6636                             0825 BIKETOWN -122.6164 45.53758
#> 11   bike_6500                             0136 BIKETOWN -122.6480 45.49828
#> 12   bike_6855                             0539 BIKETOWN -122.6449 45.50067
#> 13   bike_6043                         0639 AIX BIKETOWN -122.6057 45.54006
#> 14   bike_6182                             0317 BIKETOWN -122.6667 45.55776
#> 15   bike_6503                             0326 BIKETOWN -122.6578 45.49979
#> 16   bike_6501                         0309 AIX BIKETOWN -122.6228 45.53238
#> 17   bike_7181                             0128 BIKETOWN -122.6403 45.50489
#> 18   bike_6363                            0684 BIKETOWN  -122.6790 45.51984
#> 19   bike_6095                             0073 BIKETOWN -122.6480 45.49832
#> 20   bike_6542                             0905 BIKETOWN -122.6654 45.51116
#> 21   bike_6424                             0466 BIKETOWN -122.6786 45.52335
#> 22  bike_29251                             0135 BIKETOWN -122.6597 45.51758
#> 23   bike_7112                             0388 BIKETOWN -122.6926 45.53356
#> 24   bike_6440                             0491 BIKETOWN -122.6438 45.52883
#> 25   bike_7182                             0327 BIKETOWN -122.6297 45.49324
#> 26   bike_6575                             0053 BIKETOWN -122.6352 45.52659
#> 27   bike_7104                             0824 BIKETOWN -122.6756 45.52553
#> 28   bike_6067                             0298 BIKETOWN -122.6486 45.50211
#> 29   bike_7145                             0572 BIKETOWN -122.6769 45.52722
#> 30   bike_7206                             0553 BIKETOWN -122.6563 45.51228
#> 31   bike_6332                             0109 BIKETOWN -122.6818 45.53140
#> 32   bike_6538                             0534 BIKETOWN -122.6769 45.51569
#> 33   bike_6870                             0779 BIKETOWN -122.6604 45.52018
#> 34   bike_7233                             0041 BIKETOWN -122.6734 45.54095
#> 35   bike_6629                             0342 BIKETOWN -122.6494 45.49826
#> 36   bike_6379                             0225 BIKETOWN -122.6755 45.52571
#> 37   bike_6458                             0924 BIKETOWN -122.6749 45.54711
#> 38   bike_6238                            0774 BIKETOWN  -122.6735 45.51989
#> 39   bike_6337                             0422 BIKETOWN -122.6668 45.53977
#> 40   bike_6432                             0660 BIKETOWN -122.6754 45.51326
#> 41   bike_6415                             0441 BIKETOWN -122.6746 45.51777
#> 42   bike_7249                             0201 BIKETOWN -122.6773 45.55416
#> 43   bike_6631                             0332 BIKETOWN -122.6800 45.51693
#> 44   bike_6827                             0440 BIKETOWN -122.6607 45.52268
#> 45   bike_6647                             0386 BIKETOWN -122.6788 45.52115
#> 46   bike_6122                             0040 BIKETOWN -122.6212 45.53556
#> 47   bike_6083                             0468 BIKETOWN -122.6843 45.52048
#> 48   bike_6247                             0593 BIKETOWN -122.6544 45.52041
#> 49  bike_24811                             0033 BIKETOWN -122.6595 45.50872
#> 50   bike_6552                             0849 BIKETOWN -122.6760 45.52516
#> 51   bike_6577                             0596 BIKETOWN -122.6805 45.52280
#> 52   bike_7108                             0235 BIKETOWN -122.6665 45.50928
#> 53  bike_24823                            0557 BIKETOWN  -122.6802 45.53038
#> 54   bike_7298          0980 DESIGN BIKE, N PAUL BUNYAN  -122.6918 45.52916
#> 55  bike_19242                            0009 BIKETOWN  -122.6761 45.50669
#> 56   bike_6174                             0116 BIKETOWN -122.6781 45.52740
#> 57   bike_6166                             0036 BIKETOWN -122.6909 45.53348
#> 58  bike_28387                             0689 BIKETOWN -122.7067 45.53382
#> 59   bike_6389                             0176 BIKETOWN -122.6256 45.52021
#> 60  bike_24813                             0038 BIKETOWN -122.6928 45.52034
#> 61   bike_7170                             0019 BIKETOWN -122.6786 45.53063
#> 62   bike_7265                             0370 BIKETOWN -122.6094 45.49613
#> 63   bike_7194                             0356 BIKETOWN -122.6997 45.53485
#> 64   bike_7136                             0502 BIKETOWN -122.6834 45.52513
#> 65   bike_6031                             0204 BIKETOWN -122.6480 45.49832
#> 66   bike_6555                             0322 BIKETOWN -122.6829 45.53534
#> 67   bike_7124                             0222 BIKETOWN -122.6602 45.51729
#> 68   bike_7315                0966 WOMEN’S HISTORY MONTH -122.6744 45.52440
#> 69   bike_6516                             0060 BIKETOWN -122.6872 45.52280
#> 70   bike_6028                             0927 BIKETOWN -122.6954 45.52333
#> 71   bike_6097                             0439 BIKETOWN -122.6971 45.53840
#> 72   bike_6198                             0739 BIKETOWN -122.6928 45.52420
#> 73   bike_6861                             0541 BIKETOWN -122.6804 45.52257
#> 74   bike_6386                             0654 BIKETOWN -122.6988 45.53460
#> 75   bike_6390                             0107 BIKETOWN -122.6933 45.52913
#> 76   bike_6110                             0365 BIKETOWN -122.6842 45.52054
#> 77  bike_19237                             0133 BIKETOWN -122.6495 45.49394
#> 78   bike_7325                          0969 ASCEND BIKE -122.6785 45.52321
#> 79   bike_6059                             0816 BIKETOWN -122.6828 45.51146
#> 80   bike_6041                             0369 BIKETOWN -122.6842 45.52726
#> 81   bike_7263                             0001 BIKETOWN -122.6601 45.51730
#> 82   bike_6112                             0121 BIKETOWN -122.6970 45.53780
#> 83   bike_7278                             0208 BIKETOWN -122.6474 45.52722
#> 84   bike_6624                             0413 BIKETOWN -122.6886 45.53150
#> 85   bike_6391                             0113 BIKETOWN -122.6449 45.50064
#> 86   bike_6165                             0613 BIKETOWN -122.6799 45.51656
#> 87   bike_6026                             0442 BIKETOWN -122.6447 45.53453
#> 88   bike_7176                             0484 BIKETOWN -122.6832 45.52640
#> 89   bike_7331                       1015 RIP CITY RIDER -122.6931 45.51818
#> 90   bike_6350                             0800 BIKETOWN -122.6626 45.50997
#> 91   bike_6406                             0812 BIKETOWN -122.6578 45.51348
#> 92   bike_6652                             0518 BIKETOWN -122.7110 45.53601
#> 93   bike_6825                             0101 BIKETOWN -122.6233 45.52361
#> 94   bike_6640                            0214 BIKETOWN  -122.6439 45.52888
#> 95   bike_6445                            0504 BIKETOWN  -122.6606 45.52713
#> 96   bike_6109                             0894 BIKETOWN -122.6648 45.54192
#> 97   bike_7175                            0333 BIKETOWN  -122.6910 45.52324
#> 98   bike_6470                             0320 BIKETOWN -122.6760 45.50667
#> 99   bike_6398                             0375 BIKETOWN -122.6960 45.53446
#> 100  bike_6612                             0794 BIKETOWN -122.6732 45.51435
#> 101  bike_7151                             0129 BIKETOWN -122.6748 45.52173
#> 102  bike_7158                            0324 BIKETOWN  -122.6747 45.52180
#> 103  bike_7250                             0742 BIKETOWN -122.6904 45.52665
#> 104  bike_7355         1006 DESIGN BIKE, NE CYCLING LIFE -122.6849 45.53402
#> 105  bike_6620                             0754 BIKETOWN -122.6153 45.51194
#> 106  bike_7232                            0030 BIKETOWN  -122.6495 45.49395
#> 107  bike_6232                             0870 BIKETOWN -122.6181 45.50342
#> 108  bike_6168                             0888 BIKETOWN -122.6474 45.51175
#> 109  bike_7496                             0270 BIKETOWN -122.6454 45.53140
#> 110  bike_7155                             0869 BIKETOWN -122.6619 45.51313
#> 111  bike_6483                             0606 BIKETOWN -122.6922 45.51873
#> 112 bike_24814                             0398 BIKETOWN -122.6725 45.52417
#> 113  bike_6327                             0715 BIKETOWN -122.6832 45.52584
#> 114  bike_6244                             0886 BIKETOWN -122.6115 45.50665
#> 115  bike_6528                             0646 BIKETOWN -122.6783 45.52266
#> 116  bike_6385                            0443 BIKETOWN  -122.6466 45.50353
#> 117  bike_6653                             0651 BIKETOWN -122.6476 45.51748
#> 118  bike_7147                             0170 BIKETOWN -122.6832 45.52579
#> 119  bike_6859                             0550 BIKETOWN -122.6616 45.56358
#> 120  bike_5996                             0804 BIKETOWN -122.6859 45.53249
#> 121  bike_6221                             0093 BIKETOWN -122.6736 45.52011
#> 122  bike_7171                             0094 BIKETOWN -122.6733 45.52804
#> 123  bike_7161                            0171 BIKETOWN  -122.6498 45.50342
#> 124  bike_6054                             0744 BIKETOWN -122.6567 45.52188
#> 125  bike_7143                             0763 BIKETOWN -122.6692 45.49353
#> 126  bike_6838                            0358 BIKETOWN  -122.6812 45.52936
#> 127  bike_6369                             0361 BIKETOWN -122.6769 45.51776
#> 128  bike_6156                             0615 BIKETOWN -122.6644 45.51152
#> 129  bike_7279                            0578 BIKETOWN  -122.6800 45.52096
#> 130  bike_6314                             0617 BIKETOWN -122.6637 45.52099
#> 131  bike_6809                             0103 BIKETOWN -122.6769 45.51790
#> 132  bike_6495                             0702 BIKETOWN -122.6810 45.52729
#> 133  bike_6462                            0275 BIKETOWN  -122.6710 45.52325
#> 134  bike_7153                             0156 BIKETOWN -122.6761 45.51969
#> 135  bike_6665                             0246 BIKETOWN -122.6596 45.52388
#> 136  bike_7390 0961 DESIGN BIKE, WATERFRONT WANDERER SW  -122.6487 45.48984
#> 137  bike_6317                             0854 BIKETOWN -122.6748 45.52171
#> 138  bike_7268                            0016 BIKETOWN  -122.6804 45.52269
#> 139  bike_6281                             0213 BIKETOWN -122.6844 45.51388
#> 140  bike_7137                             0137 BIKETOWN -122.6709 45.52302
#> 141  bike_7376            0946 DESIGN BIKE, EXPLORIN' SE -122.6758 45.51993
#> 142 bike_33605                         0448 AIX BIKETOWN -122.6758 45.52970
#> 143  bike_6019                             0723 BIKETOWN -122.6802 45.53048
#> 144  bike_6572                             0665 BIKETOWN -122.6544 45.50083
#> 145  bike_7341                    1021 BLAZERS 50th BIKE -122.6895 45.53414
#> 146  bike_7142                             0149 BIKETOWN -122.6588 45.51639
#> 147  bike_7121                             0117 BIKETOWN -122.6794 45.51991
#> 148  bike_6786                           0529 BIKETOWN   -122.6820 45.52005
#> 149  bike_7313                         0985 ASCEND BIKE  -122.6770 45.50165
#> 150  bike_7255                             0013 BIKETOWN -122.6660 45.50863
#> 151  bike_6885                             0682 BIKETOWN -122.6629 45.51119
#> 152 bike_23242                             0893 BIKETOWN -122.6746 45.51780
#> 153  bike_6593                             0738 BIKETOWN -122.7003 45.52394
#> 154  bike_7354         0998 DESIGN BIKE, NE CYCLING LIFE -122.6758 45.52972
#> 155  bike_7282                             0020 BIKETOWN -122.6664 45.51505
#> 156  bike_6312                             0892 BIKETOWN -122.6688 45.53015
#> 157  bike_7254                            1031 BIKETOWN  -122.6746 45.51781
#> 158  bike_6034                             0899 BIKETOWN -122.6499 45.52778
#> 159  bike_6197                             0372 BIKETOWN -122.6748 45.52195
#> 160  bike_6777                             0266 BIKETOWN -122.6786 45.51977
#> 161  bike_7110                             0185 BIKETOWN -122.6837 45.53175
#> 162  bike_6874                             0862 BIKETOWN -122.6712 45.52327
#> 163  bike_6408                             0223 BIKETOWN -122.6785 45.52365
#> 164  bike_6834                             0191 BIKETOWN -122.6088 45.53362
#> 165  bike_6052                             0494 BIKETOWN -122.6919 45.52915
#> 166  bike_6265                             0188 BIKETOWN -122.6393 45.50486
#> 167  bike_6517                             0476 BIKETOWN -122.6781 45.52243
#> 168  bike_6099                             0481 BIKETOWN -122.6774 45.51688
#> 169 bike_24816                             0515 BIKETOWN -122.6980 45.53348
#> 170  bike_6269                             0595 BIKETOWN -122.6777 45.51454
#> 171  bike_7193                             0513 BIKETOWN -122.6805 45.52255
#> 172  bike_6228                             0580 BIKETOWN -122.6763 45.51895
#> 173  bike_6435                             0197 BIKETOWN -122.6784 45.52374
#> 174 bike_29385                             0568 BIKETOWN -122.6517 45.51928
#> 175  bike_6045                             0433 BIKETOWN -122.6726 45.50272
#> 176  bike_6788                             0444 BIKETOWN -122.6607 45.51956
#> 177  bike_7119                            0291 BIKETOWN  -122.6363 45.53404
#> 178  bike_6559                             0394 BIKETOWN -122.6438 45.52884
#> 179  bike_7326                       1012 VETERANS BIKE  -122.6425 45.51221
#> 180  bike_6660                         0335  XC BIKETOWN -122.6832 45.52590
#> 181  bike_6294                             0434 BIKETOWN -122.6627 45.52432
#> 182  bike_6176                             0587 BIKETOWN -122.6786 45.51962
#> 183  bike_6534                             0922 BIKETOWN -122.6548 45.50481
#> 184  bike_6662                             0793 BIKETOWN -122.6798 45.51899
#> 185  bike_6818                             0125 BIKETOWN -122.6802 45.52519
#> 186  bike_6224                             0727 BIKETOWN -122.6418 45.52215
#> 187  bike_6340                             0664 BIKETOWN -122.6816 45.52925
#> 188  bike_7323           0965 DESIGN BIKE, N PAUL BUNYAN -122.6804 45.51807
#> 189  bike_6243                             0579 BIKETOWN -122.6817 45.52059
#> 190  bike_6414                             0363 BIKETOWN -122.6882 45.53435
#> 191  bike_6301                             0247 BIKETOWN -122.6839 45.52799
#> 192  bike_6862                             0226 BIKETOWN -122.6558 45.51965
#> 193  bike_7381       0937 NATIVE AMERICAN HERITAGE MONTH -122.6834 45.53070
#> 194  bike_6480                         0726 AIX BIKETOWN -122.6257 45.50570
#> 195  bike_6039                             0865 BIKETOWN -122.6788 45.51898
#> 196  bike_6091                             0875 BIKETOWN -122.6931 45.52724
#> 197  bike_7267                             0011 BIKETOWN -122.6450 45.53288
#> 198  bike_6664                             0809 BIKETOWN -122.6722 45.50355
#> 199  bike_6663                             0569 BIKETOWN -122.6745 45.51809
#> 200  bike_7373            0944 DESIGN BIKE, EXPLORIN' SE -122.6752 45.52433
#> 201  bike_7382            0933 DESIGN BIKE, EXPLORIN’ SE -122.6194 45.50545
#> 202  bike_6632                             0278 BIKETOWN -122.6987 45.53366
#> 203  bike_6089                             0837 BIKETOWN -122.6805 45.53110
#> 204  bike_6426                             0582 BIKETOWN -122.6843 45.52662
#> 205  bike_7347                1028 LATINX HERITAGE MONTH -122.6948 45.52987
#> 206  bike_6646                             0644 BIKETOWN -122.6857 45.53399
#> 207  bike_6816                             0706 BIKETOWN -122.6586 45.51557
#> 208  bike_7384            0957 DESIGN BIKE, EXPLORIN' SE -122.6866 45.52876
#> 209  bike_6330                             0671 BIKETOWN -122.6820 45.52011
#> 210  bike_6070                             0847 BIKETOWN -122.6719 45.52032
#> 211  bike_6412                             0376 BIKETOWN -122.6844 45.52735
#> 212 bike_30041         1014 DESIGN BIKE, NE CYCLING LIFE -122.6845 45.53382
#> 213  bike_7307                         0979 AIR TRAINER  -122.6553 45.50865
#> 214  bike_6085                             0354 BIKETOWN -122.6760 45.50663
#> 215  bike_6847                             0716 BIKETOWN -122.6639 45.51300
#> 216  bike_6421                             0118 BIKETOWN -122.6476 45.52312
#> 217  bike_6461                             0166 BIKETOWN -122.6493 45.49396
#> 218  bike_7179                             0431 BIKETOWN -122.6765 45.52278
#> 219  bike_6588                             0338 BIKETOWN -122.6357 45.50131
#> 220  bike_7247                             0254 BIKETOWN -122.6807 45.52445
#> 221  bike_6490                             0915 BIKETOWN -122.6787 45.51888
#> 222  bike_6284                             0878 BIKETOWN -122.6783 45.51897
#> 223  bike_6167                             0284 BIKETOWN -122.6821 45.53183
#> 224  bike_6044                             0499 BIKETOWN -122.6804 45.53020
#> 225  bike_7409                             0209 BIKETOWN -122.6786 45.53225
#> 226  bike_6336                             0805 BIKETOWN -122.6920 45.52020
#> 227  bike_6810                             0097 BIKETOWN -122.6722 45.52396
#> 228  bike_6208                             0736 BIKETOWN -122.6508 45.51430
#> 229  bike_6442                             0249 BIKETOWN -122.6836 45.51915
#> 230  bike_6423                             0527 BIKETOWN -122.6876 45.53560
#> 231  bike_7394                1029 LATINX HERITAGE MONTH -122.6600 45.50939
#> 232  bike_6533                             0610 BIKETOWN -122.6785 45.52196
#> 233  bike_7125                             0172 BIKETOWN -122.6499 45.50342
#> 234  bike_6889                             0063 BIKETOWN -122.6641 45.53592
#> 235  bike_6473                             0656 BIKETOWN -122.6965 45.52524
#> 236  bike_6668                             0628 BIKETOWN -122.6754 45.49372
#> 237  bike_6821                             0198 BIKETOWN -122.6809 45.52371
#> 238  bike_6046                             0452 BIKETOWN -122.6749 45.52178
#> 239  bike_6450                            0556 BIKETOWN  -122.6922 45.52286
#> 240  bike_7198                             0018 BIKETOWN -122.6834 45.53026
#> 241  bike_6658                             0619 BIKETOWN -122.6753 45.51528
#> 242  bike_7116                             0399 BIKETOWN -122.6983 45.53482
#> 243  bike_7140                             0228 BIKETOWN -122.6818 45.52571
#> 244  bike_6634                            0583 BIKETOWN  -122.6813 45.51992
#> 245  bike_6334                             0253 BIKETOWN -122.6642 45.50859
#> 246  bike_6666                             0349 BIKETOWN -122.6760 45.50668
#> 247  bike_7374  0935 DESIGN BIKE, WATERFRONT WANDERER SW -122.6586 45.51738
#> 248  bike_6805                             0598 BIKETOWN -122.6708 45.49576
#> 249  bike_6539                             0917 BIKETOWN -122.6925 45.51957
#> 250  bike_7501                             0592 BIKETOWN -122.6396 45.50514
#> 251  bike_7240                             0007 BIKETOWN -122.6756 45.52957
#> 252  bike_6359                             0051 BIKETOWN -122.6985 45.53339
#> 253  bike_5997                             0150 BIKETOWN -122.6872 45.52148
#> 254  bike_5992                             0909 BIKETOWN -122.6820 45.52003
#> 255  bike_6785                          0074 xc BIKETOWN -122.6474 45.51128
#> 256  bike_6331                             0043 BIKETOWN -122.6612 45.55461
#> 257 bike_19240                          0042 xc BIKETOWN -122.6449 45.53302
#> 258  bike_7245                             0551 BIKETOWN -122.6836 45.52641
#> 259  bike_7388                         0945 ABILITY BIKE -122.6760 45.51783
#> 260 bike_14548                            0803 BIKETOWN  -122.6760 45.50668
#> 261  bike_7107                          0755 xc BIKETOWN -122.6839 45.52658
#> 262  bike_6894                             0252 BIKETOWN -122.6872 45.52145
#> 263  bike_7369                    0950 BLAZERS 50th BIKE -122.6780 45.52736
#> 264  bike_7364       0956 NATIVE AMERICAN HERITAGE MONTH -122.6855 45.53125
#> 265  bike_6642                             0562 BIKETOWN -122.6787 45.51863
#>     is_reserved is_disabled jump_ebike_battery_level        last_updated year
#> 1         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 2         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 3         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 4         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 5         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 6         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 7         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 8         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 9         FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 10        FALSE       FALSE                      55% 2019-12-11 17:43:37 2019
#> 11        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 12        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 13        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 14        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 15        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 16        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 17        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 18        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 19        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 20        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 21        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 22        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 23        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 24        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 25        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 26        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 27        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 28        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 29        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 30        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 31        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 32        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 33        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 34        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 35        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 36        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 37        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 38        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 39        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 40        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 41        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 42        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 43        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 44        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 45        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 46        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 47        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 48        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 49        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 50        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 51        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 52        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 53        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 54        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 55        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 56        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 57        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 58        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 59        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 60        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 61        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 62        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 63        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 64        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 65        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 66        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 67        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 68        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 69        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 70        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 71        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 72        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 73        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 74        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 75        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 76        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 77        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 78        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 79        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 80        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 81        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 82        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 83        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 84        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 85        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 86        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 87        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 88        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 89        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 90        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 91        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 92        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 93        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 94        FALSE       FALSE                      93% 2019-12-11 17:43:37 2019
#> 95        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 96        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 97        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 98        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 99        FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 100       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 101       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 102       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 103       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 104       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 105       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 106       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 107       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 108       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 109       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 110       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 111       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 112       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 113       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 114       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 115       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 116       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 117       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 118       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 119       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 120       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 121       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 122       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 123       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 124       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 125       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 126       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 127       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 128       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 129       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 130       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 131       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 132       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 133       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 134       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 135       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 136       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 137       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 138       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 139       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 140       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 141       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 142       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 143       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 144       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 145       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 146       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 147       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 148       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 149       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 150       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 151       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 152       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 153       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 154       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 155       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 156       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 157       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 158       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 159       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 160       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 161       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 162       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 163       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 164       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 165       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 166       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 167       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 168       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 169       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 170       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 171       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 172       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 173       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 174       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 175       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 176       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 177       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 178       FALSE       FALSE                      86% 2019-12-11 17:43:37 2019
#> 179       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 180       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 181       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 182       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 183       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 184       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 185       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 186       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 187       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 188       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 189       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 190       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 191       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 192       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 193       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 194       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 195       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 196       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 197       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 198       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 199       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 200       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 201       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 202       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 203       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 204       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 205       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 206       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 207       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 208       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 209       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 210       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 211       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 212       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 213       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 214       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 215       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 216       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 217       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 218       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 219       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 220       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 221       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 222       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 223       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 224       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 225       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 226       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 227       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 228       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 229       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 230       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 231       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 232       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 233       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 234       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 235       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 236       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 237       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 238       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 239       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 240       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 241       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 242       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 243       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 244       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 245       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 246       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 247       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 248       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 249       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 250       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 251       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 252       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 253       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 254       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 255       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 256       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 257       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 258       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 259       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 260       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 261       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 262       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 263       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 264       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#> 265       FALSE       FALSE                     <NA> 2019-12-11 17:43:37 2019
#>     month day hour minute
#> 1      12  11   17     43
#> 2      12  11   17     43
#> 3      12  11   17     43
#> 4      12  11   17     43
#> 5      12  11   17     43
#> 6      12  11   17     43
#> 7      12  11   17     43
#> 8      12  11   17     43
#> 9      12  11   17     43
#> 10     12  11   17     43
#> 11     12  11   17     43
#> 12     12  11   17     43
#> 13     12  11   17     43
#> 14     12  11   17     43
#> 15     12  11   17     43
#> 16     12  11   17     43
#> 17     12  11   17     43
#> 18     12  11   17     43
#> 19     12  11   17     43
#> 20     12  11   17     43
#> 21     12  11   17     43
#> 22     12  11   17     43
#> 23     12  11   17     43
#> 24     12  11   17     43
#> 25     12  11   17     43
#> 26     12  11   17     43
#> 27     12  11   17     43
#> 28     12  11   17     43
#> 29     12  11   17     43
#> 30     12  11   17     43
#> 31     12  11   17     43
#> 32     12  11   17     43
#> 33     12  11   17     43
#> 34     12  11   17     43
#> 35     12  11   17     43
#> 36     12  11   17     43
#> 37     12  11   17     43
#> 38     12  11   17     43
#> 39     12  11   17     43
#> 40     12  11   17     43
#> 41     12  11   17     43
#> 42     12  11   17     43
#> 43     12  11   17     43
#> 44     12  11   17     43
#> 45     12  11   17     43
#> 46     12  11   17     43
#> 47     12  11   17     43
#> 48     12  11   17     43
#> 49     12  11   17     43
#> 50     12  11   17     43
#> 51     12  11   17     43
#> 52     12  11   17     43
#> 53     12  11   17     43
#> 54     12  11   17     43
#> 55     12  11   17     43
#> 56     12  11   17     43
#> 57     12  11   17     43
#> 58     12  11   17     43
#> 59     12  11   17     43
#> 60     12  11   17     43
#> 61     12  11   17     43
#> 62     12  11   17     43
#> 63     12  11   17     43
#> 64     12  11   17     43
#> 65     12  11   17     43
#> 66     12  11   17     43
#> 67     12  11   17     43
#> 68     12  11   17     43
#> 69     12  11   17     43
#> 70     12  11   17     43
#> 71     12  11   17     43
#> 72     12  11   17     43
#> 73     12  11   17     43
#> 74     12  11   17     43
#> 75     12  11   17     43
#> 76     12  11   17     43
#> 77     12  11   17     43
#> 78     12  11   17     43
#> 79     12  11   17     43
#> 80     12  11   17     43
#> 81     12  11   17     43
#> 82     12  11   17     43
#> 83     12  11   17     43
#> 84     12  11   17     43
#> 85     12  11   17     43
#> 86     12  11   17     43
#> 87     12  11   17     43
#> 88     12  11   17     43
#> 89     12  11   17     43
#> 90     12  11   17     43
#> 91     12  11   17     43
#> 92     12  11   17     43
#> 93     12  11   17     43
#> 94     12  11   17     43
#> 95     12  11   17     43
#> 96     12  11   17     43
#> 97     12  11   17     43
#> 98     12  11   17     43
#> 99     12  11   17     43
#> 100    12  11   17     43
#> 101    12  11   17     43
#> 102    12  11   17     43
#> 103    12  11   17     43
#> 104    12  11   17     43
#> 105    12  11   17     43
#> 106    12  11   17     43
#> 107    12  11   17     43
#> 108    12  11   17     43
#> 109    12  11   17     43
#> 110    12  11   17     43
#> 111    12  11   17     43
#> 112    12  11   17     43
#> 113    12  11   17     43
#> 114    12  11   17     43
#> 115    12  11   17     43
#> 116    12  11   17     43
#> 117    12  11   17     43
#> 118    12  11   17     43
#> 119    12  11   17     43
#> 120    12  11   17     43
#> 121    12  11   17     43
#> 122    12  11   17     43
#> 123    12  11   17     43
#> 124    12  11   17     43
#> 125    12  11   17     43
#> 126    12  11   17     43
#> 127    12  11   17     43
#> 128    12  11   17     43
#> 129    12  11   17     43
#> 130    12  11   17     43
#> 131    12  11   17     43
#> 132    12  11   17     43
#> 133    12  11   17     43
#> 134    12  11   17     43
#> 135    12  11   17     43
#> 136    12  11   17     43
#> 137    12  11   17     43
#> 138    12  11   17     43
#> 139    12  11   17     43
#> 140    12  11   17     43
#> 141    12  11   17     43
#> 142    12  11   17     43
#> 143    12  11   17     43
#> 144    12  11   17     43
#> 145    12  11   17     43
#> 146    12  11   17     43
#> 147    12  11   17     43
#> 148    12  11   17     43
#> 149    12  11   17     43
#> 150    12  11   17     43
#> 151    12  11   17     43
#> 152    12  11   17     43
#> 153    12  11   17     43
#> 154    12  11   17     43
#> 155    12  11   17     43
#> 156    12  11   17     43
#> 157    12  11   17     43
#> 158    12  11   17     43
#> 159    12  11   17     43
#> 160    12  11   17     43
#> 161    12  11   17     43
#> 162    12  11   17     43
#> 163    12  11   17     43
#> 164    12  11   17     43
#> 165    12  11   17     43
#> 166    12  11   17     43
#> 167    12  11   17     43
#> 168    12  11   17     43
#> 169    12  11   17     43
#> 170    12  11   17     43
#> 171    12  11   17     43
#> 172    12  11   17     43
#> 173    12  11   17     43
#> 174    12  11   17     43
#> 175    12  11   17     43
#> 176    12  11   17     43
#> 177    12  11   17     43
#> 178    12  11   17     43
#> 179    12  11   17     43
#> 180    12  11   17     43
#> 181    12  11   17     43
#> 182    12  11   17     43
#> 183    12  11   17     43
#> 184    12  11   17     43
#> 185    12  11   17     43
#> 186    12  11   17     43
#> 187    12  11   17     43
#> 188    12  11   17     43
#> 189    12  11   17     43
#> 190    12  11   17     43
#> 191    12  11   17     43
#> 192    12  11   17     43
#> 193    12  11   17     43
#> 194    12  11   17     43
#> 195    12  11   17     43
#> 196    12  11   17     43
#> 197    12  11   17     43
#> 198    12  11   17     43
#> 199    12  11   17     43
#> 200    12  11   17     43
#> 201    12  11   17     43
#> 202    12  11   17     43
#> 203    12  11   17     43
#> 204    12  11   17     43
#> 205    12  11   17     43
#> 206    12  11   17     43
#> 207    12  11   17     43
#> 208    12  11   17     43
#> 209    12  11   17     43
#> 210    12  11   17     43
#> 211    12  11   17     43
#> 212    12  11   17     43
#> 213    12  11   17     43
#> 214    12  11   17     43
#> 215    12  11   17     43
#> 216    12  11   17     43
#> 217    12  11   17     43
#> 218    12  11   17     43
#> 219    12  11   17     43
#> 220    12  11   17     43
#> 221    12  11   17     43
#> 222    12  11   17     43
#> 223    12  11   17     43
#> 224    12  11   17     43
#> 225    12  11   17     43
#> 226    12  11   17     43
#> 227    12  11   17     43
#> 228    12  11   17     43
#> 229    12  11   17     43
#> 230    12  11   17     43
#> 231    12  11   17     43
#> 232    12  11   17     43
#> 233    12  11   17     43
#> 234    12  11   17     43
#> 235    12  11   17     43
#> 236    12  11   17     43
#> 237    12  11   17     43
#> 238    12  11   17     43
#> 239    12  11   17     43
#> 240    12  11   17     43
#> 241    12  11   17     43
#> 242    12  11   17     43
#> 243    12  11   17     43
#> 244    12  11   17     43
#> 245    12  11   17     43
#> 246    12  11   17     43
#> 247    12  11   17     43
#> 248    12  11   17     43
#> 249    12  11   17     43
#> 250    12  11   17     43
#> 251    12  11   17     43
#> 252    12  11   17     43
#> 253    12  11   17     43
#> 254    12  11   17     43
#> 255    12  11   17     43
#> 256    12  11   17     43
#> 257    12  11   17     43
#> 258    12  11   17     43
#> 259    12  11   17     43
#> 260    12  11   17     43
#> 261    12  11   17     43
#> 262    12  11   17     43
#> 263    12  11   17     43
#> 264    12  11   17     43
#> 265    12  11   17     43