Grab the free_bike_status feed.

Grab a dataframe giving the geographic location and other metadata of bikeshare bikes not parked at bikeshare stations. Metadata for this dataset can be found at: https://github.com/NABSA/gbfs/blob/master/gbfs.md

get_free_bike_status(
  city,
  directory = NULL,
  file = "free_bike_status.rds",
  output = NULL
)

Arguments

city
A character string that can be matched to a city or a url to an active gbfs .json feed. See [get_gbfs_cities()] for a current list of available cities.

directory
Optional. Path to a folder (or folder to be created) where the feed will be saved.

file
Optional. The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
Optional. The type of output method. If left as default, this argument is inferred from the directory argument. If output = "save", the object will be saved as an .rds object at # the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on the argument to output and directory. Either a saved .rds object generated from the current station_information feed, a dataframe object, or both. If a saved feed of the same type already exists at the filepath, the feed will be appended to rather than overwritten.

See also

Examples

# we can grab the free bike status feed for portland, 
# oregon's bikeshare program in  several ways! the most 
# straightforward way is just to supply the `city` argument
# as a string:
get_free_bike_status(city = "portland")

#> Message: Returning the output data as an object, rather than saving it, since the `directory` argument was not specified. Setting `output = "return"` will silence this message.

#>        bike_id                                name       lon      lat
#> 1    bike_6887                       0450 BIKETOWN -122.6836 45.51919
#> 2    bike_6101                       0190 BIKETOWN -122.6889 45.52931
#> 3    bike_6058                       0259 BIKETOWN -122.6736 45.51789
#> 4    bike_6109                       0894 BIKETOWN -122.6818 45.51987
#> 5    bike_6106                       0669 BIKETOWN -122.6660 45.52719
#> 6    bike_7266                       0775 BIKETOWN -122.6860 45.52787
#> 7    bike_6610                       0818 BIKETOWN -122.6757 45.52972
#> 8    bike_5987                       0487 BIKETOWN -122.6959 45.53336
#> 9    bike_7144                       0224 BIKETOWN -122.6818 45.51993
#> 10   bike_6313                       0930 BIKETOWN -122.6553 45.50863
#> 11   bike_6464                       0559 BIKETOWN -122.6427 45.49896
#> 12   bike_7253                       0177 BIKETOWN -122.7103 45.53857
#> 13   bike_6102                      0083 BIKETOWN  -122.6709 45.52506
#> 14   bike_7101                      0496 BIKETOWN  -122.6765 45.53113
#> 15   bike_7238                       0558 BIKETOWN -122.6488 45.50124
#> 16   bike_7327   1005 DESIGN BIKE, NE CYCLING LIFE -122.6781 45.52314
#> 17   bike_7325                    0969 ASCEND BIKE -122.6822 45.52025
#> 18   bike_6889                       0063 BIKETOWN -122.6715 45.52486
#> 19   bike_7300            0996 BLACK HISTORY MONTH -122.6781 45.52741
#> 20   bike_6527                       0801 BIKETOWN -122.6900 45.53897
#> 21   bike_6104                       0624 BIKETOWN -122.6768 45.52737
#> 22   bike_7149                       0207 BIKETOWN -122.6804 45.52652
#> 23   bike_7497                       0576 BIKETOWN -122.6833 45.53036
#> 24   bike_6550                       0918 BIKETOWN -122.6979 45.52980
#> 25   bike_7280                      0026 BIKETOWN  -122.6408 45.49833
#> 26   bike_6089                       0837 BIKETOWN -122.6826 45.52121
#> 27   bike_7348                 1018 RIP CITY RIDER -122.6648 45.55692
#> 28   bike_7309            0995 BETRUE 2018 TRAINER -122.6496 45.52581
#> 29   bike_6636                       0825 BIKETOWN -122.6594 45.53196
#> 30   bike_6186                       0737 BIKETOWN -122.6716 45.52503
#> 31   bike_6168                       0888 BIKETOWN -122.6374 45.50842
#> 32   bike_7383 0964 NATIVE AMERICAN HERITAGE MONTH -122.6987 45.52763
#> 33   bike_6098                       0454 BIKETOWN -122.6403 45.52815
#> 34   bike_7125                       0172 BIKETOWN -122.6601 45.55598
#> 35   bike_6870                       0779 BIKETOWN -122.6606 45.53593
#> 36   bike_6240                     0216 BIKETOWN   -122.6784 45.52081
#> 37   bike_6777                       0266 BIKETOWN -122.6651 45.52296
#> 38   bike_6034                       0899 BIKETOWN -122.6592 45.56184
#> 39   bike_7205                       0458 BIKETOWN -122.6617 45.51780
#> 40   bike_6573                       0843 BIKETOWN -122.6553 45.50859
#> 41   bike_6238                      0774 BIKETOWN  -122.6537 45.52150
#> 42   bike_6536                       0704 BIKETOWN -122.6656 45.53554
#> 43   bike_7294                    0997 ASCEND BIKE -122.6839 45.52797
#> 44   bike_6193                       0153 BIKETOWN -122.6488 45.50122
#> 45  bike_24814                       0398 BIKETOWN -122.6330 45.56158
#> 46   bike_6279                       0049 BIKETOWN -122.6831 45.52178
#> 47   bike_7211                       0004 BIKETOWN -122.6880 45.51325
#> 48  bike_19527                       0305 BIKETOWN -122.6810 45.51725
#> 49   bike_6503                       0326 BIKETOWN -122.6637 45.52161
#> 50   bike_7334                  0999 VETERANS BIKE -122.7065 45.53177
#> 51  bike_28387                       0689 BIKETOWN -122.6639 45.51302
#> 52   bike_7126                       0220 BIKETOWN -122.6852 45.53441
#> 53   bike_7275                       0264 BIKETOWN -122.6779 45.52303
#> 54   bike_6500                       0136 BIKETOWN -122.6606 45.51993
#> 55   bike_6246                      0511 BIKETOWN  -122.6209 45.53547
#> 56   bike_6878                      0141 BIKETOWN  -122.6368 45.50972
#> 57   bike_7163                       0173 BIKETOWN -122.6768 45.53093
#> 58  bike_19238                       0495 BIKETOWN -122.6709 45.52460
#> 59   bike_6494                      0734 BIKETOWN  -122.6767 45.52320
#> 60   bike_6377                       0110 BIKETOWN -122.6447 45.49957
#> 61   bike_6482                       0308 BIKETOWN -122.6802 45.52501
#> 62   bike_6195                     0662 BIKETOWN   -122.6690 45.53013
#> 63   bike_6786                     0529 BIKETOWN   -122.6488 45.52280
#> 64   bike_7297             0973 BLAZERS 50th BIKE  -122.6750 45.52999
#> 65   bike_6591                       0432 BIKETOWN -122.6089 45.52266
#> 66   bike_7255                       0013 BIKETOWN -122.6555 45.54357
#> 67   bike_6318                       0919 BIKETOWN -122.6393 45.51814
#> 68   bike_7138                       0245 BIKETOWN -122.6536 45.49886
#> 69   bike_6249                       0300 BIKETOWN -122.6446 45.49977
#> 70   bike_6838                      0358 BIKETOWN  -122.6218 45.53520
#> 71   bike_6600                      0923 BIKETOWN  -122.6099 45.50526
#> 72   bike_6155                       0032 BIKETOWN -122.6484 45.49667
#> 73   bike_6064                       0732 BIKETOWN -122.6799 45.51591
#> 74   bike_7313                   0985 ASCEND BIKE  -122.6568 45.51711
#> 75   bike_6065                       0584 BIKETOWN -122.6510 45.52220
#> 76   bike_7121                       0117 BIKETOWN -122.6350 45.50346
#> 77   bike_5997                       0150 BIKETOWN -122.6543 45.50087
#> 78   bike_6649                       0778 BIKETOWN -122.6823 45.52708
#> 79   bike_7298    0980 DESIGN BIKE, N PAUL BUNYAN  -122.6854 45.52636
#> 80   bike_7181                      0128 BIKETOWN  -122.6177 45.53500
#> 81   bike_7328                    1000 AIR SAFARI  -122.6409 45.50660
#> 82   bike_7394          1029 LATINX HERITAGE MONTH -122.6587 45.53401
#> 83   bike_6275                       0826 BIKETOWN -122.6618 45.55210
#> 84   bike_6026                       0442 BIKETOWN -122.6835 45.51909
#> 85  bike_24817                       0630 BIKETOWN -122.6615 45.54739
#> 86  bike_28106                       0319 BIKETOWN -122.6360 45.52583
#> 87   bike_7342          1016 LATINX HERITAGE MONTH -122.6545 45.52215
#> 88   bike_6223                       0353 BIKETOWN -122.6845 45.51691
#> 89  bike_24823                      0557 BIKETOWN  -122.6339 45.52204
#> 90  bike_19237                       0133 BIKETOWN -122.6768 45.53083
#> 91   bike_6892                       0206 BIKETOWN -122.6802 45.52506
#> 92  bike_19242                      0009 BIKETOWN  -122.6853 45.52802
#> 93   bike_7318            0993 BETRUE 2018 TRAINER -122.6602 45.51732
#> 94   bike_6861                       0541 BIKETOWN -122.6488 45.52279
#> 95  bike_29251                       0135 BIKETOWN -122.6316 45.52437
#> 96   bike_6354                       0374 BIKETOWN -122.6789 45.51560
#> 97   bike_6837                       0538 BIKETOWN -122.6840 45.51531
#> 98   bike_7343   1011 DESIGN BIKE, NE CYCLING LIFE -122.6689 45.53012
#> 99   bike_7319     0975 DESIGN BIKE, N PAUL BUNYAN -122.6860 45.51483
#> 100  bike_7236                      0006 BIKETOWN  -122.6054 45.53044
#> 101  bike_6272                      0685 BIKETOWN  -122.6402 45.51285
#> 102  bike_6061                       0165 BIKETOWN -122.6517 45.51996
#> 103  bike_6595                       0822 BIKETOWN -122.6781 45.52585
#> 104  bike_6170                       0336 BIKETOWN -122.6545 45.51579
#> 105  bike_6533                       0610 BIKETOWN -122.6756 45.52083
#> 106  bike_6165                       0613 BIKETOWN -122.6902 45.53142
#> 107  bike_7341              1021 BLAZERS 50th BIKE -122.6781 45.52738
#> 108  bike_7245                       0551 BIKETOWN -122.6547 45.51568
#> 109  bike_7333          1022 LATINX HERITAGE MONTH -122.6809 45.54586
#> 110 bike_29385                       0568 BIKETOWN -122.6910 45.53358
#> 111  bike_6420                       0283 BIKETOWN -122.6640 45.51937
#> 112  bike_6432                       0660 BIKETOWN -122.6607 45.51948
#> 113  bike_6113                       0227 BIKETOWN -122.6322 45.50475
#> 114  bike_6083                      0468 BIKETOWN  -122.6966 45.52759
#> 115  bike_6379                       0225 BIKETOWN -122.6548 45.50468
#> 116  bike_6793                       0703 BIKETOWN -122.6908 45.53209
#> 117  bike_7197                       0014 BIKETOWN -122.6668 45.53800
#> 118  bike_6381                       0842 BIKETOWN -122.6535 45.50375
#> 119  bike_7178                       0672 BIKETOWN -122.6620 45.56055
#> 120  bike_7351                     1003 AIR SAFARI -122.6610 45.52303
#> 121  bike_7498                      0181 BIKETOWN  -122.6538 45.55078
#> 122  bike_7122                       0276 BIKETOWN -122.6619 45.50634
#> 123  bike_7187                       0155 BIKETOWN -122.6690 45.53010
#>     is_reserved is_disabled        last_updated year month day hour minute
#> 1             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 2             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 3             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 4             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 5             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 6             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 7             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 8             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 9             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 10            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 11            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 12            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 13            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 14            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 15            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 16            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 17            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 18            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 19            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 20            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 21            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 22            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 23            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 24            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 25            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 26            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 27            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 28            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 29            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 30            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 31            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 32            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 33            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 34            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 35            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 36            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 37            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 38            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 39            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 40            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 41            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 42            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 43            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 44            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 45            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 46            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 47            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 48            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 49            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 50            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 51            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 52            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 53            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 54            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 55            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 56            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 57            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 58            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 59            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 60            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 61            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 62            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 63            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 64            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 65            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 66            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 67            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 68            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 69            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 70            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 71            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 72            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 73            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 74            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 75            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 76            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 77            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 78            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 79            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 80            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 81            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 82            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 83            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 84            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 85            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 86            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 87            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 88            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 89            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 90            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 91            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 92            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 93            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 94            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 95            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 96            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 97            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 98            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 99            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 100           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 101           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 102           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 103           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 104           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 105           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 106           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 107           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 108           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 109           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 110           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 111           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 112           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 113           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 114           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 115           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 116           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 117           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 118           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 119           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 120           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 121           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 122           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 123           0           0 2020-03-26 12:13:11 2020     3  26   12     13

# the `city` argument can also be supplied as an
# actual URL to an active .json feed
get_free_bike_status(city =
"http://biketownpdx.socialbicycles.com/opendata/free_bike_status.json",
                     directory = tempdir())

# if you'd like to save the output to file, supply a 
# `directory` argument. usually, though, we would supply a 
# character string (like "pdx", maybe,) for the `directory` 
# argument instead of `tempdir`.
get_free_bike_status(city = "portland",
                     directory = tempdir())

# the output argument can control whether the file is
# saved and/or returned explicitly
get_free_bike_status(city = "portland",
                     directory = tempdir(),
                     output = "both")

#>        bike_id                                name       lon      lat
#> 1    bike_6887                       0450 BIKETOWN -122.6836 45.51919
#> 2    bike_6101                       0190 BIKETOWN -122.6889 45.52931
#> 3    bike_6058                       0259 BIKETOWN -122.6736 45.51789
#> 4    bike_6109                       0894 BIKETOWN -122.6818 45.51987
#> 5    bike_6106                       0669 BIKETOWN -122.6660 45.52719
#> 6    bike_7266                       0775 BIKETOWN -122.6860 45.52787
#> 7    bike_6610                       0818 BIKETOWN -122.6757 45.52972
#> 8    bike_5987                       0487 BIKETOWN -122.6959 45.53336
#> 9    bike_7144                       0224 BIKETOWN -122.6818 45.51993
#> 10   bike_6313                       0930 BIKETOWN -122.6553 45.50863
#> 11   bike_6464                       0559 BIKETOWN -122.6427 45.49896
#> 12   bike_7253                       0177 BIKETOWN -122.7103 45.53857
#> 13   bike_6102                      0083 BIKETOWN  -122.6709 45.52506
#> 14   bike_7101                      0496 BIKETOWN  -122.6765 45.53113
#> 15   bike_7238                       0558 BIKETOWN -122.6488 45.50124
#> 16   bike_7327   1005 DESIGN BIKE, NE CYCLING LIFE -122.6781 45.52314
#> 17   bike_7325                    0969 ASCEND BIKE -122.6822 45.52025
#> 18   bike_6889                       0063 BIKETOWN -122.6715 45.52486
#> 19   bike_7300            0996 BLACK HISTORY MONTH -122.6781 45.52741
#> 20   bike_6527                       0801 BIKETOWN -122.6900 45.53897
#> 21   bike_6104                       0624 BIKETOWN -122.6768 45.52737
#> 22   bike_7149                       0207 BIKETOWN -122.6804 45.52652
#> 23   bike_7497                       0576 BIKETOWN -122.6833 45.53036
#> 24   bike_6550                       0918 BIKETOWN -122.6979 45.52980
#> 25   bike_7280                      0026 BIKETOWN  -122.6408 45.49833
#> 26   bike_6089                       0837 BIKETOWN -122.6826 45.52121
#> 27   bike_7348                 1018 RIP CITY RIDER -122.6648 45.55692
#> 28   bike_7309            0995 BETRUE 2018 TRAINER -122.6496 45.52581
#> 29   bike_6636                       0825 BIKETOWN -122.6594 45.53196
#> 30   bike_6186                       0737 BIKETOWN -122.6716 45.52503
#> 31   bike_6168                       0888 BIKETOWN -122.6374 45.50842
#> 32   bike_7383 0964 NATIVE AMERICAN HERITAGE MONTH -122.6987 45.52763
#> 33   bike_6098                       0454 BIKETOWN -122.6403 45.52815
#> 34   bike_7125                       0172 BIKETOWN -122.6601 45.55598
#> 35   bike_6870                       0779 BIKETOWN -122.6606 45.53593
#> 36   bike_6240                     0216 BIKETOWN   -122.6784 45.52081
#> 37   bike_6777                       0266 BIKETOWN -122.6651 45.52296
#> 38   bike_6034                       0899 BIKETOWN -122.6592 45.56184
#> 39   bike_7205                       0458 BIKETOWN -122.6617 45.51780
#> 40   bike_6573                       0843 BIKETOWN -122.6553 45.50859
#> 41   bike_6238                      0774 BIKETOWN  -122.6537 45.52150
#> 42   bike_6536                       0704 BIKETOWN -122.6656 45.53554
#> 43   bike_7294                    0997 ASCEND BIKE -122.6839 45.52797
#> 44   bike_6193                       0153 BIKETOWN -122.6488 45.50122
#> 45  bike_24814                       0398 BIKETOWN -122.6330 45.56158
#> 46   bike_6279                       0049 BIKETOWN -122.6831 45.52178
#> 47   bike_7211                       0004 BIKETOWN -122.6880 45.51325
#> 48  bike_19527                       0305 BIKETOWN -122.6810 45.51725
#> 49   bike_6503                       0326 BIKETOWN -122.6637 45.52161
#> 50   bike_7334                  0999 VETERANS BIKE -122.7065 45.53177
#> 51  bike_28387                       0689 BIKETOWN -122.6639 45.51302
#> 52   bike_7126                       0220 BIKETOWN -122.6852 45.53441
#> 53   bike_7275                       0264 BIKETOWN -122.6779 45.52303
#> 54   bike_6500                       0136 BIKETOWN -122.6606 45.51993
#> 55   bike_6246                      0511 BIKETOWN  -122.6209 45.53547
#> 56   bike_6878                      0141 BIKETOWN  -122.6368 45.50972
#> 57   bike_7163                       0173 BIKETOWN -122.6768 45.53093
#> 58  bike_19238                       0495 BIKETOWN -122.6709 45.52460
#> 59   bike_6494                      0734 BIKETOWN  -122.6767 45.52320
#> 60   bike_6377                       0110 BIKETOWN -122.6447 45.49957
#> 61   bike_6482                       0308 BIKETOWN -122.6802 45.52501
#> 62   bike_6195                     0662 BIKETOWN   -122.6690 45.53013
#> 63   bike_6786                     0529 BIKETOWN   -122.6488 45.52280
#> 64   bike_7297             0973 BLAZERS 50th BIKE  -122.6750 45.52999
#> 65   bike_6591                       0432 BIKETOWN -122.6089 45.52266
#> 66   bike_7255                       0013 BIKETOWN -122.6555 45.54357
#> 67   bike_6318                       0919 BIKETOWN -122.6393 45.51814
#> 68   bike_7138                       0245 BIKETOWN -122.6536 45.49886
#> 69   bike_6249                       0300 BIKETOWN -122.6446 45.49977
#> 70   bike_6838                      0358 BIKETOWN  -122.6218 45.53520
#> 71   bike_6600                      0923 BIKETOWN  -122.6099 45.50526
#> 72   bike_6155                       0032 BIKETOWN -122.6484 45.49667
#> 73   bike_6064                       0732 BIKETOWN -122.6799 45.51591
#> 74   bike_7313                   0985 ASCEND BIKE  -122.6568 45.51711
#> 75   bike_6065                       0584 BIKETOWN -122.6510 45.52220
#> 76   bike_7121                       0117 BIKETOWN -122.6350 45.50346
#> 77   bike_5997                       0150 BIKETOWN -122.6543 45.50087
#> 78   bike_6649                       0778 BIKETOWN -122.6823 45.52708
#> 79   bike_7298    0980 DESIGN BIKE, N PAUL BUNYAN  -122.6854 45.52636
#> 80   bike_7181                      0128 BIKETOWN  -122.6177 45.53500
#> 81   bike_7328                    1000 AIR SAFARI  -122.6409 45.50660
#> 82   bike_7394          1029 LATINX HERITAGE MONTH -122.6587 45.53401
#> 83   bike_6275                       0826 BIKETOWN -122.6618 45.55210
#> 84   bike_6026                       0442 BIKETOWN -122.6835 45.51909
#> 85  bike_24817                       0630 BIKETOWN -122.6615 45.54739
#> 86  bike_28106                       0319 BIKETOWN -122.6360 45.52583
#> 87   bike_7342          1016 LATINX HERITAGE MONTH -122.6545 45.52215
#> 88   bike_6223                       0353 BIKETOWN -122.6845 45.51691
#> 89  bike_24823                      0557 BIKETOWN  -122.6339 45.52204
#> 90  bike_19237                       0133 BIKETOWN -122.6768 45.53083
#> 91   bike_6892                       0206 BIKETOWN -122.6802 45.52506
#> 92  bike_19242                      0009 BIKETOWN  -122.6853 45.52802
#> 93   bike_7318            0993 BETRUE 2018 TRAINER -122.6602 45.51732
#> 94   bike_6861                       0541 BIKETOWN -122.6488 45.52279
#> 95  bike_29251                       0135 BIKETOWN -122.6316 45.52437
#> 96   bike_6354                       0374 BIKETOWN -122.6789 45.51560
#> 97   bike_6837                       0538 BIKETOWN -122.6840 45.51531
#> 98   bike_7343   1011 DESIGN BIKE, NE CYCLING LIFE -122.6689 45.53012
#> 99   bike_7319     0975 DESIGN BIKE, N PAUL BUNYAN -122.6860 45.51483
#> 100  bike_7236                      0006 BIKETOWN  -122.6054 45.53044
#> 101  bike_6272                      0685 BIKETOWN  -122.6402 45.51285
#> 102  bike_6061                       0165 BIKETOWN -122.6517 45.51996
#> 103  bike_6595                       0822 BIKETOWN -122.6781 45.52585
#> 104  bike_6170                       0336 BIKETOWN -122.6545 45.51579
#> 105  bike_6533                       0610 BIKETOWN -122.6756 45.52083
#> 106  bike_6165                       0613 BIKETOWN -122.6902 45.53142
#> 107  bike_7341              1021 BLAZERS 50th BIKE -122.6781 45.52738
#> 108  bike_7245                       0551 BIKETOWN -122.6547 45.51568
#> 109  bike_7333          1022 LATINX HERITAGE MONTH -122.6809 45.54586
#> 110 bike_29385                       0568 BIKETOWN -122.6910 45.53358
#> 111  bike_6420                       0283 BIKETOWN -122.6640 45.51937
#> 112  bike_6432                       0660 BIKETOWN -122.6607 45.51948
#> 113  bike_6113                       0227 BIKETOWN -122.6322 45.50475
#> 114  bike_6083                      0468 BIKETOWN  -122.6966 45.52759
#> 115  bike_6379                       0225 BIKETOWN -122.6548 45.50468
#> 116  bike_6793                       0703 BIKETOWN -122.6908 45.53209
#> 117  bike_7197                       0014 BIKETOWN -122.6668 45.53800
#> 118  bike_6381                       0842 BIKETOWN -122.6535 45.50375
#> 119  bike_7178                       0672 BIKETOWN -122.6620 45.56055
#> 120  bike_7351                     1003 AIR SAFARI -122.6610 45.52303
#> 121  bike_7498                      0181 BIKETOWN  -122.6538 45.55078
#> 122  bike_7122                       0276 BIKETOWN -122.6619 45.50634
#> 123  bike_7187                       0155 BIKETOWN -122.6690 45.53010
#>     is_reserved is_disabled        last_updated year month day hour minute
#> 1             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 2             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 3             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 4             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 5             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 6             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 7             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 8             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 9             0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 10            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 11            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 12            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 13            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 14            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 15            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 16            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 17            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 18            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 19            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 20            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 21            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 22            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 23            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 24            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 25            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 26            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 27            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 28            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 29            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 30            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 31            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 32            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 33            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 34            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 35            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 36            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 37            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 38            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 39            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 40            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 41            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 42            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 43            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 44            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 45            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 46            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 47            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 48            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 49            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 50            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 51            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 52            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 53            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 54            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 55            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 56            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 57            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 58            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 59            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 60            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 61            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 62            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 63            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 64            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 65            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 66            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 67            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 68            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 69            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 70            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 71            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 72            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 73            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 74            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 75            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 76            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 77            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 78            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 79            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 80            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 81            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 82            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 83            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 84            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 85            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 86            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 87            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 88            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 89            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 90            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 91            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 92            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 93            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 94            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 95            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 96            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 97            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 98            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 99            0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 100           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 101           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 102           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 103           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 104           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 105           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 106           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 107           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 108           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 109           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 110           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 111           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 112           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 113           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 114           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 115           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 116           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 117           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 118           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 119           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 120           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 121           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 122           0           0 2020-03-26 12:13:11 2020     3  26   12     13
#> 123           0           0 2020-03-26 12:13:11 2020     3  26   12     13