Grab the station_information feed.

get_station_information grabs and tidies the station_information feed for a given city. This dataset contains locations, capacity, and other information about bikeshare stations. Go to `https://github.com/NABSA/gbfs/blob/master/gbfs.md` to see metadata for this dataset.

get_station_information(
  city,
  directory = NULL,
  file = "station_information.rds",
  output = "save"
)

Arguments

city
A character string that can be matched to a city or a url to an active gbfs .json feed. See get_gbfs_cities for a current list of available cities.

directory
The name of an existing folder or folder to be created, where the feed will will be saved. This argument is not required if output = "return".

file
The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
The type of output method. If output = "save", the object will be saved as an .rds object at the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on argument to output. Either a saved .rds object generated from the current station_information feed, a dataframe object, or both.

Examples

# we can grab the station information feed for portland, 
# oregon in one of several ways! first, supply the `city` 
# argument as a URL, and save to file by leaving output 
# set to it's default. usually, we would supply a character 
# string (like "pdx", maybe,) for the `directory` argument 
# instead of `tempdir`.
get_station_information(city = "http://biketownpdx.socialbicycles.com/opendata/system_alerts.json",
                        directory = tempdir())

# or, instead, just supply the name of 
# the city as a string. 
get_station_information(city = "portland",
                        directory = tempdir())

# instead of saving the output as a file, we can 
# just return the output as a dataframe
get_station_information(city = "portland",
                        output = "return")

#>     station_id
#> 1     hub_1512
#> 2     hub_1513
#> 3     hub_1514
#> 4     hub_1515
#> 5     hub_1516
#> 6     hub_1517
#> 7     hub_1518
#> 8     hub_1519
#> 9     hub_1521
#> 10    hub_1522
#> 11    hub_1523
#> 12    hub_1524
#> 13    hub_1527
#> 14    hub_1528
#> 15    hub_1529
#> 16    hub_1535
#> 17    hub_1536
#> 18    hub_1537
#> 19    hub_1538
#> 20    hub_1539
#> 21    hub_1540
#> 22    hub_1541
#> 23    hub_1542
#> 24    hub_1543
#> 25    hub_1544
#> 26    hub_1545
#> 27    hub_1546
#> 28    hub_1547
#> 29    hub_1548
#> 30    hub_1549
#> 31    hub_1550
#> 32    hub_1551
#> 33    hub_1552
#> 34    hub_1554
#> 35    hub_1555
#> 36    hub_1556
#> 37    hub_1558
#> 38    hub_1559
#> 39    hub_1560
#> 40    hub_1561
#> 41    hub_1562
#> 42    hub_1563
#> 43    hub_1564
#> 44    hub_1566
#> 45    hub_1567
#> 46    hub_1568
#> 47    hub_1569
#> 48    hub_1570
#> 49    hub_1571
#> 50    hub_1572
#> 51    hub_1573
#> 52    hub_1574
#> 53    hub_1575
#> 54    hub_1577
#> 55    hub_1578
#> 56    hub_1579
#> 57    hub_1580
#> 58    hub_1581
#> 59    hub_1582
#> 60    hub_1583
#> 61    hub_1584
#> 62    hub_1585
#> 63    hub_1586
#> 64    hub_1588
#> 65    hub_1590
#> 66    hub_1591
#> 67    hub_1592
#> 68    hub_1593
#> 69    hub_1594
#> 70    hub_1595
#> 71    hub_1597
#> 72    hub_1598
#> 73    hub_1599
#> 74    hub_1600
#> 75    hub_1601
#> 76    hub_1602
#> 77    hub_1603
#> 78    hub_1604
#> 79    hub_1605
#> 80    hub_1606
#> 81    hub_1607
#> 82    hub_1608
#> 83    hub_1609
#> 84    hub_1611
#> 85    hub_1612
#> 86    hub_1613
#> 87    hub_1614
#> 88    hub_1615
#> 89    hub_1616
#> 90    hub_1617
#> 91    hub_1713
#> 92    hub_2275
#> 93    hub_2276
#> 94    hub_2598
#> 95    hub_2677
#> 96    hub_2678
#> 97    hub_2679
#> 98    hub_2680
#> 99    hub_2681
#> 100   hub_2682
#> 101   hub_2683
#> 102   hub_2684
#> 103   hub_2685
#> 104   hub_2686
#> 105   hub_2687
#> 106   hub_2688
#> 107   hub_2689
#> 108   hub_2695
#> 109   hub_2696
#> 110   hub_2698
#> 111   hub_2707
#> 112   hub_2708
#> 113   hub_2789
#> 114   hub_2798
#> 115   hub_2875
#> 116   hub_3240
#> 117   hub_3929
#> 118   hub_3957
#> 119   hub_4034
#> 120   hub_4050
#> 121   hub_4312
#> 122   hub_4313
#> 123   hub_4315
#> 124   hub_4316
#> 125   hub_4317
#> 126   hub_4318
#> 127   hub_4319
#> 128   hub_4320
#> 129   hub_4321
#> 130   hub_4322
#> 131   hub_4323
#> 132   hub_4324
#> 133   hub_4325
#> 134   hub_4326
#> 135   hub_4327
#> 136   hub_4328
#> 137   hub_4329
#> 138   hub_4368
#> 139   hub_4381
#> 140   hub_4482
#> 141   hub_5277
#> 142   hub_5311
#> 143   hub_5316
#> 144   hub_5357
#> 145   hub_5414
#> 146   hub_5470
#> 147   hub_5473
#> 148   hub_5474
#> 149   hub_5476
#> 150   hub_5491
#> 151   hub_5493
#> 152   hub_5498
#> 153   hub_5567
#> 154   hub_5597
#>                                                                  name
#> 1                                             SW 10th at Harvey Milk 
#> 2                                   SE Gideon at 12th Ave MAX station
#> 3                                                  SE Clinton at 26th
#> 4                                              SE Elliott at Division
#> 5                                                    SE Clay at Water
#> 6                                                  SE 29th at Belmont
#> 7                                                  SE Water at Taylor
#> 8                                                    SE 14th at Stark
#> 9                                            N Prescott at Interstate
#> 10                                                  NE Glisan at 24th
#> 11                                            NE Wheeler at Multnomah
#> 12                                        N Mississippi at N Skidmore
#> 13                               SE Taylor at Chavez Community Corral
#> 14                                                NW 13th at Marshall
#> 15                                                NW 18th at Northrup
#> 16                                     SW Park at Portland Art Museum
#> 17                 SW 4th at Jefferson (City Hall) - Community Corral
#> 18                        SW Broadway at Smith Memorial Student Union
#> 19                                    Urban Center Plaza at SW 6th St
#> 20                                                  SW 4th at College
#> 21                                                 SW Harrison at 1st
#> 22                                   SW Moody at Aerial Tram Terminal
#> 23                                                SW 10th at Columbia
#> 24                                                    SW 12th at Clay
#> 25                                                     SW 5th at Main
#> 26                                 NW Marshall at Tanner Springs Park
#> 27                                                 NW Thurman at 24th
#> 28                                                 NW Raleigh at 21st
#> 29                                                SE 34th at Division
#> 30                                                SE 30th at Division
#> 31                                                 SE Clinton at 21st
#> 32                                                   SE 29th at Stark
#> 33                                                 SE Belmont at 20th
#> 34                                               SE 34th at Hawthorne
#> 35                                                NE 15th at Broadway
#> 36                                                    NE MLK at Knott
#> 37                                                 NW Everett at 22nd
#> 38                                             N Vancouver at Alberta
#> 39                                          N Killingsworth at Albina
#> 40                                       NW Johnson at Jamison Square
#> 41                                               NW Station at Irving
#> 42                                                 NW 23rd at Overton
#> 43                                                NW 20th at Burnside
#> 44                                                  NW Glisan at 21st
#> 45                                                 NW 21st at Johnson
#> 46                                                 NW 10th at Everett
#> 47                                       SW Salmon at Waterfront Park
#> 48                                             SW River at Montgomery
#> 49                                         SW Harvey Milk at Burnside
#> 50                                               SW 18th at Jefferson
#> 51                                 SW 11th at Main - Community Corral
#> 52                                                     SW 2nd at Pine
#> 53                                                   SW 3rd at Ankeny
#> 54                                                 SW 5th at Morrison
#> 55                                                      SW 5th at Oak
#> 56                                  SW Market at Keller Fountain Park
#> 57                                                SW Morrison at 18th
#> 58                              SW Morrison at Pioneer Courthouse Sq.
#> 59                                        SW Yamhill at Director Park
#> 60                                              NW 11th at The Fields
#> 61                                                NW 18th at Flanders
#> 62                                                 NW 22nd at Lovejoy
#> 63                                                  NW 2nd at Everett
#> 64                                                   NW Couch at 11th
#> 65                                                NW Flanders at 14th
#> 66                                           Tilikum West at SW Moody
#> 67                                                 NW Kearney at 23rd
#> 68                                               NW Naito at Ironside
#> 69                                              N Failing at Williams
#> 70                                             N Mississippi at Beech
#> 71                           N Russell at Interstate Community Corral
#> 72                                              N Williams at Fremont
#> 73                                           NE 11th at Holladay Park
#> 74                                           SW Naito at Ankeny Plaza
#> 75                                                NE 22nd at Burnside
#> 76                                                NE Broadway at 12th
#> 77                              NE Holladay at MLK - Community Corral
#> 78                                                NE Multnomah at 9th
#> 79                                              NE Multnomah at Grand
#> 80                                                   NE Sandy at 16th
#> 81                                                  NE Going at Grand
#> 82                                               SE 29th at Hawthorne
#> 83                                           SE 2nd Pl at Tilikum Way
#> 84                                                  SE 9th at Belmont
#> 85                                                    SE Alder at 6th
#> 86                                                 SE Belmont at 33rd
#> 87                                               SE Hawthorne at 20th
#> 88                                               SE Ladd at Hawthorne
#> 89                           SW Park at W Burnside - Community Corral
#> 90                                                N Mason at Williams
#> 91                                                 NW 19th at Johnson
#> 92                                NE 28th at Couch - Community Corral
#> 93                                 SE 28th at Pine - Community Corral
#> 94                                    Metro Campus - Community Corral
#> 95                              NE Alberta at 10th - Community Corral
#> 96                              NE Alberta at 14th - Community Corral
#> 97                              NE Alberta at 17th - Community Corral
#> 98                              NE Alberta at 22nd - Community Corral
#> 99                              NE Alberta at 24th - Community Corral
#> 100                     NE Alberta at NE 29th/30th - Community Corral
#> 101            NE 33rd at Jarrett (Kennedy School) - Community Corral
#> 102                             NE 33rd at Emerson - Community Corral
#> 103                         NE 14th Pl at Prescott - Community Corral
#> 104                             NE Fremont at 15th - Community Corral
#> 105                    N Michigan at Killingsworth - Community Corral
#> 106                     N Concord at Killingsworth - Community Corral
#> 107                         N Killingsworth at Gay - Community Corral
#> 108                       NE Killingsworth at 30th - Community Corral
#> 109                       NE 17th at Killingsworth - Community Corral
#> 110                                              SE 40th at Hawthorne
#> 111                                              SW Naito at Morrison
#> 112                     N Interstate at Willamette - Community Corral
#> 113                                            NW Broadway at Everett
#> 114                                             NW Pettygrove at 15th
#> 115                                                SW Moody at Thomas
#> 116                                                   NE Couch at MLK
#> 117                                                SE Belmont at 42nd
#> 118                                             NE 2nd at NE Schuyler
#> 119 SW Park at SW Wright (Oregon Holocaust Memorial) Community Corral
#> 120                                                  NW Front at 17th
#> 121                             SE Belmont at 45th - Community Corral
#> 122                           SE 37th at Gladstone - Community Corral
#> 123                             NE 50th at Fremont - Community Corral
#> 124                           NE Cully at Skidmore - Community Corral
#> 125                          SE 29th and Gladstone - Community Corral
#> 126                            SE 50th at Division - Community Corral
#> 127                            SE 47th at Hawthorne- Community Corral
#> 128                               NE 28th at Glisan- Community Corral
#> 129                            SE 50th at Hawthorne- Community Corral
#> 130                            SE 49th at Division - Community Corral
#> 131                              NE 30th at Glisan - Community Corral
#> 132                            SE 37th at Division - Community Corral
#> 133        NE 41st at Sandy Blvd (Hollywood Theatre) Community Corral
#> 134                                                NE 42nd at Hancock
#> 135                             NE 48th at Fremont - Community Corral
#> 136    NE 22nd at Thompson - Community Corral (Irvington Tennis Club)
#> 137                       N Burrage at Killingsworth Community Corral
#> 138                             NE 10th at Halsey -- Community Corral
#> 139                           NW Nicolai at Sherlock Community Corral
#> 140                                                SE 9th at Burnside
#> 141                      NE Cully at Killingsworth - Community Corral
#> 142                                                 SE 9th at Madison
#> 143                     SE Powell Blvd at 29th Ave - Community Corral
#> 144                                                   NW 4th at Couch
#> 145                                                SW 3rd at Morrison
#> 146                                                 NE 60th at Cully 
#> 147                                                  NE 50th at Sandy
#> 148                                                 NE Glisan at 53rd
#> 149                                       NE 42nd at NE Alberta Court
#> 150                                            SE Powell Blvd at 28th
#> 151                                                  NE 43rd at Sandy
#> 152                          N Wheeler at Broadway - Community Corral
#> 153                                                SE 50th at Clinton
#> 154                              N Alberta Street at N Interstate Ave
#>      region_id       lon      lat
#> 1   region_241 -122.6811 45.52175
#> 2   region_241 -122.6535 45.50281
#> 3   region_241 -122.6395 45.50357
#> 4   region_241 -122.6534 45.50513
#> 5   region_241 -122.6663 45.51148
#> 6   region_241 -122.6361 45.51623
#> 7   region_241 -122.6658 45.51515
#> 8   region_241 -122.6517 45.51845
#> 9   region_241 -122.6819 45.55522
#> 10  region_241 -122.6416 45.52649
#> 11  region_241 -122.6659 45.53087
#> 12  region_241 -122.6755 45.55459
#> 13  region_241 -122.6229 45.51516
#> 14  region_241 -122.6844 45.53080
#> 15  region_241 -122.6896 45.53143
#> 16  region_241 -122.6833 45.51592
#> 17  region_241 -122.6784 45.51515
#> 18  region_241 -122.6837 45.51184
#> 19  region_241 -122.6823 45.51158
#> 20  region_241 -122.6812 45.50936
#> 21  region_241 -122.6777 45.50971
#> 22  region_241 -122.6716 45.49914
#> 23  region_241 -122.6847 45.51544
#> 24  region_241 -122.6870 45.51507
#> 25  region_241 -122.6791 45.51610
#> 26  region_241 -122.6816 45.53083
#> 27  region_241 -122.7007 45.53550
#> 28  region_241 -122.6949 45.53409
#> 29  region_241 -122.6296 45.50484
#> 30  region_241 -122.6346 45.50469
#> 31  region_241 -122.6447 45.50337
#> 32  region_241 -122.6360 45.51936
#> 33  region_241 -122.6460 45.51636
#> 34  region_241 -122.6295 45.51191
#> 35  region_241 -122.6506 45.53508
#> 36  region_241 -122.6615 45.54180
#> 37  region_241 -122.6965 45.52488
#> 38  region_241 -122.6685 45.55941
#> 39  region_241 -122.6747 45.56283
#> 40  region_241 -122.6820 45.52864
#> 41  region_241 -122.6766 45.52828
#> 42  region_241 -122.6986 45.53212
#> 43  region_241 -122.6933 45.52331
#> 44  region_241 -122.6941 45.52635
#> 45  region_241 -122.6946 45.52827
#> 46  region_241 -122.6811 45.52500
#> 47  region_241 -122.6739 45.51557
#> 48  region_241 -122.6735 45.50910
#> 49  region_241 -122.6840 45.52279
#> 50  region_241 -122.6920 45.51788
#> 51  region_241 -122.6847 45.51730
#> 52  region_241 -122.6726 45.52142
#> 53  region_241 -122.6733 45.52248
#> 54  region_241 -122.6774 45.51889
#> 55  region_241 -122.6764 45.52117
#> 56  region_241 -122.6789 45.51232
#> 57  region_241 -122.6897 45.52196
#> 58  region_241 -122.6795 45.51946
#> 59  region_241 -122.6813 45.51898
#> 60  region_241 -122.6824 45.53233
#> 61  region_241 -122.6893 45.52556
#> 62  region_241 -122.6966 45.52976
#> 63  region_241 -122.6725 45.52537
#> 64  region_241 -122.6818 45.52374
#> 65  region_241 -122.6856 45.52571
#> 66  region_241 -122.6714 45.50277
#> 67  region_241 -122.6987 45.52914
#> 68  region_241 -122.6742 45.52871
#> 69  region_241 -122.6669 45.55076
#> 70  region_241 -122.6757 45.54999
#> 71  region_241 -122.6770 45.54108
#> 72  region_241 -122.6668 45.54759
#> 73  region_241 -122.6547 45.53014
#> 74  region_241 -122.6703 45.52270
#> 75  region_241 -122.6434 45.52297
#> 76  region_241 -122.6536 45.53497
#> 77  region_241 -122.6618 45.52969
#> 78  region_241 -122.6575 45.53141
#> 79  region_241 -122.6597 45.53146
#> 80  region_241 -122.6498 45.52441
#> 81  region_241 -122.6604 45.55645
#> 82  region_241 -122.6363 45.51202
#> 83  region_241 -122.6633 45.50624
#> 84  region_241 -122.6566 45.51633
#> 85  region_241 -122.6601 45.51790
#> 86  region_241 -122.6308 45.51650
#> 87  region_241 -122.6454 45.51222
#> 88  region_241 -122.6533 45.51208
#> 89  region_241 -122.6790 45.52277
#> 90  region_241 -122.6669 45.55335
#> 91  region_241 -122.6905 45.52862
#> 92  region_241 -122.6372 45.52373
#> 93  region_241 -122.6372 45.52059
#> 94  region_241 -122.6604 45.52772
#> 95  region_241 -122.6554 45.55914
#> 96  region_241 -122.6514 45.55917
#> 97  region_241 -122.6475 45.55907
#> 98  region_241 -122.6427 45.55906
#> 99  region_241 -122.6408 45.55909
#> 100 region_241 -122.6355 45.55902
#> 101 region_241 -122.6305 45.56453
#> 102 region_241 -122.6303 45.56139
#> 103 region_241 -122.6507 45.55568
#> 104 region_241 -122.6508 45.54829
#> 105 region_241 -122.6770 45.56252
#> 106 region_241 -122.6848 45.56245
#> 107 region_241 -122.6890 45.56274
#> 108 region_241 -122.6350 45.56272
#> 109 region_241 -122.6481 45.56275
#> 110 region_241 -122.6210 45.51186
#> 111 region_241 -122.6725 45.51783
#> 112 region_241 -122.6820 45.56197
#> 113 region_241 -122.6776 45.52508
#> 114 region_241 -122.6863 45.53290
#> 115 region_241 -122.6719 45.49429
#> 116 region_241 -122.6619 45.52359
#> 117 region_241 -122.6190 45.51646
#> 118 region_241 -122.6636 45.53561
#> 119 region_241 -122.7044 45.52182
#> 120 region_241 -122.6881 45.53715
#> 121 region_241 -122.6160 45.51658
#> 122 region_241 -122.6234 45.49334
#> 123 region_241 -122.6112 45.54828
#> 124 region_241 -122.6021 45.55392
#> 125 region_241 -122.6358 45.49334
#> 126 region_241 -122.6109 45.50528
#> 127 region_241 -122.6142 45.51195
#> 128 region_241 -122.6362 45.52643
#> 129 region_241 -122.6113 45.51189
#> 130 region_241 -122.6121 45.50524
#> 131 region_241 -122.6347 45.52643
#> 132 region_241 -122.6253 45.50473
#> 133 region_241 -122.6206 45.53578
#> 134 region_241 -122.6200 45.53690
#> 135 region_241 -122.6137 45.54831
#> 136 region_241 -122.6433 45.53894
#> 137 region_241 -122.6946 45.56268
#> 138 region_241 -122.6556 45.53401
#> 139 region_241 -122.6967 45.54126
#> 140 region_241 -122.6566 45.52279
#> 141 region_241 -122.5944 45.56251
#> 142 region_241 -122.6567 45.51300
#> 143 region_241 -122.6354 45.49723
#> 144 region_241 -122.6747 45.52379
#> 145 region_241 -122.6755 45.51803
#> 146 region_241 -122.6008 45.55593
#> 147 region_241 -122.6119 45.53967
#> 148 region_241 -122.6093 45.52646
#> 149 region_241 -122.6204 45.55836
#> 150 region_241 -122.6375 45.49787
#> 151 region_241 -122.6188 45.53677
#> 152 region_241 -122.6687 45.53480
#> 153 region_241 -122.6111 45.50351
#> 154 region_241 -122.6824 45.55892
#>                                                                                                                                             address
#> 1                                                                                                               404 Southwest 10th Avenue, Portland
#> 2                                                                                                       1201-1241 Southeast Gideon Street, Portland
#> 3                                                                                                           2615 Southeast Clinton Street, Portland
#> 4                                                                                                      2454-2498 Southeast Elliott Avenue, Portland
#> 5                                                                                                        1601-1799 Southeast Water Avenue, Portland
#> 6                                                                                                           901-915 Southeast 29th Avenue, Portland
#> 7                                                                                                             1028 Southeast Water Avenue, Portland
#> 8                                                                                                           531-699 Southeast 14th Avenue, Portland
#> 9                                                                                                            4350 North Interstate Avenue, Portland
#> 10                                                                                                      2329-2399 Northeast Glisan Street, Portland
#> 11                                                                                                     1051-1099 Northeast Wheeler Avenue, Portland
#> 12                                                                                                          4220 North Mississippi Avenue, Portland
#> 13                                                                                                                                         Portland
#> 14                                                                                                        1100-1124 Northwest 13th Avenue, Portland
#> 15                                                                                                         1801 Northwest Northrup Street, Portland
#> 16                                                                                                             1315 Southwest Park Avenue, Portland
#> 17                                                                                                         1200-1238 Southwest 4th Avenue, Portland
#> 18                                                                                                           1801-1825 Southwest Broadway, Portland
#> 19                                                                                                       1800 Southwest Montgomery Street, Portland
#> 20                                                                            TECOTOSH, Southwest 4th Avenue, University District, Portland, Oregon
#> 21                                                                                                        86-98 Southwest Harrison Street, Portland
#> 22                                                                                                                                         Portland
#> 23                                                                                                    1001-1023 Southwest Columbia Street, Portland
#> 24                                                                                                             1520 Southwest 12th Avenue, Portland
#> 25                                                                                                          512-586 Southwest Main Street, Portland
#> 26                                                                                                         1000 Northwest Marshall Street, Portland
#> 27                                                                                                     2383-2389 Northwest Thurman Street, Portland
#> 28                                                                                                     2100-2112 Northwest Raleigh Street, Portland
#> 29                                                                                                    3377-3399 Southeast Division Street, Portland
#> 30                                                                                                        2501-2523 Southeast 30th Avenue, Portland
#> 31                                                                                                     2039-2099 Southeast Clinton Street, Portland
#> 32                                                                                                              415 Southeast 29th Avenue, Portland
#> 33                                                                                                     1900-1998 Southeast Belmont Street, Portland
#> 34                                                                                                             1510 Southeast 34th Avenue, Portland
#> 35                                                                                                             1711 Northeast 15th Avenue, Portland
#> 36                                                                                                                        2650 Oregon 99E, Portland
#> 37                                                                                                     2201-2209 Northwest Everett Street, Portland
#> 38                                                                                                               241 North Alberta Street, Portland
#> 39                                                                                                             North Killingsworth Street, Portland
#> 40                                                                                                              718 Northwest 11th Avenue, Portland
#> 41                                                                                                           700-746 Northwest 6th Avenue, Portland
#> 42                                                                                                             1310 Northwest 23rd Avenue, Portland
#> 43                                                                                                              2-90 Northwest 20th Place, Portland
#> 44                                                                                                           2066 Northwest Glisan Street, Portland
#> 45                                                                                                              721 Northwest 21st Avenue, Portland
#> 46                                                                                                              230 Northwest 10th Avenue, Portland
#> 47                                                                                                             25 Southwest Salmon Street, Portland
#> 48                                                Portland Sports Bar and Grill, 1800, Southwest River Drive, University District, Portland, Oregon
#> 49                                                                                                       1250-1298 Southwest Stark Street, Portland
#> 50                                                                               1234, Southwest 18th Avenue, University District, Portland, Oregon
#> 51                                                              The St. Francis, 1024, Southwest Main Street, University District, Portland, Oregon
#> 52                                                                                                           128-198 Southwest 2nd Avenue, Portland
#> 53                                                                                                                40 Southwest 3rd Avenue, Portland
#> 54                                                                                                               620 Southwest 5th Avenue, Portland
#> 55                                                                                                               315 Southwest 5th Avenue, Portland
#> 56                                                                                                        300-398 Southwest Market Street, Portland
#> 57                                                           Southwest Morrison & 17th, Southwest Morrison Street, Pearl District, Portland, Oregon
#> 58                                                                                                             640-698 Southwest Broadway, Portland
#> 59                                                                                                       800-898 Southwest Yamhill Street, Portland
#> 60                                                                                                             1315 Northwest 11th Avenue, Portland
#> 61                                                                                                          300-398 Northwest 18th Avenue, Portland
#> 62                                                                                                          945-993 Northwest 22nd Avenue, Portland
#> 63                                                                                                           301-399 Northwest 2nd Avenue, Portland
#> 64                                                                                                            1037 Northwest Couch Street, Portland
#> 65                                                                                                    1400-1498 Northwest Flanders Street, Portland
#> 66                                                                                                                       Tilikum Crossing, Portland
#> 67                                                                                                     2301-2305 Northwest Kearney Street, Portland
#> 68                                                                                        Northwest Naito Parkway, Pearl District, Portland, Oregon
#> 69                                                                                                              1-99 North Failing Street, Portland
#> 70                                                                                                     3731-3741 North Mississippi Avenue, Portland
#> 71                                                                                                               955 North Russell Street, Portland
#> 72                                                                       North Williams at Fremont, North Williams Avenue, Albina, Portland, Oregon
#> 73                                                                                                             1003 Northeast 11th Avenue, Portland
#> 74                                                                                                             43 Southwest Naito Parkway, Portland
#> 75                                                                                                               14 Northeast 22nd Avenue, Portland
#> 76                                                                                                    1100-1198 Northeast Broadway Street, Portland
#> 77                                                                           The Dream, Northeast Holladay Street, Lloyd District, Portland, Oregon
#> 78                                                                                                     796-842 Northeast Multnomah Street, Portland
#> 79                                                                                                     564-598 Northeast Multnomah Street, Portland
#> 80                                                                                                    1527-1541 Northeast Sandy Boulevard, Portland
#> 81                                                                                                       4600-4722 Northeast Grand Avenue, Portland
#> 82                                                                  2835, Southeast Hawthorne Boulevard, Ladd's Addition, Belmont, Portland, Oregon
#> 83                                                                                                          2254-2398 Southeast 2nd Place, Portland
#> 84                                                                                                           900-928 Southeast 9th Avenue, Portland
#> 85                                                                                                         509-599 Southeast Alder Street, Portland
#> 86                                                                                                     3301-3333 Southeast Belmont Street, Portland
#> 87                                                                                                2003-2033 Southeast Hawthorne Boulevard, Portland
#> 88                                                                                                1202-1210 Southeast Hawthorne Boulevard, Portland
#> 89                                                            Balfour–Guthrie Building, 733, Southwest Oak Street, Pearl District, Portland, Oregon
#> 90                                                                                            4073, North Williams Avenue, Albina, Portland, Oregon
#> 91                                                                                                          801-811 Northwest 19th Avenue, Portland
#> 92                                                                                                              108 Northeast 28th Avenue, Portland
#> 93                                                                                                              304 Southeast 28th Avenue, Portland
#> 94                                                                                                                        Apotheker Plaza, Portland
#> 95                                                                                                          1006 Northeast Alberta Street, Portland
#> 96                                                                                                          1413 Northeast Alberta Street, Portland
#> 97                                                          Courtesy Janitorial Services, 1705, Northeast Alberta Street, Alameda, Portland, Oregon
#> 98                                                                                                          2209 Northeast Alberta Street, Portland
#> 99                                                                    Northeast Alberta & 24th, Northeast Alberta Street, Alameda, Portland, Oregon
#> 100                                                                            Ampersand, 2916, Northeast Alberta Street, Alameda, Portland, Oregon
#> 101                                                                          Kennedy School, 5736, Northeast 33rd Avenue, Alameda, Portland, Oregon
#> 102                                                                                                            5320 Northeast 33rd Avenue, Portland
#> 103                                                                                                             4520 Northeast 14th Place, Portland
#> 104                                                                                                    1441-1499 Northeast Fremont Street, Portland
#> 105                                                                                                            5433 North Michigan Avenue, Portland
#> 106                                                                                                        5410-5498 North Concord Avenue, Portland
#> 107                                                                                                       2116 North Killingsworth Street, Portland
#> 108                                                                        Cocotte, 2930, Northeast Killingsworth Street, Alameda, Portland, Oregon
#> 109                                                                     Podnah's BBQ Pit, Northeast Killingsworth Street, Alameda, Portland, Oregon
#> 110                                                                                                    4034 Southeast Hawthorne Boulevard, Portland
#> 111                                                                                                          679-695 Pacific Highway West, Portland
#> 112                                                 Interstate Firehouse Cultural Center, 5340, North Interstate Avenue, Piedmont, Portland, Oregon
#> 113                                                                                                      700-748 Northwest Everett Street, Portland
#> 114                                                                                                 1497-1499 Northwest Pettygrove Street, Portland
#> 115                                                                                                           4000 Southwest Moody Avenue, Portland
#> 116 Northeast Martin Luther King & East Burnside, Northeast Martin Luther King, Junior Boulevard, Ladd's Addition, Lloyd District, Portland, Oregon
#> 117                                                                                                         4220 Southeast Belmont Street, Portland
#> 118                                                                                                        1760-1798 Northeast 2nd Avenue, Portland
#> 119                                                                                                             2609 Southwest Park Place, Portland
#> 120                                                                                                           2035 Northwest Front Avenue, Portland
#> 121                                                                                                         4505 Southeast Belmont Street, Portland
#> 122                                                                                                  3816-3824 Southeast Gladstone Street, Portland
#> 123                                                                                                         5005 Northeast Fremont Street, Portland
#> 124                                                                                                    5931 Northeast Skidmore Street, Portland, OR
#> 125                                                                                                       2886 Southeast Gladstone Street, Portland
#> 126                                                                                                   5008-5022 Southeast Division Street, Portland
#> 127                                                                                               4638-4698 Southeast Hawthorne Boulevard, Portland
#> 128                                                                                                          2852 Northeast Glisan Street, Portland
#> 129                                                                                                    5008 Southeast Hawthorne Boulevard, Portland
#> 130                                                                                                   4846-4848 Southeast Division Street, Portland
#> 131                                                                                                          2948 Northeast Glisan Street, Portland
#> 132                                                                                                        3652 Southeast Division Street, Portland
#> 133                                                                                                   4144-4152 Northeast Sandy Boulevard, Portland
#> 134                                                                                                            1969 Northeast 42nd Avenue, Portland
#> 135                                                                                                        4759B Northeast Fremont Street, Portland
#> 136                                                                                                   2100-2198 Northeast Thompson Street, Portland
#> 137                                                                                                        5425-5499 North Burrage Avenue, Portland
#> 138                                                                                                       1541-1599 Northeast 10th Avenue, Portland
#> 139                                                                                       Northwest Nicolai Street, Kings Heights, Portland, Oregon
#> 140                                                                                                             2-98 Southeast 9th Avenue, Portland
#> 141                                                                                                        5410 Northeast Cully Boulevard, Portland
#> 142                                                                                                        1325-1399 Southeast 9th Avenue, Portland
#> 143                                                                                                  2896-2898 Southeast Powell Boulevard, Portland
#> 144                                                                                                            400 Northwest Couch Street, Portland
#> 145                                                                                                          708-724 Southwest 3rd Avenue, Portland
#> 146                                                                                                            4523 Northeast 60th Avenue, Portland
#> 147                                                                                                            2425 Northeast 50th Avenue, Portland
#> 148                                                                                                     5258-5298 Northeast Glisan Street, Portland
#> 149                                                                                                                 Northeast 42nd Avenue, Portland
#> 150                                                                                                       3474-3498 Southeast 28th Avenue, Portland
#> 151                                                                                                            2025 Northeast 44th Avenue, Portland
#> 152                                                                                                        1656-1686 North Wheeler Avenue, Portland
#> 153                                                                                                    4901-4999 Southeast Clinton Street, Portland
#> 154                                                                                          1510, North Alberta Street, Piedmont, Portland, Oregon
#>                                                               rental_methods
#> 1   KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 2               KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 3               KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 4               KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 5               KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 6               KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 7   KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 8               KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 9               KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 10              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 11              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 12              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 13              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 14  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 15              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 16              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 17              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 18  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 19              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 20              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 21              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 22  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 23              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 24              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 25              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 26              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 27              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 28              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 29              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 30  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 31              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 32              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 33              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 34              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 35  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 36  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 37              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 38              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 39              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 40  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 41  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 42              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 43              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 44              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 45              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 46              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 47  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 48  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 49  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 50              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 51              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 52              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 53  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 54              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 55              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 56              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 57              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 58              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 59              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 60              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 61              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 62              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 63              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 64              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 65              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 66              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 67  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 68              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 69              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 70  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 71              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 72  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 73              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 74              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 75              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 76              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 77              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 78              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 79              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 80              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 81              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 82              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 83              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 84              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 85              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 86              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 87              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 88  KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 89              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 90              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 91              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 92              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 93              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 94              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 95              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 96              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 97              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 98              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 99              KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 100             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 101             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 102             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 103             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 104             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 105             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 106             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 107             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 108             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 109             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 110             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 111             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 112             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 113 KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE, CREDITCARD
#> 114             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 115             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 116             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 117             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 118             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 119             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 120             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 121             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 122             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 123             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 124             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 125             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 126             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 127             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 128             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 129             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 130             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 131             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 132             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 133             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 134             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 135             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 136             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 137             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 138             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 139             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 140             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 141             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 142             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 143             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 144             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 145             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 146             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 147             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 148             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 149             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 150             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 151             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 152             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 153             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE
#> 154             KEY, APPLEPAY, ANDROIDPAY, TRANSITCARD, ACCOUNTNUMBER, PHONE