Grab the station_status feed.

Grab a dataframe giving the geographic location and other metadata of bikeshare bikes parked at bikeshare stations. Metadata for this dataset can be found at: https://github.com/NABSA/gbfs/blob/master/gbfs.md

get_station_status(
  city,
  directory = NULL,
  file = "station_status.rds",
  output = NULL
)

Arguments

city
A character string that can be matched to a city or a url to an active gbfs .json feed. See [get_gbfs_cities()] for a current list of available cities.

directory
Optional. Path to a folder (or folder to be created) where the feed will be saved.

file
Optional. The name of the file to be saved (if output is set to "save" or "both"), as a character string. Must end in ".rds".

output
Optional. The type of output method. If left as default, this argument is inferred from the directory argument. If output = "save", the object will be saved as an .rds object at # the given path. If output = "return", the output will be returned as a dataframe object. Setting output = "both" will do both.

Value

The output of this function depends on the argument to output and directory. Either a saved .rds object generated from the current station_information feed, a dataframe object, or both. If a saved feed of the same type already exists at the filepath, the feed will be appended to rather than overwritten.

See also

Examples

# we can grab the station status feed for memphis, 
# tennessee's bikeshare program in  several ways! the most 
# straightforward way is just to supply the `city` argument
# as a string:
get_station_status(city = "memphis")

#> Message: Returning the output data as an object, rather than saving it, since the `directory` argument was not specified. Setting `output = "return"` will silence this message.

#>    is_returning is_renting is_installed num_docks_available num_bikes_available
#> 1             1          1            1                   0                   2
#> 2             1          1            1                   0                   2
#> 3             1          1            1                   0                   2
#> 4             1          1            1                   0                   3
#> 5             1          1            1                   0                   5
#> 6             1          1            1                   0                   3
#> 7             1          1            1                   0                   0
#> 8             1          1            1                   0                   1
#> 9             1          1            1                   0                   2
#> 10            0          0            1                   0                   0
#> 11            1          1            1                   0                   3
#> 12            0          0            1                   0                   0
#> 13            1          1            1                   0                   5
#> 14            1          1            1                   0                   5
#> 15            1          1            1                   0                   1
#> 16            1          1            1                   0                   3
#> 17            1          1            1                   0                   3
#> 18            1          1            1                   0                   6
#> 19            1          1            1                   0                   5
#> 20            1          1            1                   0                   3
#> 21            1          1            1                   0                   4
#> 22            1          1            1                   0                   0
#> 23            1          1            1                   0                   0
#> 24            1          1            1                   0                   3
#> 25            1          1            1                   0                   4
#> 26            1          1            1                   0                   0
#> 27            1          1            1                   0                   0
#> 28            1          1            1                   0                   0
#> 29            1          1            1                   0                   0
#> 30            1          1            1                   0                   0
#> 31            1          1            1                   0                   0
#> 32            1          1            1                   0                   0
#> 33            1          1            1                   0                   7
#> 34            1          1            1                   0                   6
#> 35            1          1            1                   0                   4
#> 36            1          1            1                   0                   0
#> 37            1          1            1                   0                   3
#> 38            1          1            1                   0                   5
#> 39            1          1            1                   0                   0
#> 40            1          1            1                   0                   2
#> 41            1          1            1                   0                   3
#> 42            1          1            1                   0                   4
#> 43            0          0            1                   0                   0
#> 44            1          1            1                   0                   0
#> 45            1          1            1                   0                   0
#> 46            1          1            1                   0                   1
#> 47            1          1            1                   0                   3
#> 48            1          1            1                   0                   0
#> 49            1          1            1                   0                   4
#> 50            1          1            1                   0                   0
#> 51            1          1            1                   0                   3
#> 52            1          1            1                   0                   4
#> 53            1          1            1                   0                   2
#> 54            1          1            1                   0                   4
#> 55            1          1            1                   0                   5
#> 56            1          1            1                   0                   0
#> 57            1          1            1                   0                   3
#> 58            1          1            1                   0                   0
#> 59            1          1            1                   0                   0
#> 60            1          1            1                   0                   7
#> 61            1          1            1                   0                   3
#> 62            1          1            1                   0                   2
#> 63            1          1            1                   0                   0
#> 64            1          1            1                   0                   6
#> 65            1          1            1                   0                   0
#> 66            1          1            1                   0                   0
#> 67            1          1            1                   0                   2
#> 68            1          1            1                   0                   1
#> 69            1          1            1                   0                   0
#> 70            1          1            1                   0                   5
#> 71            1          1            1                   0                   0
#> 72            1          1            1                   0                   1
#> 73            1          1            1                   0                   0
#> 74            1          1            1                   0                   0
#> 75            1          1            1                   0                   1
#> 76            1          1            1                   0                   0
#> 77            1          1            1                   0                   0
#> 78            1          1            1                   0                   0
#>    last_reported          station_id num_bikes_available_types.electric
#> 1     1601840576 bcycle_memphis_3864                                  0
#> 2     1601840576 bcycle_memphis_3872                                  0
#> 3     1601840576 bcycle_memphis_3873                                  0
#> 4     1601840576 bcycle_memphis_3874                                  0
#> 5     1601840576 bcycle_memphis_3875                                  0
#> 6     1601840576 bcycle_memphis_3924                                  0
#> 7     1601840576 bcycle_memphis_3925                                  0
#> 8     1601840576 bcycle_memphis_3926                                  0
#> 9     1601840576 bcycle_memphis_3927                                  0
#> 10    1601840576 bcycle_memphis_3928                                  0
#> 11    1601840576 bcycle_memphis_3929                                  0
#> 12    1601840576 bcycle_memphis_3930                                  0
#> 13    1601840576 bcycle_memphis_3931                                  0
#> 14    1601840576 bcycle_memphis_3932                                  0
#> 15    1601840576 bcycle_memphis_3933                                  0
#> 16    1601840576 bcycle_memphis_3934                                  0
#> 17    1601840576 bcycle_memphis_3935                                  0
#> 18    1601840576 bcycle_memphis_3936                                  0
#> 19    1601840576 bcycle_memphis_3937                                  0
#> 20    1601840576 bcycle_memphis_3938                                  0
#> 21    1601840576 bcycle_memphis_3940                                  0
#> 22    1601840576 bcycle_memphis_3941                                  0
#> 23    1601840576 bcycle_memphis_3942                                  0
#> 24    1601840576 bcycle_memphis_3943                                  0
#> 25    1601840576 bcycle_memphis_3944                                  0
#> 26    1601840576 bcycle_memphis_3945                                  0
#> 27    1601840576 bcycle_memphis_3947                                  0
#> 28    1601840576 bcycle_memphis_3948                                  0
#> 29    1601840576 bcycle_memphis_3949                                  0
#> 30    1601840576 bcycle_memphis_3951                                  0
#> 31    1601840576 bcycle_memphis_3952                                  0
#> 32    1601840576 bcycle_memphis_3953                                  0
#> 33    1601840576 bcycle_memphis_3955                                  0
#> 34    1601840576 bcycle_memphis_3957                                  0
#> 35    1601840576 bcycle_memphis_3959                                  0
#> 36    1601840576 bcycle_memphis_3960                                  0
#> 37    1601840576 bcycle_memphis_3961                                  0
#> 38    1601840576 bcycle_memphis_3963                                  0
#> 39    1601840576 bcycle_memphis_3965                                  0
#> 40    1601840576 bcycle_memphis_3966                                  0
#> 41    1601840576 bcycle_memphis_3967                                  0
#> 42    1601840576 bcycle_memphis_3968                                  0
#> 43    1601840576 bcycle_memphis_3970                                  0
#> 44    1601840576 bcycle_memphis_3971                                  0
#> 45    1601840576 bcycle_memphis_3972                                  0
#> 46    1601840576 bcycle_memphis_3973                                  0
#> 47    1601840576 bcycle_memphis_3974                                  0
#> 48    1601840576 bcycle_memphis_3976                                  0
#> 49    1601840576 bcycle_memphis_3984                                  0
#> 50    1601840576 bcycle_memphis_3986                                  0
#> 51    1601840576 bcycle_memphis_4101                                  0
#> 52    1601840576 bcycle_memphis_4196                                  0
#> 53    1601840576 bcycle_memphis_4197                                  0
#> 54    1601840576 bcycle_memphis_4198                                  0
#> 55    1601840576 bcycle_memphis_4201                                  0
#> 56    1601840576 bcycle_memphis_4255                                  0
#> 57    1601840576 bcycle_memphis_4264                                  0
#> 58    1601840576 bcycle_memphis_4269                                  0
#> 59    1601840576 bcycle_memphis_4279                                  0
#> 60    1601840576 bcycle_memphis_4281                                  0
#> 61    1601840576 bcycle_memphis_4285                                  0
#> 62    1601840576 bcycle_memphis_4307                                  0
#> 63    1601840576 bcycle_memphis_4308                                  0
#> 64    1601840576 bcycle_memphis_4309                                  0
#> 65    1601840576 bcycle_memphis_4313                                  0
#> 66    1601840576 bcycle_memphis_4314                                  0
#> 67    1601840576 bcycle_memphis_4328                                  0
#> 68    1601840576 bcycle_memphis_4330                                  0
#> 69    1601840576 bcycle_memphis_4368                                  0
#> 70    1601840576 bcycle_memphis_4369                                  0
#> 71    1601840576 bcycle_memphis_4433                                  0
#> 72    1601840576 bcycle_memphis_4434                                  0
#> 73    1601840576 bcycle_memphis_4435                                  0
#> 74    1601840576 bcycle_memphis_4519                                  0
#> 75    1601840576 bcycle_memphis_4520                                  0
#> 76    1601840576 bcycle_memphis_4535                                  0
#> 77    1601840576 bcycle_memphis_4536                                  0
#> 78    1601840576 bcycle_memphis_4574                                  0
#>    num_bikes_available_types.smart num_bikes_available_types.classic
#> 1                                2                                 0
#> 2                                2                                 0
#> 3                                2                                 0
#> 4                                3                                 0
#> 5                                5                                 0
#> 6                                3                                 0
#> 7                                0                                 0
#> 8                                1                                 0
#> 9                                2                                 0
#> 10                               0                                 0
#> 11                               3                                 0
#> 12                               0                                 0
#> 13                               5                                 0
#> 14                               5                                 0
#> 15                               1                                 0
#> 16                               3                                 0
#> 17                               3                                 0
#> 18                               6                                 0
#> 19                               5                                 0
#> 20                               3                                 0
#> 21                               4                                 0
#> 22                               0                                 0
#> 23                               0                                 0
#> 24                               3                                 0
#> 25                               4                                 0
#> 26                               0                                 0
#> 27                               0                                 0
#> 28                               0                                 0
#> 29                               0                                 0
#> 30                               0                                 0
#> 31                               0                                 0
#> 32                               0                                 0
#> 33                               7                                 0
#> 34                               6                                 0
#> 35                               4                                 0
#> 36                               0                                 0
#> 37                               3                                 0
#> 38                               5                                 0
#> 39                               0                                 0
#> 40                               2                                 0
#> 41                               3                                 0
#> 42                               4                                 0
#> 43                               0                                 0
#> 44                               0                                 0
#> 45                               0                                 0
#> 46                               1                                 0
#> 47                               3                                 0
#> 48                               0                                 0
#> 49                               4                                 0
#> 50                               0                                 0
#> 51                               3                                 0
#> 52                               4                                 0
#> 53                               2                                 0
#> 54                               4                                 0
#> 55                               5                                 0
#> 56                               0                                 0
#> 57                               3                                 0
#> 58                               0                                 0
#> 59                               0                                 0
#> 60                               7                                 0
#> 61                               3                                 0
#> 62                               2                                 0
#> 63                               0                                 0
#> 64                               6                                 0
#> 65                               0                                 0
#> 66                               0                                 0
#> 67                               2                                 0
#> 68                               1                                 0
#> 69                               0                                 0
#> 70                               5                                 0
#> 71                               0                                 0
#> 72                               1                                 0
#> 73                               0                                 0
#> 74                               0                                 0
#> 75                               1                                 0
#> 76                               0                                 0
#> 77                               0                                 0
#> 78                               0                                 0
#>           last_updated year month day hour minute
#> 1  2020-10-04 12:42:56 2020    10   4   12     42
#> 2  2020-10-04 12:42:56 2020    10   4   12     42
#> 3  2020-10-04 12:42:56 2020    10   4   12     42
#> 4  2020-10-04 12:42:56 2020    10   4   12     42
#> 5  2020-10-04 12:42:56 2020    10   4   12     42
#> 6  2020-10-04 12:42:56 2020    10   4   12     42
#> 7  2020-10-04 12:42:56 2020    10   4   12     42
#> 8  2020-10-04 12:42:56 2020    10   4   12     42
#> 9  2020-10-04 12:42:56 2020    10   4   12     42
#> 10 2020-10-04 12:42:56 2020    10   4   12     42
#> 11 2020-10-04 12:42:56 2020    10   4   12     42
#> 12 2020-10-04 12:42:56 2020    10   4   12     42
#> 13 2020-10-04 12:42:56 2020    10   4   12     42
#> 14 2020-10-04 12:42:56 2020    10   4   12     42
#> 15 2020-10-04 12:42:56 2020    10   4   12     42
#> 16 2020-10-04 12:42:56 2020    10   4   12     42
#> 17 2020-10-04 12:42:56 2020    10   4   12     42
#> 18 2020-10-04 12:42:56 2020    10   4   12     42
#> 19 2020-10-04 12:42:56 2020    10   4   12     42
#> 20 2020-10-04 12:42:56 2020    10   4   12     42
#> 21 2020-10-04 12:42:56 2020    10   4   12     42
#> 22 2020-10-04 12:42:56 2020    10   4   12     42
#> 23 2020-10-04 12:42:56 2020    10   4   12     42
#> 24 2020-10-04 12:42:56 2020    10   4   12     42
#> 25 2020-10-04 12:42:56 2020    10   4   12     42
#> 26 2020-10-04 12:42:56 2020    10   4   12     42
#> 27 2020-10-04 12:42:56 2020    10   4   12     42
#> 28 2020-10-04 12:42:56 2020    10   4   12     42
#> 29 2020-10-04 12:42:56 2020    10   4   12     42
#> 30 2020-10-04 12:42:56 2020    10   4   12     42
#> 31 2020-10-04 12:42:56 2020    10   4   12     42
#> 32 2020-10-04 12:42:56 2020    10   4   12     42
#> 33 2020-10-04 12:42:56 2020    10   4   12     42
#> 34 2020-10-04 12:42:56 2020    10   4   12     42
#> 35 2020-10-04 12:42:56 2020    10   4   12     42
#> 36 2020-10-04 12:42:56 2020    10   4   12     42
#> 37 2020-10-04 12:42:56 2020    10   4   12     42
#> 38 2020-10-04 12:42:56 2020    10   4   12     42
#> 39 2020-10-04 12:42:56 2020    10   4   12     42
#> 40 2020-10-04 12:42:56 2020    10   4   12     42
#> 41 2020-10-04 12:42:56 2020    10   4   12     42
#> 42 2020-10-04 12:42:56 2020    10   4   12     42
#> 43 2020-10-04 12:42:56 2020    10   4   12     42
#> 44 2020-10-04 12:42:56 2020    10   4   12     42
#> 45 2020-10-04 12:42:56 2020    10   4   12     42
#> 46 2020-10-04 12:42:56 2020    10   4   12     42
#> 47 2020-10-04 12:42:56 2020    10   4   12     42
#> 48 2020-10-04 12:42:56 2020    10   4   12     42
#> 49 2020-10-04 12:42:56 2020    10   4   12     42
#> 50 2020-10-04 12:42:56 2020    10   4   12     42
#> 51 2020-10-04 12:42:56 2020    10   4   12     42
#> 52 2020-10-04 12:42:56 2020    10   4   12     42
#> 53 2020-10-04 12:42:56 2020    10   4   12     42
#> 54 2020-10-04 12:42:56 2020    10   4   12     42
#> 55 2020-10-04 12:42:56 2020    10   4   12     42
#> 56 2020-10-04 12:42:56 2020    10   4   12     42
#> 57 2020-10-04 12:42:56 2020    10   4   12     42
#> 58 2020-10-04 12:42:56 2020    10   4   12     42
#> 59 2020-10-04 12:42:56 2020    10   4   12     42
#> 60 2020-10-04 12:42:56 2020    10   4   12     42
#> 61 2020-10-04 12:42:56 2020    10   4   12     42
#> 62 2020-10-04 12:42:56 2020    10   4   12     42
#> 63 2020-10-04 12:42:56 2020    10   4   12     42
#> 64 2020-10-04 12:42:56 2020    10   4   12     42
#> 65 2020-10-04 12:42:56 2020    10   4   12     42
#> 66 2020-10-04 12:42:56 2020    10   4   12     42
#> 67 2020-10-04 12:42:56 2020    10   4   12     42
#> 68 2020-10-04 12:42:56 2020    10   4   12     42
#> 69 2020-10-04 12:42:56 2020    10   4   12     42
#> 70 2020-10-04 12:42:56 2020    10   4   12     42
#> 71 2020-10-04 12:42:56 2020    10   4   12     42
#> 72 2020-10-04 12:42:56 2020    10   4   12     42
#> 73 2020-10-04 12:42:56 2020    10   4   12     42
#> 74 2020-10-04 12:42:56 2020    10   4   12     42
#> 75 2020-10-04 12:42:56 2020    10   4   12     42
#> 76 2020-10-04 12:42:56 2020    10   4   12     42
#> 77 2020-10-04 12:42:56 2020    10   4   12     42
#> 78 2020-10-04 12:42:56 2020    10   4   12     42

# the `city` argument can also be supplied as an
# actual URL to an active .json feed
get_station_status(city = 
"https://gbfs.bcycle.com/bcycle_memphis/station_status.json",  
                     directory = tempdir())

#> Error in datasets_can_be_row_binded(data, paste(directory, file, sep = "/")): The bikeshare data just pulled has 14 columns, while the already stored bikeshare data, at file path /var/folders/qz/jqpclkqn0l9fkr5fr11xrzlh0000gn/T//RtmpkpZ3IW/station_status.rds, has 16 columns, so they can not be row-binded.

# if you'd like to save the output to file, supply a 
# `directory` argument. usually, though, we would supply a 
# character string (like "memphis", maybe,) for the `directory` 
# argument instead of `tempdir`.
get_station_status(city = "memphis",  
                     directory = tempdir())
                     
# the output argument can control whether the file is
# saved and/or returned explicitly
get_station_status(city = "memphis",  
                     directory = tempdir(),
                     output = "both") 

#>    is_returning is_renting is_installed num_docks_available num_bikes_available
#> 1             1          1            1                   0                   2
#> 2             1          1            1                   0                   2
#> 3             1          1            1                   0                   2
#> 4             1          1            1                   0                   3
#> 5             1          1            1                   0                   5
#> 6             1          1            1                   0                   3
#> 7             1          1            1                   0                   0
#> 8             1          1            1                   0                   1
#> 9             1          1            1                   0                   2
#> 10            0          0            1                   0                   0
#> 11            1          1            1                   0                   3
#> 12            0          0            1                   0                   0
#> 13            1          1            1                   0                   5
#> 14            1          1            1                   0                   5
#> 15            1          1            1                   0                   1
#> 16            1          1            1                   0                   3
#> 17            1          1            1                   0                   3
#> 18            1          1            1                   0                   6
#> 19            1          1            1                   0                   5
#> 20            1          1            1                   0                   3
#> 21            1          1            1                   0                   4
#> 22            1          1            1                   0                   0
#> 23            1          1            1                   0                   0
#> 24            1          1            1                   0                   3
#> 25            1          1            1                   0                   4
#> 26            1          1            1                   0                   0
#> 27            1          1            1                   0                   0
#> 28            1          1            1                   0                   0
#> 29            1          1            1                   0                   0
#> 30            1          1            1                   0                   0
#> 31            1          1            1                   0                   0
#> 32            1          1            1                   0                   0
#> 33            1          1            1                   0                   7
#> 34            1          1            1                   0                   6
#> 35            1          1            1                   0                   4
#> 36            1          1            1                   0                   0
#> 37            1          1            1                   0                   3
#> 38            1          1            1                   0                   5
#> 39            1          1            1                   0                   0
#> 40            1          1            1                   0                   2
#> 41            1          1            1                   0                   3
#> 42            1          1            1                   0                   4
#> 43            0          0            1                   0                   0
#> 44            1          1            1                   0                   0
#> 45            1          1            1                   0                   0
#> 46            1          1            1                   0                   1
#> 47            1          1            1                   0                   3
#> 48            1          1            1                   0                   0
#> 49            1          1            1                   0                   4
#> 50            1          1            1                   0                   0
#> 51            1          1            1                   0                   3
#> 52            1          1            1                   0                   4
#> 53            1          1            1                   0                   2
#> 54            1          1            1                   0                   4
#> 55            1          1            1                   0                   5
#> 56            1          1            1                   0                   0
#> 57            1          1            1                   0                   3
#> 58            1          1            1                   0                   0
#> 59            1          1            1                   0                   0
#> 60            1          1            1                   0                   7
#> 61            1          1            1                   0                   3
#> 62            1          1            1                   0                   2
#> 63            1          1            1                   0                   0
#> 64            1          1            1                   0                   6
#> 65            1          1            1                   0                   0
#> 66            1          1            1                   0                   0
#> 67            1          1            1                   0                   2
#> 68            1          1            1                   0                   1
#> 69            1          1            1                   0                   0
#> 70            1          1            1                   0                   5
#> 71            1          1            1                   0                   0
#> 72            1          1            1                   0                   1
#> 73            1          1            1                   0                   0
#> 74            1          1            1                   0                   0
#> 75            1          1            1                   0                   1
#> 76            1          1            1                   0                   0
#> 77            1          1            1                   0                   0
#> 78            1          1            1                   0                   0
#>    last_reported          station_id num_bikes_available_types.electric
#> 1     1601840576 bcycle_memphis_3864                                  0
#> 2     1601840576 bcycle_memphis_3872                                  0
#> 3     1601840576 bcycle_memphis_3873                                  0
#> 4     1601840576 bcycle_memphis_3874                                  0
#> 5     1601840576 bcycle_memphis_3875                                  0
#> 6     1601840576 bcycle_memphis_3924                                  0
#> 7     1601840576 bcycle_memphis_3925                                  0
#> 8     1601840576 bcycle_memphis_3926                                  0
#> 9     1601840576 bcycle_memphis_3927                                  0
#> 10    1601840576 bcycle_memphis_3928                                  0
#> 11    1601840576 bcycle_memphis_3929                                  0
#> 12    1601840576 bcycle_memphis_3930                                  0
#> 13    1601840576 bcycle_memphis_3931                                  0
#> 14    1601840576 bcycle_memphis_3932                                  0
#> 15    1601840576 bcycle_memphis_3933                                  0
#> 16    1601840576 bcycle_memphis_3934                                  0
#> 17    1601840576 bcycle_memphis_3935                                  0
#> 18    1601840576 bcycle_memphis_3936                                  0
#> 19    1601840576 bcycle_memphis_3937                                  0
#> 20    1601840576 bcycle_memphis_3938                                  0
#> 21    1601840576 bcycle_memphis_3940                                  0
#> 22    1601840576 bcycle_memphis_3941                                  0
#> 23    1601840576 bcycle_memphis_3942                                  0
#> 24    1601840576 bcycle_memphis_3943                                  0
#> 25    1601840576 bcycle_memphis_3944                                  0
#> 26    1601840576 bcycle_memphis_3945                                  0
#> 27    1601840576 bcycle_memphis_3947                                  0
#> 28    1601840576 bcycle_memphis_3948                                  0
#> 29    1601840576 bcycle_memphis_3949                                  0
#> 30    1601840576 bcycle_memphis_3951                                  0
#> 31    1601840576 bcycle_memphis_3952                                  0
#> 32    1601840576 bcycle_memphis_3953                                  0
#> 33    1601840576 bcycle_memphis_3955                                  0
#> 34    1601840576 bcycle_memphis_3957                                  0
#> 35    1601840576 bcycle_memphis_3959                                  0
#> 36    1601840576 bcycle_memphis_3960                                  0
#> 37    1601840576 bcycle_memphis_3961                                  0
#> 38    1601840576 bcycle_memphis_3963                                  0
#> 39    1601840576 bcycle_memphis_3965                                  0
#> 40    1601840576 bcycle_memphis_3966                                  0
#> 41    1601840576 bcycle_memphis_3967                                  0
#> 42    1601840576 bcycle_memphis_3968                                  0
#> 43    1601840576 bcycle_memphis_3970                                  0
#> 44    1601840576 bcycle_memphis_3971                                  0
#> 45    1601840576 bcycle_memphis_3972                                  0
#> 46    1601840576 bcycle_memphis_3973                                  0
#> 47    1601840576 bcycle_memphis_3974                                  0
#> 48    1601840576 bcycle_memphis_3976                                  0
#> 49    1601840576 bcycle_memphis_3984                                  0
#> 50    1601840576 bcycle_memphis_3986                                  0
#> 51    1601840576 bcycle_memphis_4101                                  0
#> 52    1601840576 bcycle_memphis_4196                                  0
#> 53    1601840576 bcycle_memphis_4197                                  0
#> 54    1601840576 bcycle_memphis_4198                                  0
#> 55    1601840576 bcycle_memphis_4201                                  0
#> 56    1601840576 bcycle_memphis_4255                                  0
#> 57    1601840576 bcycle_memphis_4264                                  0
#> 58    1601840576 bcycle_memphis_4269                                  0
#> 59    1601840576 bcycle_memphis_4279                                  0
#> 60    1601840576 bcycle_memphis_4281                                  0
#> 61    1601840576 bcycle_memphis_4285                                  0
#> 62    1601840576 bcycle_memphis_4307                                  0
#> 63    1601840576 bcycle_memphis_4308                                  0
#> 64    1601840576 bcycle_memphis_4309                                  0
#> 65    1601840576 bcycle_memphis_4313                                  0
#> 66    1601840576 bcycle_memphis_4314                                  0
#> 67    1601840576 bcycle_memphis_4328                                  0
#> 68    1601840576 bcycle_memphis_4330                                  0
#> 69    1601840576 bcycle_memphis_4368                                  0
#> 70    1601840576 bcycle_memphis_4369                                  0
#> 71    1601840576 bcycle_memphis_4433                                  0
#> 72    1601840576 bcycle_memphis_4434                                  0
#> 73    1601840576 bcycle_memphis_4435                                  0
#> 74    1601840576 bcycle_memphis_4519                                  0
#> 75    1601840576 bcycle_memphis_4520                                  0
#> 76    1601840576 bcycle_memphis_4535                                  0
#> 77    1601840576 bcycle_memphis_4536                                  0
#> 78    1601840576 bcycle_memphis_4574                                  0
#>    num_bikes_available_types.smart num_bikes_available_types.classic
#> 1                                2                                 0
#> 2                                2                                 0
#> 3                                2                                 0
#> 4                                3                                 0
#> 5                                5                                 0
#> 6                                3                                 0
#> 7                                0                                 0
#> 8                                1                                 0
#> 9                                2                                 0
#> 10                               0                                 0
#> 11                               3                                 0
#> 12                               0                                 0
#> 13                               5                                 0
#> 14                               5                                 0
#> 15                               1                                 0
#> 16                               3                                 0
#> 17                               3                                 0
#> 18                               6                                 0
#> 19                               5                                 0
#> 20                               3                                 0
#> 21                               4                                 0
#> 22                               0                                 0
#> 23                               0                                 0
#> 24                               3                                 0
#> 25                               4                                 0
#> 26                               0                                 0
#> 27                               0                                 0
#> 28                               0                                 0
#> 29                               0                                 0
#> 30                               0                                 0
#> 31                               0                                 0
#> 32                               0                                 0
#> 33                               7                                 0
#> 34                               6                                 0
#> 35                               4                                 0
#> 36                               0                                 0
#> 37                               3                                 0
#> 38                               5                                 0
#> 39                               0                                 0
#> 40                               2                                 0
#> 41                               3                                 0
#> 42                               4                                 0
#> 43                               0                                 0
#> 44                               0                                 0
#> 45                               0                                 0
#> 46                               1                                 0
#> 47                               3                                 0
#> 48                               0                                 0
#> 49                               4                                 0
#> 50                               0                                 0
#> 51                               3                                 0
#> 52                               4                                 0
#> 53                               2                                 0
#> 54                               4                                 0
#> 55                               5                                 0
#> 56                               0                                 0
#> 57                               3                                 0
#> 58                               0                                 0
#> 59                               0                                 0
#> 60                               7                                 0
#> 61                               3                                 0
#> 62                               2                                 0
#> 63                               0                                 0
#> 64                               6                                 0
#> 65                               0                                 0
#> 66                               0                                 0
#> 67                               2                                 0
#> 68                               1                                 0
#> 69                               0                                 0
#> 70                               5                                 0
#> 71                               0                                 0
#> 72                               1                                 0
#> 73                               0                                 0
#> 74                               0                                 0
#> 75                               1                                 0
#> 76                               0                                 0
#> 77                               0                                 0
#> 78                               0                                 0
#>           last_updated year month day hour minute
#> 1  2020-10-04 12:42:56 2020    10   4   12     42
#> 2  2020-10-04 12:42:56 2020    10   4   12     42
#> 3  2020-10-04 12:42:56 2020    10   4   12     42
#> 4  2020-10-04 12:42:56 2020    10   4   12     42
#> 5  2020-10-04 12:42:56 2020    10   4   12     42
#> 6  2020-10-04 12:42:56 2020    10   4   12     42
#> 7  2020-10-04 12:42:56 2020    10   4   12     42
#> 8  2020-10-04 12:42:56 2020    10   4   12     42
#> 9  2020-10-04 12:42:56 2020    10   4   12     42
#> 10 2020-10-04 12:42:56 2020    10   4   12     42
#> 11 2020-10-04 12:42:56 2020    10   4   12     42
#> 12 2020-10-04 12:42:56 2020    10   4   12     42
#> 13 2020-10-04 12:42:56 2020    10   4   12     42
#> 14 2020-10-04 12:42:56 2020    10   4   12     42
#> 15 2020-10-04 12:42:56 2020    10   4   12     42
#> 16 2020-10-04 12:42:56 2020    10   4   12     42
#> 17 2020-10-04 12:42:56 2020    10   4   12     42
#> 18 2020-10-04 12:42:56 2020    10   4   12     42
#> 19 2020-10-04 12:42:56 2020    10   4   12     42
#> 20 2020-10-04 12:42:56 2020    10   4   12     42
#> 21 2020-10-04 12:42:56 2020    10   4   12     42
#> 22 2020-10-04 12:42:56 2020    10   4   12     42
#> 23 2020-10-04 12:42:56 2020    10   4   12     42
#> 24 2020-10-04 12:42:56 2020    10   4   12     42
#> 25 2020-10-04 12:42:56 2020    10   4   12     42
#> 26 2020-10-04 12:42:56 2020    10   4   12     42
#> 27 2020-10-04 12:42:56 2020    10   4   12     42
#> 28 2020-10-04 12:42:56 2020    10   4   12     42
#> 29 2020-10-04 12:42:56 2020    10   4   12     42
#> 30 2020-10-04 12:42:56 2020    10   4   12     42
#> 31 2020-10-04 12:42:56 2020    10   4   12     42
#> 32 2020-10-04 12:42:56 2020    10   4   12     42
#> 33 2020-10-04 12:42:56 2020    10   4   12     42
#> 34 2020-10-04 12:42:56 2020    10   4   12     42
#> 35 2020-10-04 12:42:56 2020    10   4   12     42
#> 36 2020-10-04 12:42:56 2020    10   4   12     42
#> 37 2020-10-04 12:42:56 2020    10   4   12     42
#> 38 2020-10-04 12:42:56 2020    10   4   12     42
#> 39 2020-10-04 12:42:56 2020    10   4   12     42
#> 40 2020-10-04 12:42:56 2020    10   4   12     42
#> 41 2020-10-04 12:42:56 2020    10   4   12     42
#> 42 2020-10-04 12:42:56 2020    10   4   12     42
#> 43 2020-10-04 12:42:56 2020    10   4   12     42
#> 44 2020-10-04 12:42:56 2020    10   4   12     42
#> 45 2020-10-04 12:42:56 2020    10   4   12     42
#> 46 2020-10-04 12:42:56 2020    10   4   12     42
#> 47 2020-10-04 12:42:56 2020    10   4   12     42
#> 48 2020-10-04 12:42:56 2020    10   4   12     42
#> 49 2020-10-04 12:42:56 2020    10   4   12     42
#> 50 2020-10-04 12:42:56 2020    10   4   12     42
#> 51 2020-10-04 12:42:56 2020    10   4   12     42
#> 52 2020-10-04 12:42:56 2020    10   4   12     42
#> 53 2020-10-04 12:42:56 2020    10   4   12     42
#> 54 2020-10-04 12:42:56 2020    10   4   12     42
#> 55 2020-10-04 12:42:56 2020    10   4   12     42
#> 56 2020-10-04 12:42:56 2020    10   4   12     42
#> 57 2020-10-04 12:42:56 2020    10   4   12     42
#> 58 2020-10-04 12:42:56 2020    10   4   12     42
#> 59 2020-10-04 12:42:56 2020    10   4   12     42
#> 60 2020-10-04 12:42:56 2020    10   4   12     42
#> 61 2020-10-04 12:42:56 2020    10   4   12     42
#> 62 2020-10-04 12:42:56 2020    10   4   12     42
#> 63 2020-10-04 12:42:56 2020    10   4   12     42
#> 64 2020-10-04 12:42:56 2020    10   4   12     42
#> 65 2020-10-04 12:42:56 2020    10   4   12     42
#> 66 2020-10-04 12:42:56 2020    10   4   12     42
#> 67 2020-10-04 12:42:56 2020    10   4   12     42
#> 68 2020-10-04 12:42:56 2020    10   4   12     42
#> 69 2020-10-04 12:42:56 2020    10   4   12     42
#> 70 2020-10-04 12:42:56 2020    10   4   12     42
#> 71 2020-10-04 12:42:56 2020    10   4   12     42
#> 72 2020-10-04 12:42:56 2020    10   4   12     42
#> 73 2020-10-04 12:42:56 2020    10   4   12     42
#> 74 2020-10-04 12:42:56 2020    10   4   12     42
#> 75 2020-10-04 12:42:56 2020    10   4   12     42
#> 76 2020-10-04 12:42:56 2020    10   4   12     42
#> 77 2020-10-04 12:42:56 2020    10   4   12     42
#> 78 2020-10-04 12:42:56 2020    10   4   12     42