Get dataframe of bikeshare feeds released by a city

Of the different types of feeds supplied by the gbfs, some are required, some are conditionally required, and some are optional. This function grabs a list of each of the feeds supplied by a given city, as well as the URLs to access them.

get_which_gbfs_feeds(city)

Arguments

city
A character string that can be matched to a city or a url to an active gbfs .json feed. See [get_gbfs_cities()] for a current list of available cities.

Source

North American Bikeshare Association, General Bikeshare Feed Specification https://github.com/NABSA/gbfs/blob/master/gbfs.md

Value

A data.frame containing the feeds supplied by a city. . The `feed` column supplies the name of the relevant .json feeds, while the entries in the `URL` column supply the feeds themselves.

Examples

# grab all of the feeds released by memphis
get_which_gbfs_feeds(city = "memphis")

#>                                                                url
#> 1                 https://gbfs.bcycle.com/bcycle_memphis/gbfs.json
#> 2   https://gbfs.bcycle.com/bcycle_memphis/system_information.json
#> 3  https://gbfs.bcycle.com/bcycle_memphis/station_information.json
#> 4       https://gbfs.bcycle.com/bcycle_memphis/station_status.json
#> 5 https://gbfs.bcycle.com/bcycle_memphis/system_pricing_plans.json
#> 6       https://gbfs.bcycle.com/bcycle_memphis/system_regions.json
#> 7     https://gbfs.bcycle.com/bcycle_memphis/free_bike_status.json
#>                   name
#> 1                 gbfs
#> 2   system_information
#> 3  station_information
#> 4       station_status
#> 5 system_pricing_plans
#> 6       system_regions
#> 7     free_bike_status